### 2022-09-01 #4730 @editorialbot
[PRE REVIEW]: MParT: Monotone Parameterization Toolkit
**Submitting author:** @mparno (<a href="http://orcid.org/0000-0002-9419-2693">Matthew Parno</a>)
**Repository:** https://github.com/MeasureTransport/MParT
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
