### 2022-09-01 #4730 @Kevin-Mattheus-Moerman
@mparno thanks for this submission. Our team will be looking for a handling editor. In the mean time could you please try to address those potentially missing/invalid DOIs :point_up: . You can update your .bib file and paper if needed and update the draft paper using `@editorialbot generate pdf` and use `@editorialbot check references` to redo the check. Thanks.
