+++ joss/paper.bib
@@ -30,7 +30,8 @@
   volume={6},
   number={68},
   pages={3076},
-  year={2021}
+  year={2021},
+  doi="10.21105/joss.03076"
 }
 
 @article{parno2018transport,
@@ -41,7 +42,8 @@
   number={2},
   pages={645--682},
   year={2018},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi="10.1137/17m1134640"
 }
 
 @article{cotter2019ensemble,
@@ -52,7 +54,8 @@
   number={2},
   pages={444--471},
   year={2019},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi="10.1137/17m1114867"
 }
 
 @article{el2012bayesian,
@@ -92,7 +95,8 @@
   number={12},
   pages={3202--3216},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi="10.1016/j.jpdc.2014.07.003"
 }
 
 @book{santambrogio2015optimal,
@@ -102,6 +106,7 @@
     publisher={Birk{\"a}user Cham},
     year={2015},
     edition={1},
+    doi="10.1007/978-3-319-20828-2"
 }
 
 @Inbook{Marzouk2016,
