### 2022-09-02 10ed1c2
Added linearized basis function class.
