### 2022-09-02 56630a5
Added linearization options to map factory.
