+++ src/MapFactory.cpp
@@ -7,6 +7,7 @@
 #include "MParT/HermiteFunction.h"
 #include "MParT/MultivariateExpansionWorker.h"
 #include "MParT/PositiveBijectors.h"
+#include "MParT/LinearizedBasis.h"
 
 using namespace mpart;
 
@@ -19,49 +20,107 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
         AdaptiveSimpson<MemorySpace> quad(opts.quadMaxSub, 1, nullptr, opts.quadAbsTol, opts.quadRelTol, QuadError::First, opts.quadMinSub);
 
         if(opts.basisType==BasisTypes::ProbabilistHermite){
+            
+            if(isinf(opts.basisLB) && isinf(opts.basisUB)){
+               
+                MultivariateExpansionWorker<ProbabilistHermite,MemorySpace> expansion(mset);
+                std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
-            MultivariateExpansionWorker<ProbabilistHermite,MemorySpace> expansion(mset);
-            std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
+                switch(opts.posFuncType) {
+                    case PosFuncTypes::SoftPlus:
+                        output = std::make_shared<MonotoneComponent<decltype(expansion), SoftPlus, decltype(quad), MemorySpace>>(mset, quad, opts.contDeriv); break;
+                    case PosFuncTypes::Exp:
+                        output = std::make_shared<MonotoneComponent<decltype(expansion), Exp, decltype(quad), MemorySpace>>(mset, quad, opts.contDeriv); break;
+                }
 
-            switch(opts.posFuncType) {
-                case PosFuncTypes::SoftPlus:
-                    output = std::make_shared<MonotoneComponent<decltype(expansion), SoftPlus, decltype(quad), MemorySpace>>(mset, quad, opts.contDeriv); break;
-                case PosFuncTypes::Exp:
-                    output = std::make_shared<MonotoneComponent<decltype(expansion), Exp, decltype(quad), MemorySpace>>(mset, quad, opts.contDeriv); break;
+                output->SetCoeffs(Kokkos::View<double*,MemorySpace>("Component Coefficients", mset.Size()));
+                return output;
+
+            }else{
+
+                LinearizedBasis<ProbabilistHermite> basis1d(opts.basisLB, opts.basisUB);
+
+                MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
+                std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
+
+                switch(opts.posFuncType) {
+                    case PosFuncTypes::SoftPlus:
+                        output = std::make_shared<MonotoneComponent<decltype(expansion), SoftPlus, decltype(quad), MemorySpace>>(mset, quad, opts.contDeriv); break;
+                    case PosFuncTypes::Exp:
+                        output = std::make_shared<MonotoneComponent<decltype(expansion), Exp, decltype(quad), MemorySpace>>(mset, quad, opts.contDeriv); break;
+                }
+
+                output->SetCoeffs(Kokkos::View<double*,MemorySpace>("Component Coefficients", mset.Size()));
+                return output;
             }
 
-            output->SetCoeffs(Kokkos::View<double*,MemorySpace>("Component Coefficients", mset.Size()));
-            return output;
-
+            
         }else if(opts.basisType==BasisTypes::PhysicistHermite){
 
-            MultivariateExpansionWorker<PhysicistHermite, MemorySpace> expansion(mset);
-            std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
+            if(isinf(opts.basisLB) && isinf(opts.basisUB)){
+                
[…]
