### 2022-09-02 5248627
Added linearization options to language bindings.
