+++ bindings/matlab/src/ParameterizedFunctionBase_mex.cpp
@@ -45,7 +45,7 @@ MEX_DEFINE(ParameterizedFunction_newMap) (int nlhs, mxArray* plhs[],
                                          input.get<std::string>(4),input.get<double>(5),
                                          input.get<double>(6),input.get<unsigned int>(7),
                                          input.get<unsigned int>(8),input.get<unsigned int>(9),
-                                         input.get<bool>(10));
+                                         input.get<bool>(10),input.get<double>(11),input.get<double>(12));
 
   output.set(0, Session<ParameterizedFunctionMex>::create(new ParameterizedFunctionMex(outputDim,mset.Fix(),opts)));
 }
