+++ bindings/python/src/MapOptions.cpp
@@ -33,6 +33,8 @@ void mpart::binding::MapOptionsWrapper(py::module &m)
     py::class_<MapOptions, std::shared_ptr<MapOptions>>(m, "MapOptions")
     .def(py::init<>())
     .def_readwrite("basisType", &MapOptions::basisType)
+    .def_readwrite("basisLB", &MapOptions::basisLB)
+    .def_readwrite("basisUB", &MapOptions::basisUB)
     .def_readwrite("posFuncType", &MapOptions::posFuncType)
     .def_readwrite("quadType", &MapOptions::quadType)
     .def_readwrite("quadAbsTol", &MapOptions::quadAbsTol)
