### 2022-09-02 8c0ccbc
replaced inf default value by large double
