### 2022-09-03 #4730 @f-t-s
I have never reviewed for joss. Where can I find the paper? Is it just the three-page article linked above?
