### 2022-09-05 072f23c
Added test to catch linearized MapFactory bug.  Fixed bug.
