### 2022-09-06 #4730 @mparno
@f-t-s It would be great if you could help review our submission.  The  [JOSS Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and the [Review Checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) might give you a better idea of what's involved in the review before accepting.
