### 2022-09-06 4a3d190
Added normalization options to orthogonal polynomials and MapFactory.
