+++ src/MapFactory.cpp
@@ -23,7 +23,8 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
             
             if(isinf(opts.basisLB) && isinf(opts.basisUB)){
                
-                MultivariateExpansionWorker<ProbabilistHermite,MemorySpace> expansion(mset);
+                ProbabilistHermite basis1d(opts.normalizePolys);
+                MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
                 switch(opts.posFuncType) {
@@ -38,8 +39,7 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
 
             }else{
 
-                LinearizedBasis<ProbabilistHermite> basis1d(opts.basisLB, opts.basisUB);
-
+                LinearizedBasis<ProbabilistHermite> basis1d(ProbabilistHermite(opts.normalizePolys), opts.basisLB, opts.basisUB);
                 MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
@@ -59,7 +59,8 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
 
             if(isinf(opts.basisLB) && isinf(opts.basisUB)){
                 
-                MultivariateExpansionWorker<PhysicistHermite, MemorySpace> expansion(mset);
+                PhysicistHermite basis1d(opts.normalizePolys);
+                MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
                 switch(opts.posFuncType) {
@@ -74,7 +75,7 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
 
             }else{
 
-                LinearizedBasis<PhysicistHermite> basis1d(opts.basisLB, opts.basisUB);
+                LinearizedBasis<PhysicistHermite> basis1d(PhysicistHermite(opts.normalizePolys), opts.basisLB, opts.basisUB);
                 MultivariateExpansionWorker<decltype(basis1d), MemorySpace> expansion(mset, basis1d);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
@@ -91,7 +92,7 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
 
         }else if(opts.basisType==BasisTypes::HermiteFunctions){
             if(isinf(opts.basisLB) && isinf(opts.basisUB)){
-                    
+                
                 MultivariateExpansionWorker<HermiteFunction, MemorySpace> expansion(mset);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
@@ -130,8 +131,9 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
         if(opts.basisType==BasisTypes::ProbabilistHermite){
             
             if(isinf(opts.basisLB) && isinf(opts.basisUB)){
-
-                MultivariateExpansionWorker<ProbabilistHermite,MemorySpace> expansion(mset);
+                
+                ProbabilistHermite basis1d(opts.normalizePolys);
+                MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
[…]
