### 2022-09-06 8f155ce
Added basisNorm option to language bindings.
