+++ tests/Test_TriangularMap.cpp
@@ -11,7 +11,7 @@ TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_
 
     MapOptions options;
     options.basisType = BasisTypes::ProbabilistHermite;
-    options.normalizePolys = false;
+    options.basisNorm = false;
     
     unsigned int numBlocks = 3;
     unsigned int maxDegree = 2;
