+++ src/MapFactory.cpp
@@ -23,7 +23,7 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
             
             if(isinf(opts.basisLB) && isinf(opts.basisUB)){
                
-                ProbabilistHermite basis1d(opts.normalizePolys);
+                ProbabilistHermite basis1d(opts.basisNorm);
                 MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
@@ -39,7 +39,7 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
 
             }else{
 
-                LinearizedBasis<ProbabilistHermite> basis1d(ProbabilistHermite(opts.normalizePolys), opts.basisLB, opts.basisUB);
+                LinearizedBasis<ProbabilistHermite> basis1d(ProbabilistHermite(opts.basisNorm), opts.basisLB, opts.basisUB);
                 MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
@@ -59,7 +59,7 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
 
             if(isinf(opts.basisLB) && isinf(opts.basisUB)){
                 
-                PhysicistHermite basis1d(opts.normalizePolys);
+                PhysicistHermite basis1d(opts.basisNorm);
                 MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
@@ -75,7 +75,7 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
 
             }else{
 
-                LinearizedBasis<PhysicistHermite> basis1d(PhysicistHermite(opts.normalizePolys), opts.basisLB, opts.basisUB);
+                LinearizedBasis<PhysicistHermite> basis1d(PhysicistHermite(opts.basisNorm), opts.basisLB, opts.basisUB);
                 MultivariateExpansionWorker<decltype(basis1d), MemorySpace> expansion(mset, basis1d);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
@@ -132,7 +132,7 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
             
             if(isinf(opts.basisLB) && isinf(opts.basisUB)){
                 
-                ProbabilistHermite basis1d(opts.normalizePolys);
+                ProbabilistHermite basis1d(opts.basisNorm);
                 MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
                 std::shared_ptr<ConditionalMapBase<MemorySpace>> output;
 
@@ -148,7 +148,7 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
 
             }else{
                 
-                LinearizedBasis<ProbabilistHermite> basis1d(ProbabilistHermite(opts.normalizePolys), opts.basisLB, opts.basisUB);
+                LinearizedBasis<ProbabilistHermite> basis1d(ProbabilistHermite(opts.basisNorm), opts.basisLB, opts.basisUB);
 
                 MultivariateExpansionWorker<decltype(basis1d),MemorySpace> expansion(mset, basis1d);
[…]
