+++ MParT/MapOptions.h
@@ -92,7 +92,7 @@ namespace mpart{
         bool contDeriv = true;
 
         /** If orthogonal polynomial basis functions should be normalized. */
-        bool normalizePolys = true;
+        bool basisNorm = true;
     }; 
 };
 
