### 2022-09-07 f04756e
Added ability to compute gradient with respect to function input.
