+++ MParT/MonotoneComponent.h
@@ -54,13 +54,13 @@ public:
     MonotoneComponent(ExpansionType  const& expansion,
                       QuadratureType const& quad,
                       bool useContDeriv=true) : ConditionalMapBase<MemorySpace>(expansion.InputSize(), 1, expansion.NumCoeffs()),
-                                                    _expansion(expansion),
-                                                    _quad(quad),
-                                                    _dim(expansion.InputSize()),
-                                                    _useContDeriv(useContDeriv){};
+                                                    expansion_(expansion),
+                                                    quad_(quad),
+                                                    dim_(expansion.InputSize()),
+                                                    useContDeriv_(useContDeriv){};
 
 
-    virtual std::shared_ptr<ParameterizedFunctionBase<MemorySpace>> GetBaseFunction() override{return std::make_shared<MultivariateExpansion<typename ExpansionType::BasisType, typename ExpansionType::KokkosSpace>>(1,_expansion);};
+    virtual std::shared_ptr<ParameterizedFunctionBase<MemorySpace>> GetBaseFunction() override{return std::make_shared<MultivariateExpansion<typename ExpansionType::BasisType, typename ExpansionType::KokkosSpace>>(1,expansion_);};
 
     /** Override the ConditionalMapBase Evaluate function. */
     virtual void EvaluateImpl(StridedMatrix<const double, MemorySpace> const& pts,
@@ -83,7 +83,7 @@ public:
                                     StridedVector<double,MemorySpace>               output) override
     {
         // First, get the diagonal derivative
-        if(_useContDeriv){
+        if(useContDeriv_){
             ContinuousDerivative(pts, this->savedCoeffs, output);
         }else{
             Kokkos::View<double*,MemorySpace> evals("Evaluations", pts.extent(1));
@@ -101,6 +101,28 @@ public:
         });
     }
 
+    virtual void GradientImpl(StridedMatrix<const double, MemorySpace> const& pts,  
+                              StridedMatrix<const double, MemorySpace> const& sens,
+                              StridedMatrix<double, MemorySpace>              output) override
+    {
+        assert(sens.extent(0)==this->outputDim);
+        assert(sens.extent(1)==pts.extent(1));
+        assert(pts.extent(0)==this->inputDim);
+        assert(output.extent(0)==this->inputDim);
+        assert(output.extent(1)==pts.extent(1));
+
+        Kokkos::View<double*,MemorySpace> evals("Map output", pts.extent(1));
+
+        InputJacobian(pts, this->savedCoeffs, evals, output);
+
+        // Scale each column by the sensitivity
+        auto policy = Kokkos::RangePolicy<typename MemoryToExecution<MemorySpace>::Space>(0,pts.extent(1));
+        Kokkos::parallel_for(policy, KOKKOS_CLASS_LAMBDA (unsigned int ptInd) {
+            for(unsigned int i=0; i<this->inputDim; ++i)
+                output(i,ptInd) *= sens(0,ptInd);
+        });
+    }
+    
     virtual void CoeffGradImpl(StridedMatrix<const double, MemorySpace> const& pts,
                                StridedMatrix<const double, MemorySpace> const& sens,
                                StridedMatrix<double, MemorySpace>              output) override
@@ -130,7 +152,7 @@ public:
         Kokkos::View<double*,MemorySpace> derivs("Diagonal Derivative", pts.extent(1));
 
         // First, get the diagonal derivative
[…]
