### 2022-09-07 e0ef140
Added language bindings for input gradient functionality.
