### 2022-09-07 5443481
Working on Gradient functions now. Nothing tested
