### 2022-09-08 fce8af9
Added AffineMap class that uses Eigen and works only on host backends.
