### 2022-09-09 86b7cec
Added support for rectangular affine map.
