### 2022-09-09 71c6923
Initial attempt at magma compilation.
