### 2022-09-09 3e560d2
Working on GradCoeff, found bug in SetCoeffs
