### 2022-09-12 030dfef
Added explicit template instatiation to MultivariateExpansionWorker to reduce compile times.
