+++ MParT/MultivariateExpansionWorker.h
@@ -7,8 +7,6 @@
 
 #include "MParT/MultiIndices/MultiIndexSet.h"
 #include "MParT/MultiIndices/FixedMultiIndexSet.h"
-#include "MParT/Utilities/KokkosSpaceMappings.h"
-
 #include "MParT/Utilities/ArrayConversions.h"
 
 namespace mpart{
@@ -16,20 +14,10 @@ namespace mpart{
 template<typename MemorySpace>
 struct MultivariateExpansionMaxDegreeFunctor {
 
-    MultivariateExpansionMaxDegreeFunctor(unsigned int dim, Kokkos::View<unsigned int*, MemorySpace> startPos, Kokkos::View<const unsigned int*, MemorySpace> maxDegreesIn) : dim(dim), startPos(startPos), maxDegrees(maxDegreesIn) 
-    {
-    };
+    MultivariateExpansionMaxDegreeFunctor(unsigned int dim, Kokkos::View<unsigned int*, MemorySpace> startPos, Kokkos::View<const unsigned int*, MemorySpace> maxDegreesIn);
 
-    KOKKOS_FUNCTION void operator()(const unsigned int i, unsigned int& update, const bool final) const{
-        if(final)
-            startPos(i) = update;
 
-        if(i<2*dim){
-            update += maxDegrees(i % dim)+1;
-        }else{
-            update += maxDegrees(dim-1)+1;
-        }
-    };
+    KOKKOS_FUNCTION void operator()(const unsigned int i, unsigned int& update, const bool final) const;
 
     unsigned int dim;
     Kokkos::View<unsigned int*, MemorySpace> startPos;
@@ -39,13 +27,12 @@ struct MultivariateExpansionMaxDegreeFunctor {
 template<typename MemorySpace>
 struct CacheSizeFunctor{
 
-    CacheSizeFunctor(Kokkos::View<unsigned int*, MemorySpace> startPosIn, Kokkos::View<unsigned int*, MemorySpace> cacheSizeIn) : startPos_(startPosIn), cacheSize_(cacheSizeIn){};
+    CacheSizeFunctor(Kokkos::View<unsigned int*, MemorySpace> startPosIn, Kokkos::View<unsigned int*, MemorySpace> cacheSizeIn);
 
     KOKKOS_INLINE_FUNCTION void operator()(const int) const{cacheSize_(0) = startPos_(startPos_.extent(0)-1);};
 
     Kokkos::View<unsigned int*, MemorySpace> startPos_;
     Kokkos::View<unsigned int*, MemorySpace> cacheSize_;
-
 };
 
 /**
@@ -82,25 +69,13 @@ public:
     using BasisType = BasisEvaluatorType;
     using KokkosSpace = MemorySpace;
 
-    MultivariateExpansionWorker() : dim_(0), multiSet_(FixedMultiIndexSet<MemorySpace>(1,0)){};
+    MultivariateExpansionWorker();
 
     MultivariateExpansionWorker(MultiIndexSet const& multiSet,
-                                BasisEvaluatorType const& basis1d = BasisEvaluatorType()) : MultivariateExpansionWorker(multiSet.Fix(), basis1d){};
+                                BasisEvaluatorType const& basis1d = BasisEvaluatorType());
 
     MultivariateExpansionWorker(FixedMultiIndexSet<MemorySpace> const& multiSet,
-                                BasisEvaluatorType const& basis1d = BasisEvaluatorType()) : dim_(multiSet.Length()),
-                                                                                      multiSet_(multiSet),
-                                                                                      basis1d_(basis1d),
-                                                                                      startPos_("Indices for start of 1d basis evaluations", 2*multiSet.Length()+2),
[…]
