+++ src/MultivariateExpansionWorker.cpp
@@ -1,0 +1,324 @@
+#include "MParT/MultivariateExpansionWorker.h"
+#include "MParT/Utilities/ArrayConversions.h"
+
+#include "MParT/Utilities/KokkosSpaceMappings.h"
+
+#include "MParT/Utilities/ArrayConversions.h"
+#include "MParT/OrthogonalPolynomial.h"
+#include "MParT/LinearizedBasis.h"
+#include "MParT/HermiteFunction.h"
+
+using namespace mpart;
+
+template<typename MemorySpace>
+MultivariateExpansionMaxDegreeFunctor<MemorySpace>::MultivariateExpansionMaxDegreeFunctor(unsigned int dim, Kokkos::View<unsigned int*, MemorySpace> startPos, Kokkos::View<const unsigned int*, MemorySpace> maxDegreesIn) : dim(dim), startPos(startPos), maxDegrees(maxDegreesIn) 
+{
+};
+
+template<typename MemorySpace>
+void MultivariateExpansionMaxDegreeFunctor<MemorySpace>::operator()(const unsigned int i, unsigned int& update, const bool final) const{
+    if(final)
+        startPos(i) = update;
+
+    if(i<2*dim){
+        update += maxDegrees(i % dim)+1;
+    }else{
+        update += maxDegrees(dim-1)+1;
+    }
+};
+
+template<typename MemorySpace>
+CacheSizeFunctor<MemorySpace>::CacheSizeFunctor(Kokkos::View<unsigned int*, MemorySpace> startPosIn, Kokkos::View<unsigned int*, MemorySpace> cacheSizeIn) : startPos_(startPosIn), cacheSize_(cacheSizeIn){};
+
+
+template<class BasisEvaluatorType, typename MemorySpace>
+MultivariateExpansionWorker<BasisEvaluatorType, MemorySpace>::MultivariateExpansionWorker() : dim_(0), multiSet_(FixedMultiIndexSet<MemorySpace>(1,0)){};
+
+template<class BasisEvaluatorType, typename MemorySpace>
+MultivariateExpansionWorker<BasisEvaluatorType, MemorySpace>::MultivariateExpansionWorker(MultiIndexSet const& multiSet,
+                                BasisEvaluatorType const& basis1d) : MultivariateExpansionWorker(multiSet.Fix(), basis1d){};
+
+template<class BasisEvaluatorType, typename MemorySpace>
+MultivariateExpansionWorker<BasisEvaluatorType, MemorySpace>::MultivariateExpansionWorker(FixedMultiIndexSet<MemorySpace> const& multiSet,
+                                BasisEvaluatorType const& basis1d) : dim_(multiSet.Length()),
+                                                                                      multiSet_(multiSet),
+                                                                                      basis1d_(basis1d),
+                                                                                      startPos_("Indices for start of 1d basis evaluations", 2*multiSet.Length()+2),
+                                                                                      maxDegrees_(multiSet_.MaxDegrees())
+{
+    Kokkos::parallel_scan(Kokkos::RangePolicy<typename MemoryToExecution<MemorySpace>::Space>(0,2*dim_+2), MultivariateExpansionMaxDegreeFunctor<MemorySpace>(dim_,startPos_, maxDegrees_));
+
+    // Compute the cache size and store in a double on the host
+    Kokkos::View<unsigned int*, MemorySpace> dCacheSize("Temporary cache size",1);
+    Kokkos::parallel_for(Kokkos::RangePolicy<typename MemoryToExecution<MemorySpace>::Space>(0,1), CacheSizeFunctor<MemorySpace>(startPos_, dCacheSize));
+    cacheSize_ = ToHost(dCacheSize)(0);
+};
+
[…]
