### 2022-09-12 caeeb3a
Attempt to speed up compilation of MapFactory by spreading it across several files.
