### 2022-09-13 bd94a63
added input gradient related functions
