### 2022-09-14 0ba15d1
Fixing MapOptions for Julia Bindings to work for MacOS
