+++ bindings/julia/src/MapOptions.cpp
@@ -7,28 +7,33 @@ using namespace mpart;
 
 JLCXX_MODULE BasisType_julia_module(jlcxx::Module& mod) {
     // BasisTypes
-    mod.add_bits<BasisTypes>("__BasisTypes", jlcxx::julia_type("CppEnum"));
-    mod.set_const("ProbabilistHermite", BasisTypes::ProbabilistHermite);
-    mod.set_const("PhysicistHermite", BasisTypes::PhysicistHermite);
-    mod.set_const("HermiteFunctions", BasisTypes::HermiteFunctions);
+    
 }
 
 JLCXX_MODULE PosFuncType_julia_module(jlcxx::Module &mod) {
-    // PosFuncTypes
-    mod.add_bits<PosFuncTypes>("__PosFuncTypes", jlcxx::julia_type("CppEnum"));
-    mod.set_const("Exp", PosFuncTypes::Exp);
-    mod.set_const("SoftPlus", PosFuncTypes::SoftPlus);
 }
 
 JLCXX_MODULE QuadType_julia_module(jlcxx::Module &mod) {
-    // QuadTypes
-    mod.add_bits<QuadTypes>("__QuadTypes", jlcxx::julia_type("CppEnum"));
-    mod.set_const("ClenshawCurtis", QuadTypes::ClenshawCurtis);
-    mod.set_const("AdaptiveSimpson", QuadTypes::AdaptiveSimpson);
-    mod.set_const("AdaptiveClenshawCurtis", QuadTypes::AdaptiveClenshawCurtis);
 }
 
 void mpart::binding::MapOptionsWrapper(jlcxx::Module &mod) {
+    // BasisTypes
+    mod.add_bits<BasisTypes>("__BasisTypes", jlcxx::julia_type("CppEnum"));
+    mod.set_const("__ProbabilistHermite", BasisTypes::ProbabilistHermite);
+    mod.set_const("__PhysicistHermite", BasisTypes::PhysicistHermite);
+    mod.set_const("__HermiteFunctions", BasisTypes::HermiteFunctions);
+    
+    // PosFuncTypes
+    mod.add_bits<PosFuncTypes>("__PosFuncTypes", jlcxx::julia_type("CppEnum"));
+    mod.set_const("__Exp", PosFuncTypes::Exp);
+    mod.set_const("__SoftPlus", PosFuncTypes::SoftPlus);
+
+    // QuadTypes
+    mod.add_bits<QuadTypes>("__QuadTypes", jlcxx::julia_type("CppEnum"));
+    mod.set_const("__ClenshawCurtis", QuadTypes::ClenshawCurtis);
+    mod.set_const("__AdaptiveSimpson", QuadTypes::AdaptiveSimpson);
+    mod.set_const("__AdaptiveClenshawCurtis", QuadTypes::AdaptiveClenshawCurtis);
+
     // MapOptions
     mod.add_type<MapOptions>("__MapOptions")
         .method("__basisType!", [](MapOptions &opts, unsigned int basis){ opts.basisType = static_cast<BasisTypes>(basis); })
