### 2022-09-14 fc548c5
Removing unnecessary code
