+++ bindings/julia/src/MultiIndex.cpp
@@ -43,6 +43,7 @@ void mpart::binding::MultiIndexWrapper(jlcxx::Module &mod) {
         .method("IsExpandable", &MultiIndexSet::IsExpandable)
         .method("NumActiveForward", &MultiIndexSet::NumActiveForward)
         .method("NumForward", &MultiIndexSet::NumForward)
+        .method("Size", &MultiIndexSet::Size)
     ;
 
     mod.method("MultiIndexSet", [](jlcxx::ArrayRef<int,2> idxs) {
@@ -54,7 +55,6 @@ void mpart::binding::MultiIndexWrapper(jlcxx::Module &mod) {
     mod.set_override_module(jl_base_module);
     mod.method("sum", [](MultiIndex const& idx){ return idx.Sum(); });
     mod.method("setindex!", [](MultiIndex& idx, unsigned int val, unsigned int ind) { return idx.Set(ind, val); });
-    // mod.method("setindex!", [](MultiIndexSet& idx, MultiIndex val, unsigned int ind) { return idx[ind] = val; });
     mod.method("getindex", [](MultiIndex const& idx, unsigned int ind) { return idx.Get(ind); });
     mod.method("maximum", [](MultiIndex const& idx){ return idx.Max(); });
     mod.method("String", [](MultiIndex const& idx){ return idx.String(); });
