### 2022-09-14 eacc7ba
WIP: Added AffineMap and MAGMA usage, but currently hangs on dgemm.
