+++ src/AffineFunction.cpp
@@ -1,0 +1,223 @@
+#include "MParT/AffineFunction.h"
+#include "MParT/Utilities/KokkosSpaceMappings.h"
+
+#include "MParT/Utilities/ArrayConversions.h"
+
+#if defined(MPART_ENABLE_GPU)
+#include "magma_v2.h"
+#endif 
+
+using namespace mpart;
+
+template<typename MemorySpace>
+AffineFunction<MemorySpace>::AffineFunction(StridedVector<double,MemorySpace> b) : ParameterizedFunctionBase<MemorySpace>(b.size(),b.size(),0), 
+                                                                  b_(b)
+{}
+
+template<typename MemorySpace>
+AffineFunction<MemorySpace>::AffineFunction(StridedMatrix<double,MemorySpace> A) : ParameterizedFunctionBase<MemorySpace>(A.extent(1),A.extent(0),0),
+                                                                  A_(A)
+{
+    // Make sure the columns of A have contiguous memory
+    if(A_.stride_0() != 1){
+        A_ = Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>("A col major", A_.extent(0), A_.extent(1));
+        Kokkos::deep_copy(A_,A);
+    }
+    ldA = A_.stride_1();
+}
+
+
+template<typename MemorySpace>
+AffineFunction<MemorySpace>::AffineFunction(StridedMatrix<double,MemorySpace> A,
+                                  StridedVector<double,MemorySpace> b) : ParameterizedFunctionBase<MemorySpace>(A.extent(1),A.extent(0),0),
+                                                                  A_(A),
+                                                                  b_(b)
+{   
+    assert(A_.extent(0) == b_.extent(0));
+
+    // Make sure the columns of A have contiguous memory
+    if(A_.stride_0() != 1){
+        A_ = Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>("A col major", A_.extent(0), A_.extent(1));
+        Kokkos::deep_copy(A_,A);
+    }
+    ldA = A_.stride_1();
+}
+
+
+template<>
+void AffineFunction<Kokkos::HostSpace>::EvaluateImpl(StridedMatrix<const double, Kokkos::HostSpace> const& pts,
+                                                     StridedMatrix<double, Kokkos::HostSpace>              output)
+{
+    auto eigOut = KokkosToMat<double>(output);
+    auto eigPts = ConstKokkosToMat<double>(pts);
+
+    // Linear part
+    if(A_.extent(0)>0){
+
+        auto eigA = KokkosToMat<double>(A_);
+        eigOut = eigA * eigPts;
+    }else{
+        eigOut = eigPts;
+    }
+
+    // Bias part
+    if(b_.size()>0){
+        auto eigB = KokkosToVec<double>(b_);
+        eigOut.colwise() += eigB;
+    }
+}
+
+template<typename MemorySpace>
+void AffineFunction<MemorySpace>::CoeffGradImpl(StridedMatrix<const double, MemorySpace> const& pts,  
+                                           StridedMatrix<const double, MemorySpace> const& sens,
+                                           StridedMatrix<double, MemorySpace>              output)
+{
+    return;
+}
+    
+template<>
+void AffineFunction<Kokkos::HostSpace>::GradientImpl(StridedMatrix<const double, Kokkos::HostSpace> const& pts,  
+                                                StridedMatrix<const double, Kokkos::HostSpace> const& sens,
+                                                StridedMatrix<double, Kokkos::HostSpace>              output)
+{
+    auto eigOut = KokkosToMat<double>(output);
+    auto eigSens = ConstKokkosToMat<double>(sens);
+
[…]
