+++ tests/Test_AffineFunction.cpp
@@ -1,0 +1,296 @@
+#include <catch2/catch_all.hpp>
+
+#include "MParT/AffineFunction.h"
+
+using namespace mpart;
+using namespace Catch;
+
+
+TEST_CASE( "Testing Shift-only AffineFunction", "[ShiftFunction]" ) {
+
+    Kokkos::View<double*, Kokkos::HostSpace> b("b", 2);
+    
+    b(0) = 1.0;
+    b(1) = 2.0;
+
+    auto map = std::make_shared<AffineFunction<Kokkos::HostSpace>>(b);
+
+    unsigned int numPts = 10;
+    Kokkos::View<double**, Kokkos::HostSpace> pts("Point", 2, numPts);
+    for(unsigned int i=0; i<numPts; ++i){
+        pts(0,i) = double(i)/double(numPts-1);
+        pts(1,i) = double(i)/double(numPts-1);
+    }
+
+    // Test the forward evaluation
+    Kokkos::View<double**, Kokkos::HostSpace> evals = map->Evaluate(pts);
+    for(unsigned int i=0; i<numPts; ++i){
+        double trueOut1 = pts(0,i) + b(0);
+        double trueOut2 = pts(0,i) + b(1);
+        CHECK(evals(0,i)==Approx(trueOut1).epsilon(1e-14));
+        CHECK(evals(1,i)==Approx(trueOut2).epsilon(1e-14));
+    }
+
+}
+
+TEST_CASE( "Testing Linear-only AffineFunction", "[LinearFunction]" ) {
+
+    Kokkos::View<double**, Kokkos::HostSpace> A("A", 2,2);
+    
+    A(0,0) = 2.0;
+    A(1,0) = 1.0;
+    A(0,1) = 1.0;
+    A(1,1) = 4.0;
+
+    auto map = std::make_shared<AffineFunction<Kokkos::HostSpace>>(A);
+
+    unsigned int numPts = 3;
+    Kokkos::View<double**, Kokkos::HostSpace> pts("Point", 2, numPts);
+    for(unsigned int i=0; i<numPts; ++i){
+        pts(0,i) = double(i)/double(numPts-1);
+        pts(1,i) = double(i)/double(numPts-1);
+    }
+
+    // Test the forward evaluation
+    Kokkos::View<double**, Kokkos::HostSpace> evals = map->Evaluate(pts);
+
+    for(unsigned int i=0; i<numPts; ++i){
+        double trueOut1 = A(0,0)*pts(0,i) + A(0,1)*pts(1,i);
+        double trueOut2 = A(1,0)*pts(0,i) + A(1,1)*pts(1,i);
+        CHECK(evals(0,i)==Approx(trueOut1).epsilon(1e-14));
+        CHECK(evals(1,i)==Approx(trueOut2).epsilon(1e-14));
+    }
+
+}
+
+
+TEST_CASE( "Testing Rectangular AffineFunction", "[Rectangular LinearFunction]" ) {
+
+    Kokkos::View<double**, Kokkos::HostSpace> A("A", 2,3);
+    
+    /* 
+     A = [[2.0, 3.0, 1.0],
+          [1.0, 1.0, 4.0]]
+    */
+    A(0,0) = 2.0;
+    A(1,0) = 1.0;
+    A(0,1) = 3.0;
+    A(1,1) = 1.0;
+    A(0,2) = 1.0;
+    A(1,2) = 4.0;
+    
+
+    auto map = std::make_shared<AffineFunction<Kokkos::HostSpace>>(A);
+    REQUIRE(map->inputDim==3);
+    REQUIRE(map->outputDim==2);
+
[…]
