### 2022-09-15 aaefee9
Working on LogDeterminantInputGrad
