+++ src/TriangularMap.cpp
@@ -288,10 +288,10 @@ void TriangularMap<MemorySpace>::LogDeterminantInputGradImpl(StridedMatrix<const
 
         subPts = Kokkos::subview(pts, std::make_pair(0,int(comps_.at(i)->inputDim)), Kokkos::ALL());
 
-        subOut = Kokkos::subview(output, std::make_pair(startParamDim,int(startParamDim+comps_.at(i)->numCoeffs)), Kokkos::ALL());
+        subOut = Kokkos::subview(output, std::make_pair(startParamDim,int(startParamDim+comps_.at(i)->inputDim)), Kokkos::ALL());
         comps_.at(i)->LogDeterminantInputGradImpl(subPts, subOut);
 
-        startParamDim += comps_.at(i)->numCoeffs;
+        startParamDim += comps_.at(i)->inputDim;
     }
 }
 
