### 2022-09-15 ee19a76
ComposedMap working. But is a bit slow
