+++ src/ComposedMap.cpp
@@ -222,53 +222,52 @@ void ComposedMap<MemorySpace>::CoeffGradImpl(StridedMatrix<const double, MemoryS
 }
 
 template<typename MemorySpace>
+void ComposedMap<MemorySpace>::EvaluateUntilK( int k, 
+                                        StridedMatrix<const double, MemorySpace> const& pts,
+                                        StridedMatrix<double, MemorySpace> intPts, 
+                                        StridedMatrix<double, MemorySpace> output){
+
+
+    Kokkos::deep_copy(intPts, pts);
+    for(int j = 0; j<k; j++){
+        // x = T_j(x)
+        comps_.at(j)->EvaluateImpl(intPts, output);
+        Kokkos::deep_copy(intPts, output);
+    }
+
+}
+
+
+template<typename MemorySpace>
 void ComposedMap<MemorySpace>::LogDeterminantCoeffGradImplUpdate( int compInd,
                                         int termInd, 
                                         StridedMatrix<const double, MemorySpace> const& pts, 
                                         StridedMatrix<double, MemorySpace> output){
 
-    
+
     // compute first sensitive vector 
     Kokkos::View<double**, MemorySpace>  intPts1("intermediate points 1", pts.extent(0), pts.extent(1));
     Kokkos::View<double**, MemorySpace>  intPts2("intermediate points 2", pts.extent(0), pts.extent(1));
+    Kokkos::View<double**, MemorySpace>  sens1("first sens (dim x numPts)", pts.extent(0), pts.extent(1));
+    Kokkos::View<double**, MemorySpace>  sens2("second sens (dim x numPts)", pts.extent(0), pts.extent(1));
 
-    Kokkos::deep_copy(intPts1, pts);
-    for(int j = 0; j<termInd; j++){
-        // x = T_j(x)
-        comps_.at(j)->EvaluateImpl(intPts1, intPts2);
-        Kokkos::deep_copy(intPts1, intPts2);
-    }
-
-    Kokkos::View<double**, MemorySpace>  intOutput("intOutput", pts.extent(0), pts.extent(1));
-    comps_.at(termInd)->LogDeterminantInputGradImpl(intPts1, intOutput);
-
+    this->EvaluateUntilK(termInd, pts, intPts1, intPts2);
+    comps_.at(termInd)->LogDeterminantInputGradImpl(intPts1, sens1);
 
     // loop through computing input gradient terms
     for(int i = termInd - 1; i>compInd; --i){
-        
-        // x* = T_{i-1} o ... o T_1(x)
 
-        // reset intPts1 to initial pts
-        Kokkos::deep_copy(intPts1, pts);
-        for(int j = 0; j<i; j++){
-            // x = T_j(x)
-            comps_.at(j)->EvaluateImpl(intPts1, intPts2);
-            Kokkos::deep_copy(intPts1, intPts2);
-        }
+        // x* = T_{i-1} o ... o T_1(x)
+        this->EvaluateUntilK(i, pts, intPts1, intPts2);
 
         //s_{i-1}^T = s_{i}^T J_i(x*)
-        comps_.at(i)->GradientImpl(intPts1, intOutput, output);
-        Kokkos::deep_copy(intOutput, output);
+        comps_.at(i)->GradientImpl(intPts1, sens1, sens2);
+        Kokkos::deep_copy(sens1, sens2);
     }
 
     // last part, coeff gradient of comp.at(compInd)
-    Kokkos::deep_copy(intPts1, pts);
-    for(int j = 0; j<compInd; j++){
-        // x = T_j(x)
-        comps_.at(j)->EvaluateImpl(intPts1, intPts2);
-        Kokkos::deep_copy(intPts1, intPts2);
[…]
