### 2022-09-16 dea8fb2
Version of AffineMap and AffineFunction using MAGMA as a device backend.  Compiles and should work, but hangs on queue_create and queue_sync, which I suspect is a result of how MAGMA is being compiled internally by MParT.  Copying the MAGMA source code directly into AffineMap.cpp seems to work.
