### 2022-09-16 97887d5
Initial attempt at conversion to cublas and cusolver.
