+++ src/Initialization.cpp
@@ -6,7 +6,8 @@ using namespace mpart;
 void mpart::Finalize()
 {
 #if defined(MPART_ENABLE_GPU)
-    magma_finalize();
+    cublasDestroy(GetInitializeStatusObject().GetCublasHandle());
+    cusolverDnDestroy(GetInitializeStatusObject().GetCusolverHandle());
 #endif 
 
     Kokkos::finalize();
