+++ src/AffineMap.cpp
@@ -3,10 +3,7 @@
 
 #include "MParT/Utilities/ArrayConversions.h"
 #include "MParT/Initialization.h"
-
-#if defined(MPART_ENABLE_GPU)
-#include "magma_v2.h"
-#endif 
+ 
 
 using namespace mpart;
 
@@ -172,26 +169,65 @@ void AffineMap<mpart::DeviceSpace>::Factorize()
 
     // Resize the space for storing the LU factorization
     LU_ = Kokkos::View<double**, Kokkos::LayoutLeft, mpart::DeviceSpace>("LU", A_.extent(0), A_.extent(0));
-    pivots_ = Kokkos::View<magma_int_t*, Kokkos::HostSpace>("Pivots", A_.extent(0));
+    pivots_ = Kokkos::View<int*, mpart::DeviceSpace>("Pivots", A_.extent(0));
 
-    magma_int_t ldLU = LU_.stride_1();
-    magma_int_t info;
+    int ldLU = LU_.stride_1();
+    int info;
 
     // Copy the right block of the matrix A_ into LU_
     Kokkos::deep_copy(LU_, Kokkos::subview(A_, Kokkos::ALL(), std::make_pair(ncols-nrows, int(A_.extent(1)))));
+
+    // Set up cuSolver options
+    CUSOLVER_CHECK(cusolverDnCreateParams(&params));
+    CUSOLVER_CHECK(cusolverDnSetAdvOptions(params, CUSOLVERDN_GETRF, CUSOLVER_ALG_0));
     
-    // Factorize the right block of the matrix
-    magma_dgetrf_gpu(static_cast<magma_int_t>(LU_.extent(0)), 
-                     static_cast<magma_int_t>(LU_.extent(1)), 
-                     reinterpret_cast<magmaDouble_ptr>(LU_.data()), 
-                     ldLU, 
-                     pivots_.data(), 
-                     &info);
+    size_t d_workSize, h_workSize;
+    cusolverDnXgetrf_bufferSize(GetInitializeStatusObject().GetCusolverHandle(), 
+                                params,
+                                LU_.extent(0),
+                                LU_.extent(1), 
+                                traits<double>::cuda_data_type, 
+                                LU_.data(),
+                                ldLU, 
+                                traits<double>::cuda_data_type, 
+                                &d_workSize, 
+                                &h_workSize);
+
+
+    Kokkos::View<double*, mpart::DeviceSpace> d_workspace("LU Workspace", d_workSize);
+    Kokkos::View<double*, Kokkos::HostSpace> h_workspace("LU Workspace", h_workSize);
+    Kokkos::View<double*, mpart::DeviceSpace> d_info("Info",1);
+
+    cusolverDnXgetrf(GetInitializeStatusObject().GetCusolverHandle(), 
+                     params, 
+                     LU_.extent(0), 
+                     LU_.extent(1), 
+                     traits<double>::cuda_data_type,
+                     LU_.data(), ldLU, 
+                     pivots_,
+                     traits<double>::cuda_data_type, 
+                     d_workspace,
+                     d_workSize, 
+                     h_workspace, 
+                     h_workSize,
+                     d_info);
+    
+    info = ToHost(d_info)(0);
+
+    // cusolverStatus_t res = 
+    // cusolverDnDgetrf(GetInitializeStatusObject().GetCublasHandle(),
+    //                  LU_.extent(0),
+    //                  LU_.extent(1),
+    //                  LU_.data(),
+    //                  ldLU,
+    //                  workspace.data(),
+    //                  pivots_.data(),
+    //                  &info);
 
     // Error handling
     if(info<0){
         std::stringstream msg;
[…]
