+++ docs/source/installation.rst
@@ -2,7 +2,6 @@
 
 Installation
 ===================
-
 .. card:: Install with Conda
 
     It is possible to install the main c++ MParT library and its python wrapper via `conda-forge <https://anaconda.org/conda-forge/mpart>`__:
@@ -11,9 +10,16 @@ Installation
 
         conda install -c conda-forge mpart
 
-    Julia, Matlab, and CUDA are currently only supported when compiling from source. 
+    Matlab, and CUDA are currently only supported when compiling from source.
 
+.. card:: Install with Julia
 
+    It is also possible to install MParT with Julia bindings using the Julia package manager:
+
+    .. code-block:: julia
+
+        using Pkg
+        Pkg.add("MParT")
 
 .. _compiling_source:
 
@@ -34,7 +40,7 @@ MParT uses CMake to handle dependencies and compiler configurations.   A basic b
    ..
    make install
 
-This will compile the main c++ :code:`mpart` library as well as any other language bindings that can be automatically configured.  If you are compiling on a multicore machine, you can use :code:`make -j N_JOBS install`, where :code:`N_JOBS` is the number of processes the computer can compile with in parallel.  
+This will compile the main c++ :code:`mpart` library as well as any other language bindings that can be automatically configured.  If you are compiling on a multicore machine, you can use :code:`make -j N_JOBS install`, where :code:`N_JOBS` is the number of processes the computer can compile with in parallel.
 
 This installation should also automatically install and build Kokkos, Eigen, and Catch2, assuming they aren't installed already. If CMake has trouble finding prior installations of these, then you can configuring CMake using:
 
@@ -53,7 +59,7 @@ Feel free to mix and match previous installations of Eigen, Kokkos, Pybind11, an
 
 You can force MParT to use previously installed versions of the dependencies by setting :code:`MPART_FETCH_DEPS=OFF`.  The default value of :code:`MPART_FETCH_DEPS=ON` will allow MParT to download and locally install any external dependencies using CMake's :code:`FetchContent` directive.
 
-Note that if you do not wish to compile bindings for Python, Julia, or Matlab, you can turn off binding compilation by setting the :code:`MPART_<language>=OFF` variable during CMake configuration.  For a default build with only the core c++ library, you can use 
+Note that if you do not wish to compile bindings for Python, Julia, or Matlab, you can turn off binding compilation by setting the :code:`MPART_<language>=OFF` variable during CMake configuration.  For a default build with only the core c++ library, you can use
 
 .. code-block:: bash
 
@@ -84,7 +90,7 @@ Or, with the additional specification of the number of Kokkos threads to use:
    ./RunTests --kokkos-threads=4
 
 
-Environment Paths 
+Environment Paths
 ------------------
 
 The final step is to set the relevant path variables to include the installation of MParT:
@@ -134,7 +140,7 @@ The final step is to set the relevant path variables to include the installation
 
 .. _compiling_julia:
 
-Julia Configuration
+Julia Source Installation
 ------------------
 
 By default, MParT will look for Julia during configuration and will attempt to build the Julia bindings if the Julia :code:`CxxWrap` package is installed.   To install :code:`CxxWrap`, run the following command in your Julia prompt:
@@ -143,30 +149,32 @@ By default, MParT will look for Julia during configuration and will attempt to b
 
     import Pkg; Pkg.add("CxxWrap")
 
[…]
