### 2022-09-16 3cecd5b
Working AffineMap class with cublas and cusparse.
