+++ tests/Test_AffineMap.cpp
@@ -327,10 +327,10 @@ TEST_CASE( "Testing Rectangular AffineMap on Device", "[DeviceRectangularLinearM
     for(unsigned int i=0; i<numPts; ++i){
         double trueOut1 = hA(0,0)*hpts(0,i) + hA(0,1)*hpts(1,i) + hA(0,2)*hpts(2,i);
         double trueOut2 = hA(1,0)*hpts(0,i) + hA(1,1)*hpts(1,i) + hA(1,2)*hpts(2,i);
-        CHECK(hevals(0,i)==Approx(trueOut1).epsilon(1e-14));
-        CHECK(hevals(1,i)==Approx(trueOut2).epsilon(1e-14));
-        CHECK(hpts2(0,i)==Approx(hpts(1,i)).epsilon(1e-14));
-        CHECK(hpts2(1,i)==Approx(hpts(2,i)).epsilon(1e-14));
+        CHECK(hevals(0,i)==Approx(trueOut1).epsilon(1e-14).margin(1e-14));
+        CHECK(hevals(1,i)==Approx(trueOut2).epsilon(1e-14).margin(1e-14));
+        CHECK(hpts2(0,i)==Approx(hpts(1,i)).epsilon(1e-14).margin(1e-14));
+        CHECK(hpts2(1,i)==Approx(hpts(2,i)).epsilon(1e-14).margin(1e-14));
     }
 
 }
@@ -376,12 +376,12 @@ TEST_CASE( "Testing Full AffineMap on Device", "[DeviceFullAffineMap]" ) {
     auto hpts2 = ToHost(dpts2);
     
     for(unsigned int i=0; i<numPts; ++i){
-        double trueOut1 = hA(0,0)*hpts(0,i) + hA(0,1)*hpts(1,i);
-        double trueOut2 = hA(1,0)*hpts(0,i) + hA(1,1)*hpts(1,i);
-        CHECK(hevals(0,i)==Approx(trueOut1).epsilon(1e-14));
-        CHECK(hevals(1,i)==Approx(trueOut2).epsilon(1e-14));
-        CHECK(hpts2(0,i)==Approx(hpts(0,i)).epsilon(1e-14));
-        CHECK(hpts2(1,i)==Approx(hpts(1,i)).epsilon(1e-14));
+        double trueOut1 = hA(0,0)*hpts(0,i) + hA(0,1)*hpts(1,i) + hb(0);
+        double trueOut2 = hA(1,0)*hpts(0,i) + hA(1,1)*hpts(1,i) + hb(1);
+        CHECK(hevals(0,i)==Approx(trueOut1).epsilon(1e-14).margin(1e-14));
+        CHECK(hevals(1,i)==Approx(trueOut2).epsilon(1e-14).margin(1e-14));
+        CHECK(hpts2(0,i)==Approx(hpts(0,i)).epsilon(1e-14).margin(1e-14));
+        CHECK(hpts2(1,i)==Approx(hpts(1,i)).epsilon(1e-14).margin(1e-14));
     }
 
 }
