+++ tests/Test_ComposedMap.cpp
@@ -1,0 +1,219 @@
+#include <catch2/catch_all.hpp>
+
+#include "MParT/ComposedMap.h"
+#include "MParT/MapFactory.h"
+
+using namespace mpart;
+using namespace Catch;
+using MemorySpace = Kokkos::HostSpace;
+
+TEST_CASE( "Testing 5 layer composed map", "[ComposedMap_Constructor]" ) {
+
+    MapOptions options;
+    options.basisType = BasisTypes::ProbabilistHermite;
+    options.basisNorm = false;
+    
+    unsigned int dim = 2;
+    unsigned int numMaps = 2;
+    unsigned int order = 2;
+    unsigned int coeffSize = 0;
+    
+
+    std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> maps(numMaps);
+    for(unsigned int i=0;i<numMaps;++i){
+
+        maps.at(i) = MapFactory::CreateTriangular<Kokkos::HostSpace>(dim, dim, order, options);
+        coeffSize += maps.at(i)->numCoeffs;
+    }
+
+    std::shared_ptr<ConditionalMapBase<MemorySpace>> composedMap = std::make_shared<ComposedMap<MemorySpace>>(maps);
+
+    CHECK(composedMap->outputDim == dim);
+    CHECK(composedMap->inputDim == dim);
+    CHECK(composedMap->numCoeffs == coeffSize);
+
+
+    Eigen::RowVectorXd coeffs(composedMap->numCoeffs);
+    for(unsigned int i=0; i<composedMap->numCoeffs; ++i)
+        coeffs(i) = 0.1*(i+1);
+        
+    SECTION("Coefficients"){
+        
+        // Set the coefficients of the triangular map
+        composedMap->SetCoeffs(coeffs);
+
+        // Now make sure that the coefficients of each block were set
+        unsigned int cumCoeffInd = 0;
+        for(unsigned int i=0; i<numMaps; ++i){
+            for(unsigned int j=0; j<maps.at(i)->numCoeffs; ++j){
+                CHECK(maps.at(i)->Coeffs()(j) == coeffs(cumCoeffInd)); // Values of coefficients should be equal
+                CHECK(&maps.at(i)->Coeffs()(j) == &composedMap->Coeffs()(cumCoeffInd)); // Memory location should also be the same (no copy)
+                cumCoeffInd++;
+            }
+        }
+    }
+
+
+    unsigned int numSamps = 10;
+    Kokkos::View<double**, Kokkos::HostSpace> in("Map Input", dim, numSamps);
+    //Eigen::RowMatrixXd in(dim, numSamps);
+    for(unsigned int i=0; i<dim; ++i){
+        for(unsigned int j=0; j<numSamps; ++j){
+            in(i,j) = double(i)/(dim) + double(j)/numSamps;
+        }
+    }
+
+    composedMap->SetCoeffs(coeffs);
+    auto out = composedMap->Evaluate(in);
+    // Kokkos::View<double*,Kokkos::HostSpace> coeffs("Coefficients", composedMap->numCoeffs);
+    // // intermediate output
+    // Kokkos::View<double**, MemorySpace> intOutput("intermediate output", output.extent(0), output.extent(1));
+    
+    // // Copy points to output, then output = map(output) looped over each component
+    // Kokkos::deep_copy(output, pts);
+    // for(unsigned int i=0; i<comps_.size(); ++i){
+        
+    //     comps_.at(i)->EvaluateImpl(output, intOutput);
+    //     Kokkos::deep_copy(output,intOutput);
+    // }
+
+
+    // SECTION("Inverse"){
+
[…]
