### 2022-09-16 2b700cb
Now passing gradient tests
