### 2022-09-16 0ab9149
Adding logdet gradient wrt to input
