### 2022-09-19 #4730 @diehlpk
> The only thing I want to mention is that I am in the process of starting a collaboration with Youssef, in case that constitutes a conflict of interest.

Has the collaboration already started? I would say if the collaboration starts before the review ends, it's a conflict. If you can wait for the collaboration until you finished the review, it would be fine.
