### 2022-09-20 26cff80
Added some linear algebra routines for working with Kokkos views.
