+++ src/Utilities/LinearAlgebra.cpp
@@ -1,0 +1,271 @@
+#include "MParT/Utilities/LinearAlgebra.h"
+#include "MParT/Initialization.h"
+
+#include <Eigen/Dense>
+using namespace mpart;
+
+
+
+template<>
+void mpart::dgemm(double                                     alpha, 
+                  TransposeObject<Kokkos::HostSpace>         A,
+                  TransposeObject<Kokkos::HostSpace>         B,
+                  double                                     beta,
+                  StridedMatrix<double, Kokkos::HostSpace>   C)
+{
+
+    auto eigC = KokkosToMat<double>(C);
+    auto eigB = ConstKokkosToMat<double>(B.view);
+    auto eigA = ConstKokkosToMat<double>(A.view);
+
+    if(A.isTransposed){
+        if(B.isTransposed){
+            eigC = (beta*eigC + alpha* eigA.transpose() * eigB.transpose() ).eval();
+        }else{
+            eigC = (beta*eigC + alpha* eigA.transpose() * eigB ).eval();
+        }
+    }else{
+        if(B.isTransposed){
+            eigC = (beta*eigC + alpha* eigA * eigB.transpose() ).eval();
+        }else{
+            eigC = (beta*eigC + alpha* eigA * eigB ).eval();
+        }
+    }
+
+}
+
+
+template<typename MemorySpace>
+Kokkos::View<double**,Kokkos::LayoutLeft,MemorySpace> PartialPivLU<MemorySpace>::solve(StridedMatrix<const double,MemorySpace> x)
+{
+     Kokkos::View<double**,  Kokkos::LayoutLeft, MemorySpace> output("solution", x.extent(0),x.extent(1));
+     Kokkos::deep_copy(output, x);
+     solveInPlace(output);
+     return output;
+}
+
+
+template<>
+void PartialPivLU<Kokkos::HostSpace>::compute(Kokkos::View<const double**,Kokkos::LayoutLeft,Kokkos::HostSpace> A)
+{
+    auto eigA = ConstKokkosToMat(A);
+    unsigned int nrows = eigA.rows();
+    unsigned int ncols = eigA.cols();
+
+    // Use eigen to compute the LU decomposition in place
+    luSolver_ = std::make_shared<Eigen::PartialPivLU<Eigen::MatrixXd>>(eigA);
+    isComputed = true;
+}
+
+template<>
+void PartialPivLU<Kokkos::HostSpace>::solveInPlace(Kokkos::View<double**,Kokkos::LayoutLeft,Kokkos::HostSpace> x)
+{   
+    auto eigX = KokkosToMat(x);
+    eigX = (luSolver_->permutationP() * eigX).eval();
+    luSolver_->matrixLU().template triangularView<Eigen::UnitLower>().solveInPlace(eigX);
+    luSolver_->matrixLU().template triangularView<Eigen::Upper>().solveInPlace(eigX);
+}
+
+
+template<>
+double PartialPivLU<Kokkos::HostSpace>::determinant() const
+{
+    assert(isComputed); 
+    return luSolver_->determinant();
+}
+
+
+
+#if defined(MPART_ENABLE_GPU)
+
+template<>
+void mpart::dgemm(double                             alpha, 
+           TransposeObject<mpart::DeviceSpace>       A,
+           TransposeObject<mpart::DeviceSpace>       B,
+           double                                    beta,
+           StridedMatrix<double, mpart::DeviceSpace> C)
+{   
+    // Versions of A,B, and C with the column major layout expected by cublas
+    StridedMatrix<const double, mpart::DeviceSpace> ALeft;
+    StridedMatrix<const double, mpart::DeviceSpace> BLeft;
[…]
