### 2022-09-20 007c07a
Fixed error in conversion of const strided kokkos view to eigen matrix.
