### 2022-09-20 98d011a
Updated AffineFunction and AffineMap to use new linear algebra tools.
