+++ src/AffineFunction.cpp
@@ -4,27 +4,21 @@
 #include "MParT/Utilities/ArrayConversions.h"
 #include "MParT/Initialization.h"
 
-#if defined(MPART_ENABLE_GPU)
-#include "magma_v2.h"
-#endif 
+#include "MParT/Utilities/LinearAlgebra.h"
 
 using namespace mpart;
 
+
 template<typename MemorySpace>
 AffineFunction<MemorySpace>::AffineFunction(StridedVector<double,MemorySpace> b) : ParameterizedFunctionBase<MemorySpace>(b.size(),b.size(),0), 
-                                                                  b_(b)
+                                                                  b_(b) 
 {}
 
 template<typename MemorySpace>
 AffineFunction<MemorySpace>::AffineFunction(StridedMatrix<double,MemorySpace> A) : ParameterizedFunctionBase<MemorySpace>(A.extent(1),A.extent(0),0),
                                                                   A_(A)
 {
-    // Make sure the columns of A have contiguous memory
-    if(A_.stride_0() != 1){
-        A_ = Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>("A col major", A_.extent(0), A_.extent(1));
-        Kokkos::deep_copy(A_,A);
-    }
-    ldA = A_.stride_1();
+    assert(A_.extent(0)<=A_.extent(1));
 }
 
 
@@ -34,37 +28,31 @@ AffineFunction<MemorySpace>::AffineFunction(StridedMatrix<double,MemorySpace> A,
                                                                   A_(A),
                                                                   b_(b)
 {   
+    assert(A_.extent(0)<=A_.extent(1));
     assert(A_.extent(0) == b_.extent(0));
-
-    // Make sure the columns of A have contiguous memory
-    if(A_.stride_0() != 1){
-        A_ = Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>("A col major", A_.extent(0), A_.extent(1));
-        Kokkos::deep_copy(A_,A);
-    }
-    ldA = A_.stride_1();
 }
 
 
-template<>
-void AffineFunction<Kokkos::HostSpace>::EvaluateImpl(StridedMatrix<const double, Kokkos::HostSpace> const& pts,
-                                                     StridedMatrix<double, Kokkos::HostSpace>              output)
+template<typename MemorySpace>
+void AffineFunction<MemorySpace>::EvaluateImpl(StridedMatrix<const double, MemorySpace> const& pts,
+                                          StridedMatrix<double, MemorySpace>              output)
 {
-    auto eigOut = KokkosToMat<double>(output);
-    auto eigPts = ConstKokkosToMat<double>(pts);
-
     // Linear part
     if(A_.extent(0)>0){
-
-        auto eigA = KokkosToMat<double>(A_);
-        eigOut = eigA * eigPts;
+        dgemm<MemorySpace>(1.0, A_, pts, 0.0, output);
     }else{
-        eigOut = eigPts;
+        Kokkos::deep_copy(output, pts);
     }
-
+    
     // Bias part
     if(b_.size()>0){
-        auto eigB = KokkosToVec<double>(b_);
-        eigOut.colwise() += eigB;
+
+        unsigned int numPts = pts.extent(1);
+        Kokkos::MDRangePolicy<Kokkos::Rank<2>, typename MemoryToExecution<MemorySpace>::Space> policy({0, 0}, {numPts, this->outputDim});
+
+        Kokkos::parallel_for(policy, KOKKOS_CLASS_LAMBDA(const int& j, const int& i) {
+            output(i,j) += b_(i);
+        });
     }
 }
 
@@ -76,159 +64,21 @@ void AffineFunction<MemorySpace>::CoeffGradImpl(StridedMatrix<const double, Memo
     return;
 }
     
-template<>
[…]
