+++ src/AffineMap.cpp
@@ -34,16 +34,23 @@ AffineMap<MemorySpace>::AffineMap(StridedMatrix<double,MemorySpace> A,
 }
 
 
-template<>
-void AffineMap<Kokkos::HostSpace>::Factorize(){
+template<typename MemorySpace>
+void AffineMap<MemorySpace>::Factorize(){
 
-    auto eigA = KokkosToMat(A_);
-    unsigned int nrows = eigA.rows();
-    unsigned int ncols = eigA.cols();
+    // If A_ is not column major, create a column major version
+    if(A_.stride_0()!=1){
+        Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace> anew("A_", A_.extent(0), A_.extent(1));
+        Kokkos::deep_copy(anew, A_);
+        A_ = anew;
+    }
 
-    // Use eigen to compute the LU decomposition in place
-    luSolver_ = std::make_shared<Eigen::PartialPivLU<Eigen::MatrixXd>>(eigA.block(0,ncols-nrows, nrows, nrows));
-    logDet_ = log(abs(luSolver_->determinant()));
+    if(A_.extent(0)!=A_.extent(1)){
+        StridedMatrix<const double, MemorySpace> Asub = Kokkos::subview(A_, Kokkos::ALL(), std::make_pair(A_.extent(1)-A_.extent(0),A_.extent(1)));
+        luSolver_.compute(Asub);
+    }else{
+        luSolver_.compute(A_);
+    }
+    logDet_ = log(abs(luSolver_.determinant()));
 }
 
 
@@ -58,208 +65,22 @@ void AffineMap<MemorySpace>::LogDeterminantImpl(StridedMatrix<const double, Memo
     });
 }
 
-template<>
-void AffineMap<Kokkos::HostSpace>::InverseImpl(StridedMatrix<const double, Kokkos::HostSpace> const& x1,
-                                               StridedMatrix<const double, Kokkos::HostSpace> const& r,
-                                               StridedMatrix<double, Kokkos::HostSpace>              output)
-{
-
-    auto eigOut = KokkosToMat<double>(output);
-    auto eigR = ConstKokkosToMat<double>(r);
-    
-    // Bias part
-    if(b_.size()>0){
-        auto eigB = KokkosToVec<double>(b_);
-        eigOut = eigR.colwise() - eigB;
-
-    }else{
-        eigOut = eigR;
-    }
-
-    if(A_.extent(0)>0){
-        auto eigA = KokkosToMat(A_);
-        unsigned int nrows = eigA.rows();
-        unsigned int ncols = eigA.cols();
-
-        // If the matrix is rectangular, treat it as the lower part of a block triangular matrix
-        // The value of x1 contains the compute inverse for the first block
-        if(nrows != ncols){
-            auto eigX = ConstKokkosToMat<double>(x1);
-            eigOut -= eigA.block(0,0,nrows,ncols-nrows) * eigX.topRows(ncols-nrows);
-        }
-
-        eigOut = luSolver_->solve(eigOut);
-    }
-}
-
-
-
 template<typename MemorySpace>
-void AffineMap<MemorySpace>::LogDeterminantCoeffGradImpl(StridedMatrix<const double, MemorySpace> const& pts, 
-                                                         StridedMatrix<double, MemorySpace>              output)
+void AffineMap<MemorySpace>::InverseImpl(StridedMatrix<const double, MemorySpace> const& x1,
+                                         StridedMatrix<const double, MemorySpace> const& r,
+                                         StridedMatrix<double, MemorySpace>              output)
 {
-    return;
-}
 
-template<>
[…]
