### 2022-09-20 c30cf41
Updated install docs to mention cublas and cusolver.
