### 2022-09-20 a4204b4
Added docs for Linear algebra.
