+++ docs/source/api/utilities/linearalgebra.rst
@@ -1,0 +1,173 @@
+===================
+Linear Algebra
+===================
+
+Kokkos provides the :code:`Kokkos::View` class to store multi-dimensional arrays.  Kokkos itself does not provide tools BLAS (e.g., Mat-Mat multiplication) or LAPACK (e.g., LU factorizations and solves) operations.   MParT provides a small collection of c++ classes and functions for easily working with one- and two-dimensional :code:`Kokkos::View` objects as if they were vectors and matrices, respectively.    Below is a comparison of the MParT linear algebra tools and <Eigen `https://eigen.tuxfamily.org/index.php?title=Main_Page`>_.  These examples focus on views defined in the :code:`Kokkos::HostSpace` memory space, but these operators are defined for views in device memory (e.g., :code:`mpart::DeviceSpace`) as well.
+
+The eigen code snippets require :code:`#include <Eigen/Core>` and :code:`#include <Eigen/Dense>` while the MParT snippets require :code:`#include <Kokkos_Core.hpp>`, :code`#include "MParT/Utilities/LinearAlgebra.h"`, and the addition of :code:`using namespace mpart;`.
+
+Construction
+-------------
+
+For completeness, we first show how to construct two matrices :math:`A` and :math:`B` using either Eigen or Kokkos.
+
+.. grid:: 2
+
+    .. grid-item-card::  Matrix Construction (Eigen)
+
+        Eigen::MatrixXd A(2,2);
+        Eigen::MatrixXd B(2,2);
+
+        A << 1.0, 2.0,
+             0.5, 1.0;
+
+        B << 1.0, 2.0,
+             0.125, 0.5;
+
+        
+    .. grid-item-card::  Matrix Construction (Kokkos)
+
+        Kokkos::View<double**, Kokkos::HostSpace> A("A", 2, 2);
+        Kokkos::View<double** ,Kokkos::HostSpace> B("B", 2, 2);
+
+        A(0,0) = 1.0; A(0,1) = 2.0;
+        A(1,0) = 0.5; A(1,1) = 1.0;
+
+        B(0,0) = 1.0; B(0,1) = 2.0;
+        B(1,0) = 0.125; B(1,1) = 0.5;
+
+Summation
+----------
+
+We will now compute :math:`C=2A+B` by first computing :math:`A+B` and then adding an additional :math:`A` to the result.  This is simply to demonstrate the use of both the :code:`+` and :code`+=` operators.
+
+.. grid:: 2
+
+    .. grid-item-card::  Matrix Sum (Eigen)
+
+        Eigen::MatrixXd C;
+        C = A + B;
+        C += A;
+
+
+    .. grid-item-card::  Matrix Sum (Kokkos+MParT)
+
+        Kokkos::View<double** ,Kokkos::HostSpace> C;
+        C = A + B;
+        C += A;
+
+Multiplication
+---------------
+
+The :code:`*` operator is overloaded in both Eigen and MParT to implement matrix multiplication.  Currently MParT does not provide any built-in tools for componentwise multiplication.  The snippet below computes the product :math:`AB`.
+
+.. grid:: 2
+
+    .. grid-item-card::  Matrix Sum (Eigen)
+
+        Eigen::MatrixXd C;
+        C = A*B;
+
+
+    .. grid-item-card::  Matrix Sum (Kokkos+MParT)
+
+        Kokkos::View<double** ,Kokkos::HostSpace> C;
+        C = A*B;
+
[…]
