### 2022-09-20 234c57e
Refined Linear algebra docs.
