+++ docs/source/api/utilities/linearalgebra.rst
@@ -2,9 +2,11 @@
 Linear Algebra
 ===================
 
-Kokkos provides the :code:`Kokkos::View` class to store multi-dimensional arrays.  Kokkos itself does not provide tools BLAS (e.g., Mat-Mat multiplication) or LAPACK (e.g., LU factorizations and solves) operations.   MParT provides a small collection of c++ classes and functions for easily working with one- and two-dimensional :code:`Kokkos::View` objects as if they were vectors and matrices, respectively.    Below is a comparison of the MParT linear algebra tools and <Eigen `https://eigen.tuxfamily.org/index.php?title=Main_Page`>_.  These examples focus on views defined in the :code:`Kokkos::HostSpace` memory space, but these operators are defined for views in device memory (e.g., :code:`mpart::DeviceSpace`) as well.
+Kokkos provides the :code:`Kokkos::View` class to store multi-dimensional arrays.  Kokkos itself does not provide tools BLAS (e.g., Mat-Mat multiplication) or LAPACK (e.g., LU factorizations and solves) operations.   MParT provides a small collection of c++ classes and functions for easily working with one- and two-dimensional :code:`Kokkos::View` objects as if they were vectors and matrices, respectively.   
 
-The eigen code snippets require :code:`#include <Eigen/Core>` and :code:`#include <Eigen/Dense>` while the MParT snippets require :code:`#include <Kokkos_Core.hpp>`, :code`#include "MParT/Utilities/LinearAlgebra.h"`, and the addition of :code:`using namespace mpart;`.
+Below is a comparison of the MParT linear algebra tools and `Eigen <https://eigen.tuxfamily.org/index.php?title=Main_Page>`_.  These examples focus on views defined in the :code:`Kokkos::HostSpace` memory space, but these operators are defined for views in device memory (e.g., :code:`mpart::DeviceSpace`) as well.
+
+The eigen code snippets require :code:`#include <Eigen/Core>` and :code:`#include <Eigen/Dense>` while the MParT snippets require :code:`#include <Kokkos_Core.hpp>`, :code:`#include "MParT/Utilities/LinearAlgebra.h"`, and the addition of :code:`using namespace mpart;`.
 
 Construction
 -------------
@@ -14,47 +16,55 @@ For completeness, we first show how to construct two matrices :math:`A` and :mat
 .. grid:: 2
 
     .. grid-item-card::  Matrix Construction (Eigen)
+        
+        .. code-block:: cpp
+                
+            Eigen::MatrixXd A(2,2);
+            Eigen::MatrixXd B(2,2);
 
-        Eigen::MatrixXd A(2,2);
-        Eigen::MatrixXd B(2,2);
+            A << 1.0, 2.0,
+                 0.5, 1.0;
 
-        A << 1.0, 2.0,
-             0.5, 1.0;
-
-        B << 1.0, 2.0,
-             0.125, 0.5;
+            B << 1.0,   2.0,
+                 0.125, 0.5;
 
         
     .. grid-item-card::  Matrix Construction (Kokkos)
 
-        Kokkos::View<double**, Kokkos::HostSpace> A("A", 2, 2);
-        Kokkos::View<double** ,Kokkos::HostSpace> B("B", 2, 2);
+        .. code-block:: cpp
 
-        A(0,0) = 1.0; A(0,1) = 2.0;
-        A(1,0) = 0.5; A(1,1) = 1.0;
+            Kokkos::View<double**, Kokkos::HostSpace> A("A", 2, 2);
+            Kokkos::View<double** ,Kokkos::HostSpace> B("B", 2, 2);
 
[…]
