### 2022-09-20 e664f4c
Added math.h header to try to circumvent isinf issue.
