### 2022-09-21 d104892
Add more memoryspace'ing
