### 2022-09-21 39439fa
Add more device to kokkos
