### 2022-09-21 d969d2b
More array conversions
