### 2022-09-21 388779f
Fixed bug in triangular map related to zero initialization of gradient.
