### 2022-09-21 2ff26f3
Removed deep copies from ComposedMap class and simplified implementations of logdeterminant gradient functions.
