+++ tests/Test_ComposedMap.cpp
@@ -3,6 +3,8 @@
 #include "MParT/ComposedMap.h"
 #include "MParT/MapFactory.h"
 
+#include "MParT/Utilities/LinearAlgebra.h"
+
 using namespace mpart;
 using namespace Catch;
 using MemorySpace = Kokkos::HostSpace;
@@ -36,7 +38,7 @@ TEST_CASE( "Testing 5 layer composed map", "[ComposedMap_Constructor]" ) {
     Eigen::RowVectorXd coeffs(composedMap->numCoeffs);
     for(unsigned int i=0; i<composedMap->numCoeffs; ++i)
         coeffs(i) = 0.1*(i+1);
-        
+    
     SECTION("Coefficients"){
         
         // Set the coefficients of the triangular map
@@ -65,46 +67,48 @@ TEST_CASE( "Testing 5 layer composed map", "[ComposedMap_Constructor]" ) {
 
     composedMap->SetCoeffs(coeffs);
     auto out = composedMap->Evaluate(in);
-    // Kokkos::View<double*,Kokkos::HostSpace> coeffs("Coefficients", composedMap->numCoeffs);
-    // // intermediate output
-    // Kokkos::View<double**, MemorySpace> intOutput("intermediate output", output.extent(0), output.extent(1));
     
-    // // Copy points to output, then output = map(output) looped over each component
-    // Kokkos::deep_copy(output, pts);
-    // for(unsigned int i=0; i<comps_.size(); ++i){
-        
-    //     comps_.at(i)->EvaluateImpl(output, intOutput);
-    //     Kokkos::deep_copy(output,intOutput);
-    // }
+    SECTION("Evaluate"){
+        Kokkos::View<double**, Kokkos::HostSpace> trueOut("True output", dim, numSamps);
+        Kokkos::deep_copy(trueOut, in);
+        for(auto& comp : maps)
+            trueOut = comp->Evaluate(trueOut);
 
+        for(unsigned int i=0; i<in.extent(0); ++i){
+            for(unsigned int j=0; j<numSamps; ++j)
+                CHECK( out(i,j) == Approx(trueOut(i,j)).epsilon(1e-7).margin(1e-7));
+        }
+    }
 
-    // SECTION("Inverse"){
+    SECTION("Inverse"){
 
-    //     auto inv = triMap->Inverse(in,out);
+        auto inv = composedMap->Inverse(in,out);
 
-    //     for(unsigned int i=0; i<numBlocks; ++i){
-    //         for(unsigned int j=0; j<numSamps; ++j)
-    //             CHECK( inv(i,j) == Approx(in(i+extraInputs,j)).epsilon(1e-6));
-    //     }
-    // }
+        for(unsigned int i=0; i<in.extent(0); ++i){
+            for(unsigned int j=0; j<numSamps; ++j)
+                CHECK( inv(i,j) == Approx(in(i,j)).epsilon(1e-5).margin(1e-5));
+        }
+    }
 
-    // SECTION("LogDeterminant"){
-    //     auto logDet = triMap->LogDeterminant(in);
+    SECTION("LogDeterminant"){
+        auto logDet = composedMap->LogDeterminant(in);
 
-    //     REQUIRE(logDet.extent(0)==numSamps);
-    //     Kokkos::View<double*, Kokkos::HostSpace> truth("True Log Det", numSamps);
+        REQUIRE(logDet.extent(0)==numSamps);
+        Kokkos::View<double*, Kokkos::HostSpace> truth("True Log Det", numSamps);
+        Kokkos::View<double*, Kokkos::HostSpace> partialDet("Partial log det", numSamps);
[…]
