+++ src/ComposedMap.cpp
@@ -1,5 +1,8 @@
 #include "MParT/ComposedMap.h"
 
+#include "MParT/Utilities/Miscellaneous.h"
+#include "MParT/Utilities/LinearAlgebra.h"
+
 #include <numeric>
 
 using namespace mpart;
@@ -59,7 +62,7 @@ void ComposedMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, Kokkos::DefaultEx
 
 template<typename MemorySpace>
 void ComposedMap<MemorySpace>::LogDeterminantImpl(StridedMatrix<const double, MemorySpace> const& pts,
-                                                    StridedVector<double, MemorySpace>              output)
+                                                  StridedVector<double, MemorySpace>              output)
 {
 
 
@@ -69,11 +72,11 @@ void ComposedMap<MemorySpace>::LogDeterminantImpl(StridedMatrix<const double, Me
         return;
 
     // intermediate points and variable to hold logdet increments 
-    Kokkos::View<double**, MemorySpace>  intPts1("intermediate points 1", pts.extent(0), pts.extent(1));
-    Kokkos::View<double**, MemorySpace>  intPts2("intermediate points 2", pts.extent(0), pts.extent(1));
+    Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>  intPts1("intermediate points 1", pts.extent(0), pts.extent(1));
+    Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>  intPts2("intermediate points 2", pts.extent(0), pts.extent(1));
     Kokkos::deep_copy(intPts1, pts);
 
-    Kokkos::View<double*, MemorySpace> compDetIncrement("Log Determinant", output.extent(0));
+    Kokkos::View<double*, Kokkos::LayoutLeft, MemorySpace> compDetIncrement("Log Determinant", output.extent(0));
     for(unsigned int i=1; i<comps_.size(); ++i){
         
         // Compute x_i = T_{i-1}(x_{i-1})
@@ -83,36 +86,30 @@ void ComposedMap<MemorySpace>::LogDeterminantImpl(StridedMatrix<const double, Me
         comps_.at(i)->LogDeterminantImpl(intPts2, compDetIncrement);
 
         // Add to logdet of full map
-        for(unsigned int j=0; j<output.size(); ++j)
-            output(j) += compDetIncrement(j);
+        output += compDetIncrement;
 
-        // update current x_{i-1} <-- x_{i}
-        Kokkos::deep_copy(intPts1, intPts2);
+        // x_{i-1} <-- x_{i} without memory copy
+        simple_swap<decltype(intPts1)>(intPts1, intPts2);
     }
 
 }
 
 template<typename MemorySpace>
 void ComposedMap<MemorySpace>::EvaluateImpl(StridedMatrix<const double, MemorySpace> const& pts,
-                                              StridedMatrix<double, MemorySpace>              output)
+                                            StridedMatrix<double, MemorySpace>              output)
 {
     // intermediate output
-    Kokkos::View<double**, MemorySpace> intOutput("intermediate output", output.extent(0), output.extent(1));
+    Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>  intPts1("intermediate points 1", pts.extent(0), pts.extent(1));
+    Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>  intPts2("intermediate points 2", pts.extent(0), pts.extent(1));
     
-    // Copy points to output, then output = map(output) looped over each component
-    Kokkos::deep_copy(output, pts);
-    for(unsigned int i=0; i<comps_.size(); ++i){
-        
[…]
