+++ src/IdentityMap.cpp
@@ -3,6 +3,8 @@
 #include <numeric>
 
 using namespace mpart;
+using MemorySpace = Kokkos::HostSpace;
+
 
 template<typename MemorySpace>
 IdentityMap<MemorySpace>::IdentityMap(unsigned int inDim, unsigned int outDim) : ConditionalMapBase<MemorySpace>(inDim, outDim, 0)
@@ -54,21 +56,22 @@ void IdentityMap<MemorySpace>::CoeffGradImpl(StridedMatrix<const double, MemoryS
     assert(false);
 }
 
-
+template<typename MemorySpace>
 void IdentityMap<MemorySpace>::GradientImpl(StridedMatrix<const double, MemorySpace> const& pts,  
                             StridedMatrix<const double, MemorySpace> const& sens,
-                            StridedMatrix<double, MemorySpace>              output) override
+                            StridedMatrix<double, MemorySpace>              output)
 {
 
 
-    // zero until inputDim-outDim
-    for(unsigned int j=0; j<int(this->inputDim - this->outputDim); ++j)
-        output(j) = 0.0;
+    //zero until inputDim-outDim
+    for(unsigned int i=0; i<int(this->inputDim - this->outputDim); ++i)
+        for(unsigned int j=0; j<int(this->inputDim); ++j)
+            output(i,j) = 0.0;
 
-    StridedMatrix<const double, MemorySpace> tailOut = Kokkos::subview(
+    StridedMatrix<double, MemorySpace> tailOut = Kokkos::subview(
         output, std::make_pair(int(this->inputDim - this->outputDim), int(this->inputDim)), Kokkos::ALL());
 
-    Kokkos::deep_copy(tailOutput, sens);
+    Kokkos::deep_copy(tailOut, sens);
 
 }
 
@@ -79,12 +82,14 @@ void IdentityMap<MemorySpace>::LogDeterminantCoeffGradImpl(StridedMatrix<const d
     assert(false);
 }
 
+template<typename MemorySpace>
 void IdentityMap<MemorySpace>::LogDeterminantInputGradImpl(StridedMatrix<const double, MemorySpace> const& pts, 
-                                            StridedMatrix<double, MemorySpace>              output) override
+                                            StridedMatrix<double, MemorySpace>              output)
 {   
     // Add to logdet of full map
-    for(unsigned int j=0; j<output.size(); ++j)
-        output(j) = 0.0;
+    for(unsigned int i=0; i<output.extent(0); ++i)
+        for(unsigned int j=0; j<output.extent(1); ++j)
+            output(i,j) = 0.0;
 }
 
 // Explicit template instantiation
