### 2022-09-22 7b6d24f
Work on array conversions
