+++ tests/Test_ArrayConversions.cpp
@@ -20,7 +20,7 @@ TEST_CASE( "Testing Pointer to Kokkos Conversions in 1D", "[ArrayConversions1D]"
             data[i] = i;
 
 
-        auto view = ToKokkos(&data[0], length);
+        auto view = ToKokkos<double,Kokkos::HostSpace>(&data[0], length);
         REQUIRE(view.extent(0) == length);
         for(unsigned int i=0; i<length; ++i){
 
@@ -40,7 +40,7 @@ TEST_CASE( "Testing Pointer to Kokkos Conversions in 1D", "[ArrayConversions1D]"
         for(unsigned int i=0; i<length; ++i)
             data[i] = i;
 
-        auto view = ToKokkos(&data[0], length);
+        auto view = ToKokkos<int,Kokkos::HostSpace>(&data[0], length);
         REQUIRE(view.extent(0) == length);
         for(unsigned int i=0; i<length; ++i){
 
@@ -125,10 +125,10 @@ TEST_CASE( "Testing functions that copy views between host and device", "[ArrayC
         CHECK( hostVec2(i) ==i );
 
     // Copy a slice back to host
-    auto slice1 = ToHost(deviceVec, std::make_pair(1,3));
-    REQUIRE( slice1.extent(0) == 2);
-    CHECK(slice1(0)==1);
-    CHECK(slice1(1)==2);
+    // auto slice1 = ToHost(deviceVec, std::make_pair(1,3));
+    // REQUIRE( slice1.extent(0) == 2);
+    // CHECK(slice1(0)==1);
+    // CHECK(slice1(1)==2);
 }
 
 
@@ -157,6 +157,87 @@ TEST_CASE( "Testing mapping from memory space to valid execution space.", "[Kokk
     REQUIRE(hostVec2.extent(0)==N1);
     for(unsigned int i=0; i<N1; ++i)
         CHECK( hostVec2(i) ==i );
+}
+
+
+TEST_CASE( "Testing Pointer to Kokkos Conversions in 1D, Device", "[ArrayConversions1D]" ) {
+
+    unsigned int length = 10;
+
+    SECTION("double"){
+
+        // Initialize a std vector
+        std::vector<double> data(length);
+        for(unsigned int i=0; i<length; ++i)
+            data[i] = i;
+
+
+        auto view_d = ToKokkos<double,DeviceSpace>(&data[0], length);
+        auto view = ToHost(view_d);
+        REQUIRE(view.extent(0) == length);
+        for(unsigned int i=0; i<length; ++i){
+
+            // Make sure the values are the same
+            CHECK( data[i] == view(i) );
+
+            // Check sure the memory address is the same  (i.e., we're not copying)
+            CHECK( &data[i] == &view(i) );
+        }
+    }
+
+
+    SECTION("int"){
+
+        // Initialize a std vector
+        std::vector<int> data(length);
+        for(unsigned int i=0; i<length; ++i)
+            data[i] = i;
+
+        auto view_d = ToKokkos<int,DeviceSpace>(&data[0], length);
+        auto view = ToHost(view_d);
+        REQUIRE(view.extent(0) == length);
+        for(unsigned int i=0; i<length; ++i){
+
+            // Make sure the values are the same
+            CHECK( data[i] == view(i) );
+
+            // Check sure the memory address is the same  (i.e., we're not copying)
+            CHECK( &data[i] == &view(i) );
+        }
+    }
+
+}
+
+TEST_CASE( "Testing Pointer to Kokkos Conversions in 2D, Device", "[ArrayConversions2D]" ) {
+
+    unsigned int rows = 10;
+    unsigned int cols = 20;
+
+
+    SECTION("double"){
+
+        // Initialize a std vector
[…]
