+++ src/ComposedMap.cpp
@@ -8,17 +8,17 @@
 using namespace mpart;
 
 template<typename MemorySpace>
-ComposedMap<MemorySpace>::ComposedMap(std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> const& components) : ConditionalMapBase<MemorySpace>(components.front()->inputDim,
-                                                                                                                                                         components.back()->outputDim,
-                        std::accumulate(components.begin(), components.end(), 0, [](size_t sum, std::shared_ptr<ConditionalMapBase<MemorySpace>> const& comp){ return sum + comp->numCoeffs; })),
-                        comps_(components)
+ComposedMap<MemorySpace>::ComposedMap(std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> const& maps) : ConditionalMapBase<MemorySpace>(maps.front()->inputDim,
+                                                                                                                                                         maps.back()->outputDim,
+                        std::accumulate(maps.begin(), maps.end(), 0, [](size_t sum, std::shared_ptr<ConditionalMapBase<MemorySpace>> const& comp){ return sum + comp->numCoeffs; })),
+                        maps_(maps)
 {
 
     // Check the sizes of all the inputs
-    for(unsigned int i=0; i<comps_.size()-1; ++i){
-        if(comps_.at(i)->outputDim != comps_.at(i+1)->inputDim){
+    for(unsigned int i=0; i<maps_.size()-1; ++i){
+        if(maps_.at(i)->outputDim != maps_.at(i+1)->inputDim){
             std::stringstream msg;
-            msg << "In ComposedMap constructor, the output dimension (" << comps_.at(i)->outputDim << ") of component " << i << " is not equal to the input dimension (" << comps_.at(i+1)->inputDim << ").";
+            msg << "In ComposedMap constructor, the output dimension (" << maps_.at(i)->outputDim << ") of component " << i << " is not equal to the input dimension (" << maps_.at(i+1)->inputDim << ").";
             throw std::invalid_argument(msg.str());
         }
     }
@@ -32,12 +32,12 @@ void ComposedMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, Kokkos::HostSpace
 
     // Now create subviews for each of the components
     unsigned int cumNumCoeffs = 0;
-    for(unsigned int i=0; i<comps_.size(); ++i){
-        assert(cumNumCoeffs+comps_.at(i)->numCoeffs <= this->savedCoeffs.size());
+    for(unsigned int i=0; i<maps_.size(); ++i){
+        assert(cumNumCoeffs+maps_.at(i)->numCoeffs <= this->savedCoeffs.size());
 
-        comps_.at(i)->WrapCoeffs(Kokkos::subview(this->savedCoeffs,
-            std::make_pair(cumNumCoeffs, cumNumCoeffs+comps_.at(i)->numCoeffs)));
-        cumNumCoeffs += comps_.at(i)->numCoeffs;
+        maps_.at(i)->WrapCoeffs(Kokkos::subview(this->savedCoeffs,
+            std::make_pair(cumNumCoeffs, cumNumCoeffs+maps_.at(i)->numCoeffs)));
+        cumNumCoeffs += maps_.at(i)->numCoeffs;
     }
 }
 
[…]
