### 2022-09-22 2ddc13f
Removed old friend declarations
