+++ MParT/ConditionalMapBase.h
@@ -10,15 +10,6 @@
 
 namespace mpart {
 
-    template<typename MemorySpace>
-    class TriangularMap;
-
-    template<typename MemorySpace>
-    class ComposedMap;
-
-    template<typename MemorySpace>
-    class IdentityMap;
-
     /**
      @brief Provides an abstract base class for conditional transport maps where the input dimension might be larger than output dimension.
      @details
@@ -32,10 +23,6 @@ namespace mpart {
      */
     template<typename MemorySpace>
     class ConditionalMapBase : public ParameterizedFunctionBase<MemorySpace>{
-
-        friend class TriangularMap<MemorySpace>;
-        friend class ComposedMap<MemorySpace>;
-        friend class IdentityMap<MemorySpace>;
 
 
     public:
