### 2022-09-22 73a19ce
Replaces for loops with Kokkos::parallel_for loops
