### 2022-09-22 c674ec3
Work on new Array Conversions
