### 2022-09-23 1597088
Work on array conversions
