+++ bindings/julia/include/CommonJuliaUtilities.h
@@ -48,6 +48,18 @@ void ConditionalMapBaseWrapper(jlcxx::Module&);
 void TriangularMapWrapper(jlcxx::Module&);
 
 /**
+ * @brief Adds AffineMap bindings to the existing module m.
+ * @param mod CxxWrap.jl module
+ */
+void AffineMapWrapper(jlcxx::Module&);
+
+/**
+ * @brief Adds AffineFunction bindings to the existing module m.
+ * @param mod CxxWrap.jl module
+ */
+void AffineFunctionWrapper(jlcxx::Module&);
+
+/**
  * @brief Adds MapFactory bindings to the existing module m.
  * @param mod CxxWrap.jl module
  */
