+++ bindings/julia/CMakeLists.txt
@@ -8,6 +8,7 @@ set(JULIA_BINDING_SOURCES
   src/ParameterizedFunctionBase.cpp
   src/TriangularMap.cpp
   src/Wrapper.cpp
+  src/AffineMap.cpp
   ../common/src/CommonUtilities.cpp
 )
 
