### 2022-09-23 688fc76
Added start of checkpointer class in composed map.  Currently just evaluates from beginning.
