+++ src/ComposedMap.cpp
@@ -8,6 +8,56 @@
 using namespace mpart;
 
 template<typename MemorySpace>
+ComposedMap<MemorySpace>::Checkpointer::Checkpointer(unsigned int maxSaves, 
+                                                     StridedMatrix<const double, MemorySpace> initialPts,
+                                                     std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>>& comps) : maxSaves_(maxSaves),
+                                                                                                                            comps_(comps)
+{
+    if(maxSaves_==0){
+        throw std::runtime_error("In ComposedMap::Checkpointer: The maximum number of checkpoints must be larger than 0.");
+    }
+
+    // Store the initial points as the first checkpoint
+    checkpoints_.push_back(Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>("x0", initialPts.extent(0), initialPts.extent(1)));
+    Kokkos::deep_copy(checkpoints_.back(), initialPts);
+
+    checkpointLayers_.push_back(0);
+
+    // Allocate the workspace 
+    workspace1_ = Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>("xi", initialPts.extent(0), initialPts.extent(1));
+    workspace2_ = Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>("xj", initialPts.extent(0), initialPts.extent(1));
+}
+
+
+template<typename MemorySpace>
+Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace> ComposedMap<MemorySpace>::Checkpointer::GetLayerInput(unsigned int layerInd)
+{   
+    // Check if the layer asked for is the last checkpoint
+    if(layerInd==checkpointLayers_.back()){
+        return checkpoints_.back();
+
+    // If the requested layer is before the last checkpoint, free up the checkpoint
+    }else if(layerInd<checkpointLayers_.back()){
+        checkpoints_.pop_back();
+        checkpointLayers_.pop_back();
+        return GetLayerInput(layerInd);
+
+    }else{
+
+        // Compute the input by reevaluating from the last checkpoint
+        unsigned int i = checkpointLayers_.back();
+        comps_.at(i)->EvaluateImpl(checkpoints_.back(), workspace1_);
+        ++i;
+        for(; i<layerInd; ++i){
+            comps_.at(i)->EvaluateImpl(workspace1_, workspace2_);
+            simple_swap(workspace1_,workspace2_);
+        }
+        return workspace1_;
+    }
+}
+
+
+template<typename MemorySpace>
 ComposedMap<MemorySpace>::ComposedMap(std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> const& components) : ConditionalMapBase<MemorySpace>(components.front()->inputDim,
                                                                                                                                                          components.back()->outputDim,
                         std::accumulate(components.begin(), components.end(), 0, [](size_t sum, std::shared_ptr<ConditionalMapBase<MemorySpace>> const& comp){ return sum + comp->numCoeffs; })),
@@ -102,7 +152,13 @@ void ComposedMap<MemorySpace>::EvaluateImpl(StridedMatrix<const double, MemorySp
     Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>  intPts1("intermediate points 1", pts.extent(0), pts.extent(1));
[…]
