### 2022-09-23 3ba5184
Adding ComposedMap
