+++ bindings/julia/CMakeLists.txt
@@ -9,6 +9,7 @@ set(JULIA_BINDING_SOURCES
   src/TriangularMap.cpp
   src/Wrapper.cpp
   src/AffineMap.cpp
+  src/ComposedMap.cpp
   ../common/src/CommonUtilities.cpp
 )
 
