### 2022-09-23 0045929
Added checkpointing to composed map class.
