### 2022-09-28 #4730 @ansgarwenzel
Hi Patrick,

Sure.

Kind regards,

Ansgar

Patrick Diehl ***@***.***> schrieb am Mi., 28. Sept. 2022,
16:48:

> Hi @ansgarwenzel <https://github.com/ansgarwenzel> would you have time to
> review this paper?
>
