### 2022-09-30 f763802
Create C++ for Slice of triangular map
