+++ bindings/julia/src/TriangularMap.cpp
@@ -14,7 +14,7 @@ void mpart::binding::TriangularMapWrapper(jlcxx::Module &mod) {
        .method("InverseInplace", [](TriangularMap<Kokkos::HostSpace> &map, jlcxx::ArrayRef<double,2> x, jlcxx::ArrayRef<double,2> r){
             map.InverseInplace(JuliaToKokkos(x), JuliaToKokkos(r));
        })
-       .method("GetComponent", &TriangularMap<Kokkos::HostSpace>::GetComponent)
-       .method("Slice", &TriangularMap<Kokkos::HostSpace>::Slice)
+       .method("GetComponent", [](TriangularMap<Kokkos::HostSpace>& map, int i){ return map.GetComponent(i-1); })
+       .method("Slice", [](TriangularMap<Kokkos::HostSpace>& map, int begin, int end){ return map.Slice(begin-1, end); })
     ;
 }
