### 2022-10-11 #4730 @dmbates
I tried to reply to your message in the issue but received a message that I
cannot comment at this time.  My reply was:

I regret that I am not qualified to review this project.  I have
(thankfully) forgotten most of the C++ that I ever knew and I am not
familiar with the applications area.

On Mon, Oct 3, 2022 at 10:38 AM Patrick Diehl ***@***.***>
wrote:

> Hi @dmbates <https://github.com/dmbates> would you have time to review
> this paper?
>
