### 2022-10-11 #4730 @diehlpk
> > @f-t-s can you please reply to my question above?
>
> Sorry, I overlooked this. We had an initial meeting but have not been much in touch since, since the main driver of the collaboration still had to wrap up a previous program. However, he is about to get started on this. If this is considered fine, I am happy to review it. I just don't feel comfortable giving guarantees about delaying the start of (work on) a collaboration until a review is finished.

Sounds good with me if you could finish the review in a timely manner,
