### 2022-10-11 73d75f9
created mex and mat functions
