### 2022-10-11 1417b78
Return Shared Pointer
