+++ bindings/julia/src/AffineMap.cpp
@@ -18,13 +18,13 @@ void mpart::binding::AffineMapWrapper(jlcxx::Module &mod)
 {
     mod.add_type<AffineMap<Kokkos::HostSpace>>("AffineMap", jlcxx::julia_base_type<ConditionalMapBase<Kokkos::HostSpace>>());
     mod.method("AffineMap", [](jlcxx::ArrayRef<double> b){
-        return jlcxx::create<AffineMap<Kokkos::HostSpace>>(JuliaToKokkos(b));
+        return std::make_shared<AffineMap<Kokkos::HostSpace>>(JuliaToKokkos(b));
     });
     mod.method("AffineMap", [](jlcxx::ArrayRef<double,2> A, jlcxx::ArrayRef<double> b){
-        return jlcxx::create<AffineMap<Kokkos::HostSpace>>(JuliaToKokkos(A), JuliaToKokkos(b));
+        return std::make_shared<AffineMap<Kokkos::HostSpace>>(JuliaToKokkos(A), JuliaToKokkos(b));
     });
     mod.method("AffineMap", [](jlcxx::ArrayRef<double,2> A){
-        return jlcxx::create<AffineMap<Kokkos::HostSpace>>(JuliaToKokkos(A));
+        return std::make_shared<AffineMap<Kokkos::HostSpace>>(JuliaToKokkos(A));
     });
 }
 
@@ -33,12 +33,12 @@ void mpart::binding::AffineFunctionWrapper(jlcxx::Module &mod)
 {
     mod.add_type<AffineFunction<Kokkos::HostSpace>>("AffineFunction", jlcxx::julia_base_type<ParameterizedFunctionBase<Kokkos::HostSpace>>());
     mod.method("AffineFunction", [](jlcxx::ArrayRef<double> b){
-        return jlcxx::create<AffineFunction<Kokkos::HostSpace>>(JuliaToKokkos(b));
+        return std::make_shared<AffineFunction<Kokkos::HostSpace>>(JuliaToKokkos(b));
     });
     mod.method("AffineFunction", [](jlcxx::ArrayRef<double,2> A, jlcxx::ArrayRef<double> b){
-        return jlcxx::create<AffineFunction<Kokkos::HostSpace>>(JuliaToKokkos(A), JuliaToKokkos(b));
+        return std::make_shared<AffineFunction<Kokkos::HostSpace>>(JuliaToKokkos(A), JuliaToKokkos(b));
     });
     mod.method("AffineFunction", [](jlcxx::ArrayRef<double,2> A){
-        return jlcxx::create<AffineFunction<Kokkos::HostSpace>>(JuliaToKokkos(A));
+        return std::make_shared<AffineFunction<Kokkos::HostSpace>>(JuliaToKokkos(A));
     });
 }
