### 2022-10-12 a8f5e8b
compiles with affine map constructor
