+++ bindings/matlab/mat/ConditionalMap.m
@@ -29,6 +29,10 @@ methods
             this.id_=varargin{1};
           elseif(varargin{2}=="compose")
             this.id_=MParT_('ConditionalMap_newComposedMap', varargin{1});
+          elseif(varargin{2}=="A")
+            this.id_=MParT_('ConditionalMap_newAffineMapA', varargin{1});
+          elseif(varargin{2}=="b")
+            this.id_=MParT_('ConditionalMap_newAffineMapb', varargin{1});
           end
         else
           mset = varargin{1};
@@ -51,9 +55,15 @@ methods
               input_str=[input_str,')'];
               this.id_ = eval(input_str);
           else
-              error("Wrong input argument");
+              error("Wrong input arguments");
           end
         end
+    elseif(nargin==3)
+      if(varargin{3}=="Ab")
+          this.id_=MParT_('ConditionalMap_newAffineMapAb', varargin{1},varargin{2});
+      else
+        error("Wrong input arguments");
+      end
     elseif(nargin==4)
       inputDim = varargin{1};
       outputDim = varargin{2};
