+++ bindings/matlab/tests/Test_ComposedMap.m
@@ -11,4 +11,4 @@ T2 = CreateTriangular(2,2,2,opts);
 
 T = ComposedMap([T1 T2]);
 
-
+Y=T.Evaluate(randn(2,10));
