### 2022-10-12 4270220
Ensure output is zeroed, Evaluate
