+++ bindings/matlab/external/mexplus/mexplus/mxarray.h
@@ -931,8 +931,15 @@ class MxArray {
                          R
                        >::type* value) {
     if (mxIsComplex(array)) {
+      
+      #if MX_HAS_INTERLEAVED_COMPLEX
+      mxComplexDouble* pc = mxGetComplexDoubles(array);
+      T real_part = pc[index].real;
+      T imag_part = pc[index].imag;
+      #else
       T real_part = *(reinterpret_cast<T*>(mxGetPr(array)) + index);
       T imag_part = *(reinterpret_cast<T*>(mxGetPi(array)) + index);
+      #endif
       *value = std::abs(std::complex<R>(real_part, imag_part));
     } else {
       *value = *(reinterpret_cast<T*>(mxGetData(array)) + index);
@@ -950,8 +957,14 @@ class MxArray {
                        >::type* value) {
     typename R::value_type real_part, imag_part;
     if (mxIsComplex(array)) {
+      #if MX_HAS_INTERLEAVED_COMPLEX
+      mxComplexDouble* pc = mxGetComplexDoubles(array);
+      real_part = pc[index].real;
+      imag_part = pc[index].imag;
+      #else
       real_part = *(reinterpret_cast<T*>(mxGetPr(array)) + index);
       imag_part = *(reinterpret_cast<T*>(mxGetPi(array)) + index);
+      #endif
     } else {
       real_part = *(reinterpret_cast<T*>(mxGetData(array)) + index);
       imag_part = 0.0;
@@ -1009,6 +1022,14 @@ class MxArray {
       T* data_pointer = reinterpret_cast<T*>(mxGetData(array));
       value->assign(data_pointer, data_pointer + array_size);
     } else {
+      #if MX_HAS_INTERLEAVED_COMPLEX
+      mxComplexDouble* pc = mxGetComplexDoubles(array);
+      value->resize(array_size);
+      for (mwSize i = 0; i < array_size; ++i) {
+        double mag = std::abs(std::complex<double>(pc[i].real, pc[i].imag));
+        (*value)[i] = static_cast<T>(mag);
+      }
+      #else
       T* real_part = reinterpret_cast<T*>(mxGetPr(array));
       T* imag_part = reinterpret_cast<T*>(mxGetPi(array));
       value->resize(array_size);
@@ -1018,6 +1039,7 @@ class MxArray {
             static_cast<double>(*(imag_part++))));
         (*value)[i] = static_cast<T>(mag);
       }
+      #endif
     }
   }
   #pragma warning( pop )
@@ -1037,11 +1059,18 @@ class MxArray {
         (*value)[i] = typename R::value_type(*(data_pointer++), 0.0f);
       }
     } else {
+      #if MX_HAS_INTERLEAVED_COMPLEX
+      mxComplexDouble* pc = mxGetComplexDoubles(array);
+      for (mwSize i = 0; i < array_size; ++i) {
+        (*value)[i] = typename R::value_type(pc[i].real, pc[i].imag);
+      }
+      #else
       T* real_part = reinterpret_cast<T*>(mxGetPr(array));
       T* imag_part = reinterpret_cast<T*>(mxGetPi(array));
       for (mwSize i = 0; i < array_size; ++i) {
         (*value)[i] = typename R::value_type(*(real_part++), *(imag_part++));
       }
+      #endif
     }
   }
   /** Explicit char (signed) array assignment.
@@ -1098,8 +1127,14 @@ class MxArray {
                            T
                          >::type& value) {
     if (mxIsComplex(array)) {
+      #if MX_HAS_INTERLEAVED_COMPLEX
+      mxComplexDouble* pc = mxGetComplexDoubles(array);
[…]
