### 2022-10-14 c11b6e3
Change some casts
