### 2022-10-14 0577e6f
Remove default version for 2017
