### 2022-10-17 3058bf4
Change to TriangularMap constructor
