### 2022-10-17 f13131e
Changes to ComposedMap descriptions
