+++ MParT/ComposedMap.h
@@ -13,19 +13,13 @@
 namespace mpart{
 
 /**
- @brief Provides a definition of block lower triangular transport maps.
+ @brief Provides a definition of a transport map defined by the composition of several maps.
  @details
-This class defines a map \f$T:\mathbb{R}^N\rightarrow \mathbb{R}^M\f$ with the block triangular structure
+This class defines a map \f$T:\mathbb{R}^N\rightarrow \mathbb{R}^M\f$ with the composed structure
 \f[
-T(x) = \left[\begin{array}{l} T_1(x_{1:N_1})\\ \vdots \\ T_k(x_{1:N_2})\\ \vdots T_K(x_{1:N}) \end{array} \right],
+T(x) = T_L(x) \circ \ \dots \ \circ T_1(x),
 \f]
-where each component \f$T_i(x_{1:N_i}):\mathbb{R}^{N_i}\rightarrow \mathbb{R}^{M_i}\f$ is a function depending on
-the first \f$N_i\f$ inputs and returning \f$M_i\f$ outputs.  Note that this function must be invertible in the last
-\f$M_i\f$ input arguments.  For example, it must be possible to solve \f$T_i(x_1:{N_i-M_i}, x_{N_i-M_i:N_i}) = r\f$
-for \f$x_{N_i-M_i:N_i}\f$ given a vector \f$r\in\mathbb{R}^{M_i}\f$ and previous components \f$x_1:{N_i-M_i}\f$.
-
-This block triangular form is analogous to a block triangular matrix where each \f$M_i\timesM_i\f$ diagonal block
-is positive definite.
+where each map \f$T_i:\mathbb{R}^{N}\rightarrow \mathbb{R}^{N}\f$ is a square transport map.
 
 Checkpointing can be used for deep maps to reduce the amount of memory that is required for gradient computations.
 The checkpointing logic was adapted from the :code:`revolve` algorithm of `[Griewank and Walther, 2000] <https://dl.acm.org/doi/10.1145/347837.347846>`_,
@@ -39,11 +33,10 @@ public:
 
     /** @brief Construct a block triangular map from a collection of other ConditionalMapBase objects.
 
-    @param components A vector of ConditionalMapBase objects defining each \f$T_k\f$ in the block triangular map.
-                      To maintain the correct block structure, the dimensions of the components must satisfy \f$N_k = N_{k-1}+M_{k}\f$.
+    @param maps A vector of ConditionalMapBase objects defining each \f$T_k\f$ in the composition. Note: each map must be square (i.e., have equal input and output dimension).
     @param maxChecks The maximum number of checkpoints to use during gradient computations.  If maxChecks==1, then no checkpointing will be utilized and all forward states will be recomputed.  If maxChecks==components.size(), then all states will be stored and reused during the backward pass.  This is the most efficient option, but can require an intractable amount of memory for high-dimensional or deep parameterizations.   The default value is -1, which will set the maximum number of checkpoints to be equal to the number of layers (i.e., map.size()). 
     */
-    ComposedMap(std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> const& components,
+    ComposedMap(std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> const& maps,
                 int maxChecks=-1);
 
     virtual ~ComposedMap() = default;
@@ -86,18 +79,6 @@ public:
                       StridedMatrix<double, MemorySpace>              output) override;
 
 
-    /** @brief Evaluates the map inverse.
-
[…]
