### 2022-10-17 2e85320
Adding composedmap to API
