### 2022-10-17 7853057
Switch to DeviceSpace for GPU
