### 2022-10-19 b9d6fcf
Creating slice within ConditionalMapBase
