### 2022-10-19 34e304f
Make Identity, MonoComp, TriMap abide by C.128
