### 2022-10-19 80de6e5
Add in more TriangularMap
