### 2022-10-19 aa63b75
Adding SummarizedMap class and tests, fixed some ComposedMap tests
