+++ tests/Test_SummarizedMap.cpp
@@ -1,0 +1,242 @@
+#include <catch2/catch_all.hpp>
+
+#include "MParT/SummarizedMap.h"
+#include "MParT/MapFactory.h"
+#include "MParT/AffineFunction.h"
+#include <MParT/Utilities/ArrayConversions.h>
+using namespace mpart;
+using namespace Catch;
+using MemorySpace = Kokkos::HostSpace;
+
+TEST_CASE( "SummarizedMap", "[SummarizedMap_MonotoneComponent]" ) {
+
+    MapOptions options;
+    options.basisType = BasisTypes::ProbabilistHermite;
+    options.basisNorm = false;
+
+    unsigned int maxDegree = 2;
+    unsigned int dim = 7;
+    unsigned int lrcRank = 2;
+    FixedMultiIndexSet<MemorySpace> mset(lrcRank+1, maxDegree);
+
+    // make an affine function
+    Eigen::RowMatrixXd A = Eigen::RowMatrixXd::Random(lrcRank, dim-1);
+    Kokkos::View<double**, MemorySpace> viewA(A.data(), lrcRank, dim-1);
+    std::shared_ptr<ParameterizedFunctionBase<MemorySpace>> affineFunc = std::make_shared<AffineFunction<Kokkos::HostSpace>>(viewA);
+    std::shared_ptr<ConditionalMapBase<MemorySpace>> comp = MapFactory::CreateComponent<Kokkos::HostSpace>(mset, options);
+    std::shared_ptr<ConditionalMapBase<MemorySpace>> sumMap = std::make_shared<SummarizedMap<MemorySpace>>(affineFunc, comp);
+
+    CHECK(sumMap->outputDim == 1);
+    CHECK(sumMap->inputDim == dim);
+    CHECK(sumMap->numCoeffs == comp->numCoeffs);
+
+
+    Kokkos::View<double*,Kokkos::HostSpace> coeffs("Coefficients", sumMap->numCoeffs);
+    for(unsigned int i=0; i<sumMap->numCoeffs; ++i)
+        coeffs(i) = 0.1*(i+1);
+        
+    SECTION("Coefficients"){
+        
+        // Set the coefficients of the triangular map
+        sumMap->SetCoeffs(coeffs);
+
+        // Now make sure that the coefficients of each block were set
+        unsigned int cumCoeffInd = 0;
+        for(unsigned int i=0; i<sumMap->numCoeffs; ++i){
+            CHECK(sumMap->Coeffs()(i) == 0.1*(i+1)); // Values of coefficients should be correct
+            CHECK(sumMap->Coeffs()(i) == comp->Coeffs()(i)); // Values of coefficients should be equal to those of comp
+            CHECK(&sumMap->Coeffs()(i) == &comp->Coeffs()(i)); // Memory location should also be the same (no copy)
+        
+        }
+    }
+
+
+    unsigned int numSamps = 10;
+    Kokkos::View<double**, Kokkos::HostSpace> in("Map Input", dim, numSamps);
+    for(unsigned int i=0; i<dim; ++i){
+        for(unsigned int j=0; j<numSamps; ++j){
+            in(i,j) = double(i)/dim + double(j)/numSamps;
+        }
+    }
+
+    sumMap->SetCoeffs(coeffs);
+    Kokkos::View<double**, Kokkos::HostSpace> out = sumMap->Evaluate(in);
+    
+    // SECTION("Evaluation"){
+
+    //     for(unsigned int i=0; i<numBlocks; ++i){
+            
+    //         auto outBlock = blocks.at(i)->Evaluate(Kokkos::subview(in, std::make_pair(0,int(i+1+extraInputs)), Kokkos::ALL()));
+
[…]
