+++ src/CMakeLists.txt
@@ -36,4 +36,5 @@ target_sources(mpart
     Initialization.cpp
     ComposedMap.cpp
     IdentityMap.cpp
+    SummarizedMap.cpp
 )

[…]
