### 2022-10-19 fd5dd91
Add Slicing for a lot of classes
