### 2022-10-19 f8d4219
affix kokkos to 3.6.01
