+++ tests/Test_TriangularMap.cpp
@@ -226,15 +226,13 @@ TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_
         auto slice = triMap->Slice(sliceBegin, sliceEnd);
         REQUIRE(slice->inputDim == triMap->inputDim);
         REQUIRE(slice->outputDim == sliceEnd-sliceBegin);
+
+        auto outSlice = slice->Evaluate(in);
         // Just test the evaluation
         for(unsigned int i=sliceBegin; i<sliceEnd; ++i){
-
-            auto outBlock = blocks.at(i)->Evaluate(Kokkos::subview(in, std::make_pair(0,int(i+1+extraInputs)), Kokkos::ALL()));
-
-            REQUIRE(outBlock.extent(1)==numSamps);
-            REQUIRE(outBlock.extent(0)==1);
-            for(unsigned int j=0; j<numSamps; ++j)
-                CHECK( out(i,j) == Approx(outBlock(0,j)).epsilon(1e-6));
+            for(unsigned int j=0; j<numSamps; ++j){
+                CHECK(outSlice(i-sliceBegin,j) == Approx(out(i,j)));
+            }
         }
     }*/
 
