+++ bindings/julia/src/TriangularMap.cpp
@@ -15,6 +15,5 @@ void mpart::binding::TriangularMapWrapper(jlcxx::Module &mod) {
             map.InverseInplace(JuliaToKokkos(x), JuliaToKokkos(r));
        })
        .method("GetComponent", [](TriangularMap<Kokkos::HostSpace>& map, int i){ return map.GetComponent(i-1); })
-       .method("Slice", [](TriangularMap<Kokkos::HostSpace>& map, int begin, int end){ return map.Slice(begin-1, end); })
     ;
 }
