### 2022-10-19 cde0fe6
Working on slice
