### 2022-10-21 a4d2ab4
Testing SingleComponemtMap class/factory method
Not sure which one to use yet.
