+++ src/SingleEntryMap.cpp
@@ -1,0 +1,171 @@
+#include "MParT/SingleEntryMap.h"
+#include "MParT/TriangularMap.h"
+#include "MParT/IdentityMap.h"
+#include "MParT/Utilities/KokkosSpaceMappings.h"
+
+#include <numeric>
+
+using namespace mpart;
+
+template<typename MemorySpace>
+SingleEntryMap<MemorySpace>::SingleEntryMap(const unsigned int dim, const unsigned int activeInd, std::shared_ptr<ConditionalMapBase<MemorySpace>> const& component) : 
+                        ConditionalMapBase<MemorySpace>(dim, dim, component->numCoeffs)
+{
+
+    // Check that active index is not greater than map dimension
+    if(dim < activeInd){
+        std::stringstream msg;
+        msg << "In SingleEntryMap constructor, the active index can't be greater than map dimension. Got dim = " << dim << " and activeInd = " << activeInd << ".";
+        throw std::invalid_argument(msg.str());
+    }
+
+    // Check that the input dimension of the component matches the activeInd
+    if(activeInd != component->inputDim){
+        std::stringstream msg;
+        msg << "In SingleEntryMap constructor, the component input dimension must be equal to the active index. Got dim = " << component->inputDim << " and activeInd = " << activeInd << ".";
+        throw std::invalid_argument(msg.str());
+    }
+
+    std::shared_ptr<ConditionalMapBase<MemorySpace>> map;
+    // Construct map using SingleEntryMap constructor
+    
+    if(activeInd == 1){  // special case if activeInd = 1, map is of form [T_1; Id]
+
+        // Bottom identity map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> botIdMap = std::make_shared<IdentityMap<Kokkos::HostSpace>>(dim, dim-activeInd);
+        
+        // fill a vector of components with identity, active component, identity
+        std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> blocks(2);
+        blocks.at(0) = component;
+        blocks.at(1) = botIdMap;
+
+        // make map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> map = std::make_shared<TriangularMap<MemorySpace>>(blocks);
+
+
+    }
+    else if (activeInd == dim){  // special case if activeInd = dim, map is of form [Id; T_d]
+        // Top identity map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> topIdMap = std::make_shared<IdentityMap<Kokkos::HostSpace>>(activeInd-1, activeInd-1);
+
+        // fill a vector of components with identity, active component, identity
+        std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> blocks(2);
+        blocks.at(0) = topIdMap;
+        blocks.at(1) = component;
+
+        // make map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> map = std::make_shared<TriangularMap<MemorySpace>>(blocks);
+    }
+    else{ // general case, map is of form [Id; T_i; Id]
+
+        // Top identity map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> topIdMap = std::make_shared<IdentityMap<Kokkos::HostSpace>>(activeInd-1, activeInd-1);
+        
+        // Bottom identity map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> botIdMap = std::make_shared<IdentityMap<Kokkos::HostSpace>>(dim, dim-activeInd);
+        
+        // fill a vector of components with identity, active component, identity
[…]
