+++ tests/Test_SingleEntryMap.cpp
@@ -1,0 +1,252 @@
+#include <catch2/catch_all.hpp>
+
+#include "MParT/SingleEntryMap.h"
+#include "MParT/MapFactory.h"
+
+#include "MParT/Utilities/LinearAlgebra.h"
+
+using namespace mpart;
+using namespace Catch;
+using MemorySpace = Kokkos::HostSpace;
+
+TEST_CASE( "Testing SingleEntryMap", "[ShallowComposedMap]" ) {
+
+    MapOptions options;
+    options.basisType = BasisTypes::ProbabilistHermite;
+    options.basisNorm = false;
+
+    unsigned int maxDegree = 2;
+    unsigned int dim = 7;
+    unsigned int activeInd = 2;
+    FixedMultiIndexSet<MemorySpace> mset(activeInd, maxDegree);
+    std::shared_ptr<ConditionalMapBase<MemorySpace>> comp = MapFactory::CreateComponent<Kokkos::HostSpace>(mset, options);
+    std::shared_ptr<ConditionalMapBase<MemorySpace>> singEntMap = std::make_shared<SingleEntryMap<Kokkos::HostSpace>>(dim,activeInd,comp);
+    
+
+    // std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> maps(numMaps);
+    // for(unsigned int i=0;i<numMaps;++i){
+
+    //     maps.at(i) = MapFactory::CreateTriangular<Kokkos::HostSpace>(dim, dim, order, options);
+    //     coeffSize += maps.at(i)->numCoeffs;
+    // }
+
+    // std::shared_ptr<ConditionalMapBase<MemorySpace>> composedMap = std::make_shared<ComposedMap<MemorySpace>>(maps);
+
+    // CHECK(composedMap->outputDim == dim);
+    // CHECK(composedMap->inputDim == dim);
+    // CHECK(composedMap->numCoeffs == coeffSize);
+
+
+    // Eigen::RowVectorXd coeffs(composedMap->numCoeffs);
+    // for(unsigned int i=0; i<composedMap->numCoeffs; ++i)
+    //     coeffs(i) = 0.1*(i+1);
+    
+    // SECTION("Coefficients"){
+        
+    //     // Set the coefficients of the triangular map
+    //     composedMap->SetCoeffs(coeffs);
+
+    //     // Now make sure that the coefficients of each block were set
+    //     unsigned int cumCoeffInd = 0;
+    //     for(unsigned int i=0; i<numMaps; ++i){
+    //         for(unsigned int j=0; j<maps.at(i)->numCoeffs; ++j){
+    //             CHECK(maps.at(i)->Coeffs()(j) == coeffs(cumCoeffInd)); // Values of coefficients should be equal
+    //             CHECK(&maps.at(i)->Coeffs()(j) == &composedMap->Coeffs()(cumCoeffInd)); // Memory location should also be the same (no copy)
+    //             cumCoeffInd++;
+    //         }
+    //     }
+    // }
+
+
+    // unsigned int numSamps = 10;
+    // Kokkos::View<double**, Kokkos::HostSpace> in("Map Input", dim, numSamps);
+    // //Eigen::RowMatrixXd in(dim, numSamps);
+    // for(unsigned int i=0; i<dim; ++i){
+    //     for(unsigned int j=0; j<numSamps; ++j){
+    //         in(i,j) = double(i)/(dim) + double(j)/numSamps;
+    //     }
+    // }
+
+    // composedMap->SetCoeffs(coeffs);
+    // auto out = composedMap->Evaluate(in);
+    
+    // SECTION("Evaluate"){
+    //     Kokkos::View<double**, Kokkos::HostSpace> trueOut("True output", dim, numSamps);
[…]
