+++ src/CMakeLists.txt
@@ -37,4 +37,5 @@ target_sources(mpart
     ComposedMap.cpp
     IdentityMap.cpp
     SummarizedMap.cpp
+    SingleEntryMap.cpp
 )

[…]
