### 2022-10-24 ded46a4
Error in SingleEntryMap factory method
