+++ MParT/MapFactory.h
@@ -61,6 +61,25 @@ namespace mpart{
             @param options Additional options that will be passed on to CreateComponent to construct each MonotoneComponent.
          */
         template<typename MemorySpace>
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> CreateSingleEntryMap(unsigned int dim,
+                                                                              unsigned int activeInd,
+                                                                              std::shared_ptr<ConditionalMapBase<MemorySpace>> const &comp);
+
+                                                                                  /**
+            @brief Constructs a triangular map with MonotoneComponents for each block.  A total order multiindex
+                   set is used to define the MonotoneComponent.
+
+            @details For more control over the individual components, consider constructing components with
+                     MapFactory::CreateComponent function and then manually constructing and instance of the
+                     TriangularMap class.
+            @param inputDim The dimension of the input to this map.  Will be the number of inputs passed to
+                            the last component of the triangular map.
+            @param outputDim The output dimension of the map.  Note that this must be less than or equal to
+                             the input dimension.
+            @param totalOrder The total order used to define the parameterization of each MonotoneComponent.
+            @param options Additional options that will be passed on to CreateComponent to construct each MonotoneComponent.
+         */
+        template<typename MemorySpace>
         std::shared_ptr<ConditionalMapBase<MemorySpace>> CreateTriangular(unsigned int inputDim,
                                                                           unsigned int outputDim,
                                                                           unsigned int totalOrder,
