+++ src/MapFactory.cpp
@@ -20,6 +20,82 @@ std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateCompon
     return CompFactoryImpl<MemorySpace>::GetFactoryFunction(opts)(mset,opts);
 }
 
+
+template<typename MemorySpace>
+std::shared_ptr<ConditionalMapBase<MemorySpace>> mpart::MapFactory::CreateSingleEntryMap(unsigned int dim,
+                                                                                         unsigned int activeInd,
+                                                                                         std::shared_ptr<ConditionalMapBase<MemorySpace>> const &comp)
+{
+
+    // Check that active index is not greater than map dimension
+    if(dim < activeInd){
+        std::stringstream msg;
+        msg << "In CreateSingleEntryMap, the active index can't be greater than map dimension. Got dim = " << dim << " and activeInd = " << activeInd << ".";
+        throw std::invalid_argument(msg.str());
+    }
+
+    // Check that the input dimension of the component matches the activeInd
+    if(activeInd != comp->inputDim){
+        std::stringstream msg;
+        msg << "In CreateSingleEntryMap, the component input dimension must be equal to the active index. Got dim = " << comp->inputDim << " and activeInd = " << activeInd << ".";
+        throw std::invalid_argument(msg.str());
+    }
+
+    std::shared_ptr<ParameterizedFunctionBase<MemorySpace>> output;
+    // Construct map using TriangularMap constructor
+    
+    if(activeInd == 1){  // special case if activeInd = 1, map is of form [T_1; Id]
+
+        // Bottom identity map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> botIdMap = std::make_shared<IdentityMap<Kokkos::HostSpace>>(dim, dim-activeInd);
+        
+        // fill a vector of components with identity, active component, identity
+        std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> blocks(2);
+        blocks.at(0) = comp;
+        blocks.at(1) = botIdMap;
+
+        // make map
+        output = std::make_shared<TriangularMap<MemorySpace>>(blocks);
+
+
+    }
+    else if (activeInd == dim){  // special case if activeInd = dim, map is of form [Id; T_d]
+        // Top identity map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> topIdMap = std::make_shared<IdentityMap<Kokkos::HostSpace>>(activeInd-1, activeInd-1);
+
+        // fill a vector of components with identity, active component, identity
+        std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> blocks(2);
+        blocks.at(0) = topIdMap;
+        blocks.at(1) = comp;
+
+        // make map
+        output = std::make_shared<TriangularMap<MemorySpace>>(blocks);
+    }
+    else{ // general case, map is of form [Id; T_i; Id]
+
+        // Top identity map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> topIdMap = std::make_shared<IdentityMap<Kokkos::HostSpace>>(activeInd-1, activeInd-1);
+        
+        // Bottom identity map
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> botIdMap = std::make_shared<IdentityMap<Kokkos::HostSpace>>(dim, dim-activeInd);
+        
+        // fill a vector of components with identity, active component, identity
+        std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> blocks(3);
+        blocks.at(0) = topIdMap;
+        blocks.at(1) = comp;
+        blocks.at(2) = botIdMap;
+
+        // make map
[…]
