### 2022-10-24 2c7f6bc
Instantiate CreateSingleEntryMap
