+++ bindings/python/tests/test_MapFactory.py
@@ -70,3 +70,13 @@ def test_CreateTriangular():
     triangular = mpart.CreateTriangular(2,2,2,opts)
     assert triangular.numCoeffs == 2 + 7 
 
+
+def test_CreateSingleEntryMap_1():
+    dim = 7
+    activeInd = 1
+
+    mset_csemap = mpart.MultiIndexSet.CreateTotalOrder(activeInd, 3, noneLim)  
+    component = mpart.CreateComponent(mset_csemap.fix(True), opts)
+
+    single_entry_map = mpart.CreateSingleEntryMap(dim, activeInd, component)
+    assert single_entry_map.numCoeffs == component.numCoeffs
