+++ src/SummarizedMap.cpp
@@ -34,8 +34,15 @@ template<typename MemorySpace>
 void SummarizedMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, Kokkos::HostSpace> coeffs)
 {
     // SetCoeffs for the component (not the summary function)
+    std::cout << "start SetCoeffs" << std::endl;
     ConditionalMapBase<MemorySpace>::SetCoeffs(coeffs);
+    std::cout << "called SetCoeffs of CondMapBase" << std::endl;
+    std::cout << "num coeffs of map_" << std::endl;
+    std::cout << map_->numCoeffs << std::endl;
+    std::cout << "num coeffs of this->savedCoeffs" << std::endl;
+    std::cout << this->savedCoeffs.size() << std::endl;
     map_->WrapCoeffs(this->savedCoeffs);
+    std::cout << "called WrapCoeffs of component" << std::endl;
 
     
 }
