+++ bindings/python/src/Wrapper.cpp
@@ -13,9 +13,12 @@ PYBIND11_MODULE(pympart, m) {
     ParameterizedFunctionBaseWrapper<Kokkos::HostSpace>(m);
     ConditionalMapBaseWrapper<Kokkos::HostSpace>(m);
     TriangularMapWrapper<Kokkos::HostSpace>(m);
+    ComposedMapWrapper<Kokkos::HostSpace>(m);
+    SummarizedMapWrapper<Kokkos::HostSpace>(m);
     MapFactoryWrapper<Kokkos::HostSpace>(m);
     AffineMapWrapperHost(m);
     AffineFunctionWrapperHost(m);
+
 
 #if defined(MPART_ENABLE_GPU)
     ParameterizedFunctionBaseWrapper<mpart::DeviceSpace>(m);
