+++ docs/source/installation.rst
@@ -38,7 +38,7 @@ MParT uses CMake to handle dependencies and compiler configurations.   A basic b
    cd build
    cmake                                               \
      -DCMAKE_INSTALL_PREFIX=<your/MParT/install/path>  \
-     -DKokkos_ENABLE_PTHREAD=ON                        \
+     -DKokkos_ENABLE_THREADS=ON                        \
    ..
    make install
 
@@ -53,7 +53,7 @@ This installation should also automatically install and build Kokkos, Eigen, and
      -DKokkos_ROOT=<your/kokkos/install/root>          \
      -DEigen3_ROOT=<your/eigen3/install/root>          \
      -DCatch2_ROOT=<your/catch2/install/root>          \
-     -DKokkos_ENABLE_PTHREAD=ON                        \
+     -DKokkos_ENABLE_THREADS=ON                        \
      -DKokkos_ENABLE_SERIAL=ON                         \
    ..
 
@@ -67,14 +67,14 @@ Note that if you do not wish to compile bindings for Python, Julia, or Matlab, y
 
     cmake                                              \
      -DCMAKE_INSTALL_PREFIX=<your/MParT/install/path>  \
-     -DKokkos_ENABLE_PTHREAD=ON                        \
+     -DKokkos_ENABLE_THREADS=ON                        \
      -DMPART_PYTHON=OFF                                \
      -DMPART_MATLAB=OFF                                \
      -DMPART_JULIA=OFF                                 \
    ..
 
 
-MParT is built on Kokkos, which provides a single interface to many different multithreading capabilities like pthreads, OpenMP, CUDA, and OpenCL.   A list of available backends can be found on the [Kokkos wiki](https://github.com/kokkos/kokkos/blob/master/BUILD.md#device-backends).   The `Kokkos_ENABLE_PTHREAD` option in the CMake configuration above can be changed to reflect different choices in device backends.   The OSX-provided clang compiler does not support OpenMP, so `PTHREAD` is a natural choice for CPU-based multithreading on OSX.   However, you may find that OpenMP has slightly better performance with other compilers and operating systems.
+MParT is built on Kokkos, which provides a single interface to many different multithreading capabilities like pthreads, OpenMP, CUDA, and OpenCL.   A list of available backends can be found on the [Kokkos wiki](https://github.com/kokkos/kokkos/blob/master/BUILD.md#device-backends).   The `Kokkos_ENABLE_THREADS` option in the CMake configuration above can be changed to reflect different choices in device backends.   The OSX-provided clang compiler does not support OpenMP, so `PTHREAD` is a natural choice for CPU-based multithreading on OSX.   However, you may find that OpenMP has slightly better performance with other compilers and operating systems.
 
 Tests
 ---------
@@ -169,7 +169,7 @@ Make sure that this file includes a full installation path from root! At this po
 Compiling with CUDA Support
 ----------------------------
 
-Building the Kokkos Dependency 
+Building the Kokkos Dependency
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 To support a GPU at the moment, you need a few special requirements. Due to the way that Kokkos handles GPU code, MParT must be compiled using a special wrapper around NVCC that Kokkos provides.  Because of this, MParT cannot use an internal build of Kokkos and Kokkos must therefore be compiled (or otherwise installed) manually.
 
[…]
