### 2022-10-26 d710eb2
Print out pointer info
