+++ src/SummarizedMap.cpp
@@ -7,10 +7,10 @@
 using namespace mpart;
 
 template<typename MemorySpace>
-SummarizedMap<MemorySpace>::SummarizedMap(std::shared_ptr<ParameterizedFunctionBase<MemorySpace>> const& summaryFunction, 
-                                          std::shared_ptr<ConditionalMapBase<MemorySpace>> const& map) : 
+SummarizedMap<MemorySpace>::SummarizedMap(std::shared_ptr<ParameterizedFunctionBase<MemorySpace>> const& summaryFunction,
+                                          std::shared_ptr<ConditionalMapBase<MemorySpace>> const& map) :
                                           ConditionalMapBase<MemorySpace>(summaryFunction->inputDim + 1, map->outputDim, map->numCoeffs),
-                                          summaryFunction_(summaryFunction), 
+                                          summaryFunction_(summaryFunction),
                                           map_(map)
 {
 
@@ -26,6 +26,7 @@ SummarizedMap<MemorySpace>::SummarizedMap(std::shared_ptr<ParameterizedFunctionB
         msg << "SummarizedMap: input dimension of map component must be 1 + output dimension of summaryFunction, but was given map->inputDim = " << map_->inputDim << " and summaryFunction->outputDim + 1 = " << summaryFunction_->outputDim + 1 << ".";
         throw std::invalid_argument(msg.str());
     }
+    std::cout << "constructor: map_.get() = " << map_.get() << std::endl;
 }
 
 
@@ -39,11 +40,11 @@ void SummarizedMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, Kokkos::HostSpa
     ConditionalMapBase<MemorySpace>::SetCoeffs(coeffs);
     std::cout << "called SetCoeffs of CondMapBase" << std::endl;
     std::cout << "num coeffs of map_" << std::endl;
-    
+
     std::cout << map_->numCoeffs << std::endl;
     std::cout << "num coeffs of this->savedCoeffs" << std::endl;
     std::cout << this->savedCoeffs.size() << std::endl;
-    
+
     map_->WrapCoeffs(this->savedCoeffs);
     std::cout << "called WrapCoeffs of component" << std::endl;
 
@@ -81,7 +82,7 @@ void SummarizedMap<MemorySpace>::WrapCoeffs(Kokkos::View<double*, Kokkos::Defaul
 template<typename MemorySpace>
 void SummarizedMap<MemorySpace>::SummarizePts(StridedMatrix<const double, MemorySpace> const& pts,
                                               StridedMatrix<double, MemorySpace>              output)
-{   
+{
     // Split pts into the part that is summarized and the part that is not
 
     StridedMatrix<const double, MemorySpace> ptsToSummarize = Kokkos::subview(pts, std::make_pair(0, int(summaryFunction_->inputDim)), Kokkos::ALL());
@@ -95,7 +96,7 @@ void SummarizedMap<MemorySpace>::SummarizePts(StridedMatrix<const double, Memory
     Kokkos::deep_copy(outputSummary, summary);
 
     // Copy non summarized pts
-    Kokkos::View<double**, MemorySpace> outputAfterSummary = Kokkos::subview(output, std::make_pair(int(summaryFunction_->outputDim),int(summaryFunction_->outputDim+1)), Kokkos::ALL());        
[…]
