### 2022-10-26 857e9c1
Work on Gaussian
