### 2022-10-27 eec8103
Work on Gaussian
