+++ src/Distributions/GaussianDistribution.cpp
@@ -16,16 +16,19 @@ template<typename MemorySpace>
 GaussianDistribution<MemorySpace>::GaussianDistribution(StridedVector<double, MemorySpace> mean): mean_(mean), idCov_(true) {}
 
 template<typename MemorySpace>
-StridedVector<double, MemorySpace> GaussianDistribution<MemorySpace>::LogDensityImpl(StridedMatrix<const double, MemorySpace> const &pts, StridedVector<double, MemorySpace> out) {
+void GaussianDistribution<MemorySpace>::LogDensityImpl(StridedMatrix<const double, MemorySpace> const &pts, StridedVector<double, MemorySpace> output) {
     // Compute the log density
-    StridedVector<double, MemorySpace> diff;
+    StridedMatrix<double, MemorySpace> diff;
+    int M = pts.extent(0);
+    int N = pts.extent(1);
+    Kokkos::MDRangePolicy<Kokkos::Rank<2>, typename MemoryToExecution<MemorySpace>::Space> policy({{0, 0}}, {{N, M}});
     if(mean_.extent(0) == 0){
         diff = pts;
     }
     else {
-        diff = StridedVector<double, MemorySpace>("diff", pts.extent(0));
-        Kokkos::parallel_for("diff", pts.extent(0), KOKKOS_LAMBDA(const int i) {
-            diff(i) = pts(i) - mean_(i);
+        diff = StridedMatrix<double, MemorySpace>("diff", M, N);
+        Kokkos::parallel_for(policy, KOKKOS_LAMBDA(const int& j, const int& i) {
+            diff(i,j) = pts(i,j) - mean_(i);
         });
     }
 
@@ -33,16 +36,15 @@ StridedVector<double, MemorySpace> GaussianDistribution<MemorySpace>::LogDensity
         covLU_.solveInPlace(diff);
     }
 
-    double norm = 0.;
-    Kokkos::parallel_reduce( "norm", diff.extent(0), KOKKOS_LAMBDA (const int i, double& lsum) {
-        lsum += diff(i)*diff(i);
-    }, norm);
+    Kokkos::parallel_for( "log terms", N, KOKKOS_LAMBDA (const int& j) {
+        output(j) = -0.5*( M*logtau_ + logDet_ );
+    });
 
+    Kokkos::parallel_for( "weighted norm", policy, KOKKOS_LAMBDA (const int& j, const int& i) {
+        output(j) -= 0.5*diff(i,j)*diff(i,j);
+    });
+}
 
-    auto ptsMinusMeanT = ptsMinusMean.transpose();
-    auto covarInv = covLU_.solve(StridedMatrix<double, MemorySpace>(ptsMinusMeanT));
-    auto logDensity = -0.5*(ptsMinusMeanT*covarInv).diagonal() - 0.5*mean_.size()*log(tau) - 0.5*logDetCov_;
-    // Copy the output
-    Kokkos::deep_copy(out, logDensity);
-    return out;
+void GaussianDistribution<MemorySpace>::GradLogDensityImpl(StridedMatrix<const double, MemorySpace> const &pts, StridedMatrix<double, MemorySpace> output) {
+
 }
