+++ src/Utilities/LinearAlgebra.cpp
@@ -7,7 +7,7 @@ using namespace mpart;
 
 
 template<>
-void mpart::dgemm(double                                     alpha, 
+void mpart::dgemm(double                                     alpha,
                   TransposeObject<Kokkos::HostSpace>         A,
                   TransposeObject<Kokkos::HostSpace>         B,
                   double                                     beta,
@@ -59,7 +59,7 @@ void PartialPivLU<Kokkos::HostSpace>::compute(Kokkos::View<const double**,Kokkos
 
 template<>
 void PartialPivLU<Kokkos::HostSpace>::solveInPlace(Kokkos::View<double**,Kokkos::LayoutLeft,Kokkos::HostSpace> x)
-{   
+{
     auto eigX = KokkosToMat(x);
     eigX = (luSolver_->permutationP() * eigX).eval();
     luSolver_->matrixLU().template triangularView<Eigen::UnitLower>().solveInPlace(eigX);
@@ -70,7 +70,7 @@ void PartialPivLU<Kokkos::HostSpace>::solveInPlace(Kokkos::View<double**,Kokkos:
 template<>
 double PartialPivLU<Kokkos::HostSpace>::determinant() const
 {
-    assert(isComputed); 
+    assert(isComputed);
     return luSolver_->determinant();
 }
 
@@ -79,17 +79,17 @@ double PartialPivLU<Kokkos::HostSpace>::determinant() const
 #if defined(MPART_ENABLE_GPU)
 
 template<>
-void mpart::dgemm(double                             alpha, 
+void mpart::dgemm(double                             alpha,
            TransposeObject<mpart::DeviceSpace>       A,
            TransposeObject<mpart::DeviceSpace>       B,
            double                                    beta,
            StridedMatrix<double, mpart::DeviceSpace> C)
-{   
+{
     // Versions of A,B, and C with the column major layout expected by cublas
     StridedMatrix<const double, mpart::DeviceSpace> ALeft;
     StridedMatrix<const double, mpart::DeviceSpace> BLeft;
     StridedMatrix<double, mpart::DeviceSpace> CLeft;
-    
+
     // Make sure the ptsLeft array is column major
     if(A.view.stride_0()==1){
         ALeft = A.view;
@@ -106,7 +106,7 @@ void mpart::dgemm(double                             alpha,
         Kokkos::deep_copy(temp, B.view);
         BLeft = temp;
     }
-    
+
     bool copyOut = false;
     if(C.stride_0()==1){
         CLeft = C;
@@ -115,14 +115,14 @@ void mpart::dgemm(double                             alpha,
         Kokkos::deep_copy(CLeft, C);
         copyOut = true;
     }
-    
+
     int ldA = ALeft.stride_1(); // Spacing between columns int A
     int ldB = BLeft.stride_1(); // Spacing between columns int B
     int ldC = CLeft.stride_1(); // Spacing between columns int C
 
     // Perform the matrix multiplication
     cublasStatus_t info =  cublasDgemm(GetInitializeStatusObject().GetCublasHandle(),
-                                       A.isTransposed ? CUBLAS_OP_T : CUBLAS_OP_N, 
+                                       A.isTransposed ? CUBLAS_OP_T : CUBLAS_OP_N,
                                        B.isTransposed ? CUBLAS_OP_T : CUBLAS_OP_N,
                                        A.rows(), B.cols(), A.cols(),
                                        &alpha,
@@ -130,7 +130,7 @@ void mpart::dgemm(double                             alpha,
                                        BLeft.data(), ldB,
                                        &beta,
                                        CLeft.data(), ldC);
-    
+
     // The layouts didn't match, so we have to copy back
     if(copyOut)
[…]
