+++ MParT/Utilities/LinearAlgebra.h
@@ -11,7 +11,7 @@
 #include <cuda_runtime.h>
 #include <cublas_v2.h>
 #include <cusolverDn.h>
-#endif 
+#endif
 
 namespace mpart{
 
@@ -20,11 +20,11 @@ namespace mpart{
  x += y
  */
 template<typename... Traits1, typename... Traits2>
-void AddInPlace(Kokkos::View<double*, Traits1...> x, 
+void AddInPlace(Kokkos::View<double*, Traits1...> x,
                 Kokkos::View<const double*, Traits2...> y)
-{   
+{
     assert(x.extent(0)==y.extent(0));
-    
+
     using ExecSpace = typename MemoryToExecution<typename Kokkos::View<double*, Traits1...>::memory_space>::Space;
     Kokkos::RangePolicy<ExecSpace> policy(0, x.extent(0));
 
@@ -46,7 +46,7 @@ void AddInPlace(Kokkos::View<double*, Traits1...> x,
  x += y
  */
 template<typename... Traits1, typename... Traits2>
-void AddInPlace(Kokkos::View<double**, Traits1...> x, 
+void AddInPlace(Kokkos::View<double**, Traits1...> x,
                 Kokkos::View<const double**, Traits2...> y)
 {
     assert(x.extent(0)==y.extent(0));
@@ -70,10 +70,10 @@ void AddInPlace(Kokkos::View<double**, Traits1...> x,
 }
 
 /**
-z = x + y 
+z = x + y
 */
 template<typename ScalarType, typename... Traits1, typename... Traits2>
-Kokkos::View<ScalarType*, typename Kokkos::View<ScalarType*, Traits1...>::memory_space> operator+(Kokkos::View<const ScalarType*, Traits1...> x, 
+Kokkos::View<ScalarType*, typename Kokkos::View<ScalarType*, Traits1...>::memory_space> operator+(Kokkos::View<const ScalarType*, Traits1...> x,
                                                  Kokkos::View<const ScalarType*, Traits2...> y)
 {
     Kokkos::View<ScalarType*, typename Kokkos::View<ScalarType*, Traits1...>::memory_space> z("x+y", x.extent(0));
@@ -82,7 +82,7 @@ Kokkos::View<ScalarType*, typename Kokkos::View<ScalarType*, Traits1...>::memory
     return z;
 }
 template<typename ScalarType, typename... Traits1, typename... Traits2>
-Kokkos::View<ScalarType**, typename Kokkos::View<ScalarType**, Traits1...>::memory_space> operator+(Kokkos::View<const ScalarType**, Traits1...> x, 
+Kokkos::View<ScalarType**, typename Kokkos::View<ScalarType**, Traits1...>::memory_space> operator+(Kokkos::View<const ScalarType**, Traits1...> x,
                                                   Kokkos::View<const ScalarType**, Traits2...> y)
 {
     Kokkos::View<ScalarType**, typename Kokkos::View<ScalarType**, Traits1...>::memory_space> z("x+y", x.extent(0), y.extent(1));
@@ -91,7 +91,7 @@ Kokkos::View<ScalarType**, typename Kokkos::View<ScalarType**, Traits1...>::memo
     return z;
 }
 template<typename ScalarType, typename... Traits1, typename... Traits2>
-Kokkos::View<ScalarType*, typename Kokkos::View<ScalarType*, Traits1...>::memory_space> operator+(Kokkos::View<ScalarType*, Traits1...> x, 
+Kokkos::View<ScalarType*, typename Kokkos::View<ScalarType*, Traits1...>::memory_space> operator+(Kokkos::View<ScalarType*, Traits1...> x,
[…]
