### 2022-10-27 36a5569
Working binding error
Similar error occurs for a factory method that avoids python making a summarized component
