+++ bindings/python/src/MapFactory.cpp
@@ -1,5 +1,6 @@
 #include "CommonPybindUtilities.h"
 #include "MParT/MapFactory.h"
+#include "MParT/ConditionalMapBase.h"
 #include <pybind11/stl.h>
 #include <pybind11/eigen.h>
 
@@ -22,7 +23,19 @@ void mpart::binding::MapFactoryWrapper(py::module &m)
     // CreateSingleEntryMap
     m.def(isDevice? "dCreateSingleEntryMap" : "CreateSingleEntryMap", &MapFactory::CreateSingleEntryMap<MemorySpace>);
 
+    // CreateSingleEntryMap
+    m.def("CreateAffineLRCMap", [] (unsigned int dim,
+                                    unsigned int activeInd,
+                                    Eigen::Ref<Eigen::MatrixXd> const& summaryMatrix,
+                                    unsigned int maxDegree,
+                                    MapOptions options)
+        {
+            
+            Kokkos::View<double**, MemorySpace> summaryMatrixView = MatToKokkos<double, Kokkos::HostSpace>(summaryMatrix);
+            std::shared_ptr<ConditionalMapBase<MemorySpace>> lrcMap = MapFactory::CreateAffineLRCMap(dim, activeInd, summaryMatrixView, maxDegree, options);
 
+            return lrcMap;
+        });
 }
 
 template void mpart::binding::MapFactoryWrapper<Kokkos::HostSpace>(py::module&);
