### 2022-10-28 082134a
Add CreateSummarizedMap
