### 2022-10-28 a43206d
Quest to find the weird binding bug in summarizedmap.
Plan is to progressively build up the SummarizedMap class to see when point mischief starts. DebugMap0 take a single component as input and checks its pointer
