+++ bindings/python/src/MapFactory.cpp
@@ -24,6 +24,8 @@ void mpart::binding::MapFactoryWrapper(py::module &m)
     m.def(isDevice? "dCreateSingleEntryMap" : "CreateSingleEntryMap", &MapFactory::CreateSingleEntryMap<MemorySpace>);
 
     m.def(isDevice? "dCreateSummarizedMap" : "CreateSummarizedMap", &MapFactory::CreateSummarizedMap<MemorySpace>);
+
+    m.def(isDevice? "dCreateDebugMap" : "CreateDebugMap", &MapFactory::CreateDebugMap<MemorySpace>);
     // CreateSingleEntryMap
     m.def("CreateAffineLRCMap", [] (unsigned int dim,
                                     unsigned int activeInd,
