+++ MParT/DebugMap.h
@@ -1,5 +1,5 @@
-#ifndef MPART_TRIANGULARMAP_H
-#define MPART_TRIANGULARMAP_H
+#ifndef MPART_DEBUGMAP_H
+#define MPART_DEBUGMAP_H
 
 #include "MParT/ConditionalMapBase.h"
 #include "MParT/Utilities/Miscellaneous.h"
@@ -28,7 +28,7 @@ is positive definite.
 
  */
 template<typename MemorySpace>
-class TriangularMap : public ConditionalMapBase<MemorySpace>{
+class DebugMap : public ConditionalMapBase<MemorySpace>{
         
 public:
 
@@ -37,13 +37,13 @@ public:
     @param components A vector of ConditionalMapBase objects defining each \f$T_k\f$ in the block triangular map.
                       To maintain the correct block structure, the dimensions of the components must satisfy \f$N_k = N_{k-1}+M_{k}\f$.
     */
-    TriangularMap(std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> const& components);
+    DebugMap(std::shared_ptr<ConditionalMapBase<MemorySpace>> const& component);
 
-    virtual ~TriangularMap() = default;
+    virtual ~DebugMap() = default;
 
     /** @brief Sets the coefficients for all components of the map.
 
-    @details This function will copy the provided coeffs vectors into the savedCoeffs object in the TriangularMap class.   To avoid
+    @details This function will copy the provided coeffs vectors into the savedCoeffs object in the DebugMap class.   To avoid
     duplicating the coefficients, the savedCoeffs member variable for each component will then be set to a subview of this vector.
     @param coeffs A vector containing coefficients for all components.  If component \f$k\f$ is defined by \f$C_k\f$ coefficients,
                   then this vector should have length \f$\sum_{k=1}^K C_i\f$ and the coefficients for component \f$k\f$ should
@@ -57,45 +57,20 @@ public:
     virtual void WrapCoeffs(Kokkos::View<double*, mpart::DeviceSpace> coeffs) override;
     #endif 
     
-    virtual std::shared_ptr<ConditionalMapBase<MemorySpace>> GetComponent(unsigned int i){ return comps_.at(i);}
 
-    /** @brief Computes the log determinant of the Jacobian matrix of this map.
 
-    @details
-    @param pts The points where the jacobian should be evaluated.  Should have \f$N\f$ rows.  Each column contains a single point.
-    @param output A vector with length equal to the number of columns in pts containing the log determinant of the Jacobian.  This
-                  vector should be correctly allocated and sized before calling this function.
-    */
     virtual void LogDeterminantImpl(StridedMatrix<const double, MemorySpace> const& pts,
                                     StridedVector<double, MemorySpace>              output) override;
 
 
-    /** @brief Evaluates the map.
-
-    @details
-    @param pts The points where the map should be evaluated.  Should have \f$N\f$ rows.  Each column contains a single point.
-    @param output A matrix with \f$M\f$ rows to store the map output.  Should have the same number of columns as pts and be
-                  allocated before calling this function.
-    */
-    void EvaluateImpl(StridedMatrix<const double, MemorySpace> const& pts,
+    virtual void EvaluateImpl(StridedMatrix<const double, MemorySpace> const& pts,
                       StridedMatrix<double, MemorySpace>              output) override;
 
     virtual void GradientImpl(StridedMatrix<const double, MemorySpace> const& pts,  
[…]
