+++ src/DebugMap0.cpp
@@ -1,0 +1,257 @@
+#include "MParT/DebugMap.h"
+
+#include "MParT/Utilities/KokkosSpaceMappings.h"
+
+#include <numeric>
+
+using namespace mpart;
+
+template<typename MemorySpace>
+DebugMap<MemorySpace>::DebugMap(std::shared_ptr<ConditionalMapBase<MemorySpace>> const& component) : ConditionalMapBase<MemorySpace>(component->inputDim, component->outputDim, component->numCoeffs),
+                        comp_(component)
+{
+
+
+    std::cout << "constructor: comp_.get() = " << comp_.get() << std::endl; 
+}
+
+template<typename MemorySpace>
+void DebugMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, Kokkos::HostSpace> coeffs)
+{
+    // First, call the ConditionalMapBase version of this function to copy the view into the savedCoeffs member variable
+    ConditionalMapBase<MemorySpace>::SetCoeffs(coeffs);
+
+    comp_->SetCoeffs(coeffs);
+}
+
+template<typename MemorySpace>
+void DebugMap<MemorySpace>::WrapCoeffs(Kokkos::View<double*, Kokkos::HostSpace> coeffs)
+{
+
+}
+
+#if defined(MPART_ENABLE_GPU)
+template<typename MemorySpace>
+void DebugMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, Kokkos::DefaultExecutionSpace::memory_space> coeffs)
+{
+
+}
+
+template<typename MemorySpace>
+void DebugMap<MemorySpace>::WrapCoeffs(Kokkos::View<double*, Kokkos::DefaultExecutionSpace::memory_space> coeffs)
+{
+
+}
+#endif
+
+template<typename MemorySpace>
+void DebugMap<MemorySpace>::LogDeterminantImpl(StridedMatrix<const double, MemorySpace> const& pts,
+                                                    StridedVector<double, MemorySpace>              output)
+{   
+    // // Evaluate the log determinant for the first component
+    // StridedMatrix<const double, MemorySpace> subPts = Kokkos::subview(pts, std::make_pair(0,int(comps_.at(0)->inputDim)), Kokkos::ALL());
+    // comps_.at(0)->LogDeterminantImpl(subPts, output);
+
+    // if(comps_.size()==1)
+    //     return;
+
+    // // Vector to hold log determinant for a single component
+    // Kokkos::View<double*, MemorySpace> compDet("Log Determinant", output.extent(0));
+
+    // Kokkos::RangePolicy<typename MemoryToExecution<MemorySpace>::Space> policy(0,output.size());
+
+
+    // for(unsigned int i=1; i<comps_.size(); ++i){
+    //     subPts = Kokkos::subview(pts, std::make_pair(0,int(comps_.at(i)->inputDim)), Kokkos::ALL());
+    //     comps_.at(i)->LogDeterminantImpl(subPts, compDet);
+
+    //     // Add to the output
+    //     Kokkos::parallel_for(policy, KOKKOS_LAMBDA(const int& j){
+    //         output(j) += compDet(j);
+    //     });
+    // }
+}
+
+template<typename MemorySpace>
+void DebugMap<MemorySpace>::EvaluateImpl(StridedMatrix<const double, MemorySpace> const& pts,
+                                              StridedMatrix<double, MemorySpace>              output)
+{
+    // // Evaluate the output for each component
+    // StridedMatrix<const double, MemorySpace> subPts;
+    // StridedMatrix<double, MemorySpace> subOut;
+
+    // int startOutDim = 0;
+    // for(unsigned int i=0; i<comps_.size(); ++i){
[…]
