### 2022-10-28 a68d02c
DebugMap1 added map functions, still works
