### 2022-10-28 2bb7710
DebugMap2 has both a summary function and comp as inputs
Pointers are still good
