### 2022-10-28 a04a7f9
Replaced SummarizedMap with working version.
