+++ bindings/python/include/CommonPybindUtilities.h
@@ -40,8 +40,8 @@ void ComposedMapWrapper(pybind11::module &m);
 template<typename MemorySpace>
 void SummarizedMapWrapper(pybind11::module &m);
 
-template<typename MemorySpace>
-void DebugMapWrapper(pybind11::module &m);
+// template<typename MemorySpace>
+// void DebugMapWrapper(pybind11::module &m);
 
 template<typename MemorySpace>
 void MapFactoryWrapper(pybind11::module &m);
