### 2022-10-28 f0a4c28
Fix input dimension
