### 2022-10-30 649ed3e
Const to sumFunc_ and comp_
