+++ MParT/MapFactory.h
@@ -122,16 +122,16 @@ namespace mpart{
         // std::shared_ptr<ConditionalMapBase<MemorySpace>> CreateDebugMap(std::shared_ptr<ConditionalMapBase<MemorySpace>> const &comp) { return std::make_shared<DebugMap<MemorySpace>>(comp); }
 
 
-        // /**
-        // @brief Constructs a (generally) non-monotone multivariate expansion.
-        // @param outputDim The output dimension of the expansion.  Each output will be defined by the same multiindex set but will have different coefficients.
-        // @param mset The multiindex set specifying which terms should be used in the multivariate expansion.
-        // @param options Options specifying the 1d basis functions used in the parameterization.
-        // */
-        // template<typename MemorySpace>
-        // std::shared_ptr<ConditionalMapBase<MemorySpace>> CreateSingleEntryMap(unsigned int dim,
-        //                                                                              unsigned int activeInd,
-        //                                                                              std::shared_ptr<ConditionalMapBase<MemorySpace>> &comp);
+        /**
+        @brief Constructs a (generally) non-monotone multivariate expansion.
+        @param outputDim The output dimension of the expansion.  Each output will be defined by the same multiindex set but will have different coefficients.
+        @param mset The multiindex set specifying which terms should be used in the multivariate expansion.
+        @param options Options specifying the 1d basis functions used in the parameterization.
+        */
+        template<typename MemorySpace>
+        std::shared_ptr<ConditionalMapBase<MemorySpace>> CreateSingleEntryMap(unsigned int dim,
+                                                                                     unsigned int activeInd,
+                                                                                     std::shared_ptr<ConditionalMapBase<MemorySpace>> const &comp);
 
         /** This struct is used to map the options to functions that can create a map component with types corresponding 
             to the options.
