+++ bindings/python/tests/test_SummarizedMap.py
@@ -6,7 +6,7 @@ import numpy as np
 
 noneLim = mpart.NoneLim()
 opts = mpart.MapOptions()
-maxDegree = 2
+maxDegree = 1
 dim = 7
 lrcRank = 2
 summary_mat = np.random.randn(lrcRank, dim-1)
@@ -18,30 +18,46 @@ component = mpart.CreateComponent(mset_csemap.fix(True), opts)
 summarized_component = mpart.SummarizedMap(summary_fun, component)
 
 num_samples = 100
-x = np.random.randn(dim, num_samples)
-print(summarized_component.numCoeffs)
-summarized_component.SetCoeffs(np.zeros(summarized_component.numCoeffs))
+pts = np.random.randn(dim, num_samples)
+
+coeffs = np.random.randn(summarized_component.numCoeffs)
+summarized_component.SetCoeffs(coeffs)
 
 def test_numCoeffs():
-    assert summarized_component.numCoeffs == 2 + 3
+    assert summarized_component.numCoeffs == component.numCoeffs
 
 
 def test_CoeffsMap():
-    summarized_component.SetCoeffs(np.zeros(summarized_component.numCoeffs))
-    assert summarized_component.CoeffMap().tolist() == np.zeros(summarized_component.numCoeffs).tolist()
+    
+    coeffs = np.random.randn(summarized_component.numCoeffs)
+    summarized_component.SetCoeffs(coeffs)
+    assert np.all(summarized_component.CoeffMap() == coeffs)
 
 
 def test_Evaluate():
-    assert summarized_component.Evaluate(x).shape == (1,num_samples)
+
+    eval = summarized_component.Evaluate(pts)
+    summary = summary_fun.Evaluate(pts[:dim-1,:])
+    new_pts = np.vstack([summary,pts[dim-1,:]])
+    eval_ = component.Evaluate(new_pts)
+
+    assert eval.shape == (1,num_samples)
+    assert np.allclose(eval_, eval, atol=1E-6)
 
 
 def test_LogDeterminant():
-    assert summarized_component.LogDeterminant(x).shape == (num_samples,)
+
+    logdet = summarized_component.LogDeterminant(pts)
+    summary = summary_fun.Evaluate(pts[:dim-1,:])
+    new_pts = np.vstack([summary,pts[dim-1,:]])
+    logdet_ = component.LogDeterminant(new_pts)
+
+    assert logdet.shape == (num_samples,)
+    assert np.allclose(logdet_, logdet, atol=1E-6)
     
 
 def test_Inverse():
-    coeffs = np.random.randn(summarized_component.numCoeffs)
-    summarized_component.SetCoeffs(coeffs)
-    y = summarized_component.Evaluate(x)
-    x_ = summarized_component.Inverse(np.zeros((0,num_samples)),y)
-    assert np.allclose(x_, x, atol=1E-3)
+
+    y = summarized_component.Evaluate(pts)
+    pts_ = summarized_component.Inverse(pts[:dim-1,:], y)
+    assert np.allclose(pts_, pts[-1,:], atol=1E-3)
