+++ bindings/python/tests/test_ComposedMap.py
@@ -4,13 +4,16 @@ import numpy as np
 print("made map")
 opts = mpart.MapOptions()
 
-num_maps = 8
+num_maps = 4
 dim = 2
 maxOrder = 1
 
 maps = []
+numCoeffs = 0
 for i in range(num_maps):
-    maps.append(mpart.CreateTriangular(dim,dim,maxOrder,opts))
+    map = mpart.CreateTriangular(dim,dim,maxOrder,opts)
+    maps.append(map)
+    numCoeffs += map.numCoeffs
 
 
 composed_map = mpart.ComposedMap(maps, num_maps)
@@ -20,13 +23,14 @@ num_samples = 100
 x = np.random.randn(dim,num_samples)
 
 
-# def test_numCoeffs():
-#     assert triangular.numCoeffs == 2 + 3
+def test_numCoeffs():
+    assert composed_map.numCoeffs == numCoeffs
 
-
-# def test_CoeffsMap():
+def test_CoeffsMap():
     
-#     assert triangular.CoeffMap().tolist() == [0,0,0,0,0]
+    coeffs = np.random.randn(composed_map.numCoeffs)
+    composed_map.SetCoeffs(coeffs)
+    assert np.all(composed_map.CoeffMap() == coeffs)
 
 
 def test_Evaluate():
@@ -37,8 +41,7 @@ def test_LogDeterminant():
     assert composed_map.LogDeterminant(x).shape == (num_samples,)
     
 def test_Inverse():
-    coeffs = np.random.randn(composed_map.numCoeffs)
-    composed_map.SetCoeffs(coeffs)
+
     y = composed_map.Evaluate(x)
     x_ = composed_map.Inverse(np.zeros((0,num_samples)),y)
     assert np.allclose(x_, x, atol=1E-3)
