### 2022-10-31 d70b714
Increase maxDegree
