### 2022-10-31 406a86e
Remove SingleEntryMap class
plan is to use the Factory function
