+++ MParT/TriangularMap.h
@@ -110,16 +110,16 @@ public:
                        StridedMatrix<double, MemorySpace>              output) override;
 
 
-    virtual void LogDeterminantCoeffGradImpl(StridedMatrix<const double, MemorySpace> const& pts, 
+    void LogDeterminantCoeffGradImpl(StridedMatrix<const double, MemorySpace> const& pts,
                                              StridedMatrix<double, MemorySpace>              output) override;
-    
-    virtual void LogDeterminantInputGradImpl(StridedMatrix<const double, MemorySpace> const& pts, 
+
+    void LogDeterminantInputGradImpl(StridedMatrix<const double, MemorySpace> const& pts,
                                              StridedMatrix<double, MemorySpace>              output) override;
 
 
     void print_ptr() { std::cout << "print_ptr(): comps_.at(0).get() = " << comps_.at(0).get() << std::endl; }
 private:
- 
+
     std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> comps_;
 
 
