+++ MParT/SummarizedMap.h
@@ -29,7 +29,7 @@ is positive definite.
  */
 template<typename MemorySpace>
 class SummarizedMap : public ConditionalMapBase<MemorySpace>{
-        
+
 public:
 
     /** @brief Construct a block triangular map from a collection of other ConditionalMapBase objects.
@@ -53,10 +53,10 @@ public:
     virtual void SetCoeffs(Kokkos::View<double*, Kokkos::HostSpace> coeffs) override;
     virtual void WrapCoeffs(Kokkos::View<double*, Kokkos::HostSpace> coeffs) override;
     #if defined(MPART_ENABLE_GPU)
-    virtual void SetCoeffs(Kokkos::View<double*, Kokkos::DefaultExecutionSpace::memory_space> coeffs) override;
-    virtual void WrapCoeffs(Kokkos::View<double*, mpart::DeviceSpace> coeffs) override;
-    #endif 
-    
+    virtual void SetCoeffs(Kokkos::View<double*, DeviceSpace> coeffs) override;
+    virtual void WrapCoeffs(Kokkos::View<double*, DeviceSpace> coeffs) override;
+    #endif
+
     void SummarizePts(StridedMatrix<const double, MemorySpace> const& pts,
                                     StridedMatrix<double, MemorySpace>              output);
 
@@ -68,34 +68,34 @@ public:
     virtual void EvaluateImpl(StridedMatrix<const double, MemorySpace> const& pts,
                       StridedMatrix<double, MemorySpace>              output) override;
 
-    virtual void GradientImpl(StridedMatrix<const double, MemorySpace> const& pts,  
+    virtual void GradientImpl(StridedMatrix<const double, MemorySpace> const& pts,
                               StridedMatrix<const double, MemorySpace> const& sens,
                               StridedMatrix<double, MemorySpace>              output) override;
-    
+
 
     virtual void InverseImpl(StridedMatrix<const double, MemorySpace> const& x1,
                              StridedMatrix<const double, MemorySpace> const& r,
                              StridedMatrix<double, MemorySpace>              output) override;
 
 
-    virtual void CoeffGradImpl(StridedMatrix<const double, MemorySpace> const& pts,  
+    virtual void CoeffGradImpl(StridedMatrix<const double, MemorySpace> const& pts,
                                StridedMatrix<const double, MemorySpace> const& sens,
                                StridedMatrix<double, MemorySpace>              output) override;
 
 
-    virtual void LogDeterminantCoeffGradImpl(StridedMatrix<const double, MemorySpace> const& pts, 
+    virtual void LogDeterminantCoeffGradImpl(StridedMatrix<const double, MemorySpace> const& pts,
                                              StridedMatrix<double, MemorySpace>              output) override;
-    
-    virtual void LogDeterminantInputGradImpl(StridedMatrix<const double, MemorySpace> const& pts, 
+
+    virtual void LogDeterminantInputGradImpl(StridedMatrix<const double, MemorySpace> const& pts,
                                              StridedMatrix<double, MemorySpace>              output) override;
 
 
     void print_ptr() { std::cout << "print_ptr(): comp_.get() = " << comp_.get() << std::endl; std::cout << "print_ptr(): sumFunc_.get() = " << sumFunc_.get() << std::endl; }
 private:
- 
+
     std::shared_ptr<ParameterizedFunctionBase<MemorySpace>> const sumFunc_;
     std::shared_ptr<ConditionalMapBase<MemorySpace>> const comp_;
-    
+
 
 
 }; // class SummarizedMap
