+++ src/TriangularMap.cpp
@@ -32,8 +32,6 @@ TriangularMap<MemorySpace>::TriangularMap(std::vector<std::shared_ptr<Conditiona
             throw std::invalid_argument(msg.str());
         }
     }
-
-    // std::cout << "constructor: comps_.at(0).get() = " << comps_.at(0).get() << std::endl; 
 }
 
 template<typename MemorySpace>
@@ -109,7 +107,7 @@ void TriangularMap<MemorySpace>::WrapCoeffs(Kokkos::View<double*, Kokkos::Defaul
 template<typename MemorySpace>
 void TriangularMap<MemorySpace>::LogDeterminantImpl(StridedMatrix<const double, MemorySpace> const& pts,
                                                     StridedVector<double, MemorySpace>              output)
-{   
+{
     // Evaluate the log determinant for the first component
     StridedMatrix<const double, MemorySpace> subPts = Kokkos::subview(pts, std::make_pair(0,int(comps_.at(0)->inputDim)), Kokkos::ALL());
     comps_.at(0)->LogDeterminantImpl(subPts, output);
@@ -203,8 +201,8 @@ void TriangularMap<MemorySpace>::GradientImpl(StridedMatrix<const double, Memory
     StridedMatrix<const double, MemorySpace> subPts;
     StridedMatrix<const double, MemorySpace> subSens;
 
-    
-    
+
+
     Kokkos::RangePolicy<typename MemoryToExecution<MemorySpace>::Space> policy(0,pts.extent(1));
     unsigned int dim = pts.extent(0);
     Kokkos::parallel_for(policy, KOKKOS_LAMBDA(const int& ptInd){
@@ -212,7 +210,7 @@ void TriangularMap<MemorySpace>::GradientImpl(StridedMatrix<const double, Memory
             output(d,ptInd) = 0.0;
     });
     Kokkos::fence();
-    
+
     int startOutDim = 0;
     for(unsigned int i=0; i<comps_.size(); ++i){
 
@@ -282,7 +280,7 @@ template<typename MemorySpace>
 void TriangularMap<MemorySpace>::LogDeterminantInputGradImpl(StridedMatrix<const double, MemorySpace> const& pts,
                                                              StridedMatrix<double, MemorySpace>              output)
 {
-    // Initialize the output to zero 
+    // Initialize the output to zero
     Kokkos::MDRangePolicy<Kokkos::Rank<2>, typename MemoryToExecution<MemorySpace>::Space> zeroPolicy({0, 0}, {output.extent(0), output.extent(1)});
     Kokkos::parallel_for(zeroPolicy, KOKKOS_LAMBDA(const int& i, const int& j) {
         output(i,j) = 0.0;
@@ -292,10 +290,10 @@ void TriangularMap<MemorySpace>::LogDeterminantInputGradImpl(StridedMatrix<const
     StridedMatrix<const double, MemorySpace> subPts;
     StridedMatrix<double, MemorySpace> subOut;
 
-    int numPts = pts.extent(1);    
+    int numPts = pts.extent(1);
     Kokkos::View<double**,MemorySpace> compGrad("Component Gradient", this->inputDim, numPts);
     Kokkos::View<double**,MemorySpace> subGrad;
-    
+
     for(unsigned int i=0; i<comps_.size(); ++i){
         int compDim = comps_.at(i)->inputDim;
         subPts = Kokkos::subview(pts, std::make_pair(0,compDim), Kokkos::ALL());

[…]
