### 2022-11-04 #4843 @mparno
Thanks @f-t-s !   We fixed the bug you mentioned in [Issue #40](https://github.com/MeasureTransport/MParT-examples/issues/40) in MParT release v1.2 and the fix is currently making its way through conda and the julia package manager.
