+++ src/Utilities/LinearAlgebra.cpp
@@ -66,7 +66,6 @@ void PartialPivLU<Kokkos::HostSpace>::solveInPlace(Kokkos::View<double**,Kokkos:
     luSolver_->matrixLU().template triangularView<Eigen::Upper>().solveInPlace(eigX);
 }
 
-
 template<>
 double PartialPivLU<Kokkos::HostSpace>::determinant() const
 {
@@ -74,7 +73,46 @@ double PartialPivLU<Kokkos::HostSpace>::determinant() const
     return luSolver_->determinant();
 }
 
+template<>
+void Cholesky<Kokkos::HostSpace>::compute(Kokkos::View<const double**,Kokkos::LayoutLeft,Kokkos::HostSpace> A)
+{
+    auto eigA = ConstKokkosToMat(A);
+    // Use eigen to compute the Cholesky decomposition in place
+    cholSolver_ = std::make_shared<Eigen::LLT<Eigen::MatrixXd>>(eigA);
+    isComputed = true;
+}
 
+template<>
+void Cholesky<Kokkos::HostSpace>::solveInPlace(Kokkos::View<double**,Kokkos::LayoutLeft,Kokkos::HostSpace> x)
+{
+    auto eigX = KokkosToMat(x);
+    cholSolver_->matrixL().solveInPlace(eigX);
+    cholSolver_->matrixU().solveInPlace(eigX);
+}
+
+template<>
+void Cholesky<Kokkos::HostSpace>::solveLInPlace(Kokkos::View<double**,Kokkos::LayoutLeft,Kokkos::HostSpace> x)
+{
+    auto eigX = KokkosToMat(x);
+    cholSolver_->matrixL().solveInPlace(eigX);
+}
+
+template<typename MemorySpace>
+Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace> Cholesky<MemorySpace>::solve(StridedMatrix<const double, MemorySpace> x)
+{
+    Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace> output("solution", x.extent(0),x.extent(1));
+    Kokkos::deep_copy(output, x);
+    solveInPlace(output);
+    return output;
+}
+
+template<>
+double Cholesky<Kokkos::HostSpace>::determinant() const
+{
+    assert(isComputed);
+    double sqrt_det = cholSolver_->matrixL().determinant();
+    return sqrt_det*sqrt_det;
+}
 
 #if defined(MPART_ENABLE_GPU)
 
@@ -252,6 +290,155 @@ double PartialPivLU<mpart::DeviceSpace>::determinant() const
     },Kokkos::Prod<double>(det));
 
     return det;
+}
+//
+
+template<>
+void Cholesky<mpart::DeviceSpace>::compute(Kokkos::View<const double**,Kokkos::LayoutLeft,mpart::DeviceSpace> A)
+{
+
+    ldA = A.stride_1();
+
+    int nrows = A.extent(0);
+    int ncols = A.extent(1);
+    assert(nrows==ncols);
+
+    // Resize the space for storing the LU factorization
+    LLT_ = Kokkos::View<double**, Kokkos::LayoutLeft, mpart::DeviceSpace>("LLT", A.extent(0), A.extent(0));
+
+    int ldLLT = LLT_.stride_1();
+    int info;
+
+    // Copy the right block of the matrix A_ into LU_
+    Kokkos::deep_copy(LLT_, A);
+
+    // Set up cuSolver options
+    cusolverDnCreateParams(&params);
+    cusolverDnSetAdvOptions(params, CUSOLVERDN_GETRF, CUSOLVER_ALG_0);
+
+    size_t d_workSize, h_workSize;
+    cusolverDnXpotrf_bufferSize(GetInitializeStatusObject().GetCusolverHandle(),
+                                params,
[…]
