+++ src/Distributions/GaussianDistribution.cpp
@@ -3,17 +3,20 @@
 using namespace mpart;
 
 template<typename MemorySpace>
-GaussianDistribution<MemorySpace>::GaussianDistribution(StridedVector<double, MemorySpace> mean, StridedMatrix<double, MemorySpace> covar): mean_(mean) {
+GaussianDistribution<MemorySpace>::GaussianDistribution(StridedVector<double, MemorySpace> mean, StridedMatrix<double, MemorySpace> covar): mean_(mean), dim_(mean.extent(0)) {
     Factorize(covar);
 }
 
 template<typename MemorySpace>
-GaussianDistribution<MemorySpace>::GaussianDistribution(StridedMatrix<double, MemorySpace> covar) {
+GaussianDistribution<MemorySpace>::GaussianDistribution(StridedMatrix<double, MemorySpace> covar): dim_(covar.extent(0)) {
     Factorize(covar);
 }
 
 template<typename MemorySpace>
-GaussianDistribution<MemorySpace>::GaussianDistribution(StridedVector<double, MemorySpace> mean): mean_(mean), idCov_(true) {}
+GaussianDistribution<MemorySpace>::GaussianDistribution(StridedVector<double, MemorySpace> mean): mean_(mean), dim_(mean.extent(0)), idCov_(true) {}
+
+template<typename MemorySpace>
+GaussianDistribution<MemorySpace>::GaussianDistribution(): idCov_(true) {}
 
 template<typename MemorySpace>
 void GaussianDistribution<MemorySpace>::LogDensityImpl(StridedMatrix<const double, MemorySpace> const &pts, StridedVector<double, MemorySpace> output) {
@@ -21,6 +24,9 @@ void GaussianDistribution<MemorySpace>::LogDensityImpl(StridedMatrix<const doubl
     StridedMatrix<double, MemorySpace> diff;
     int M = pts.extent(0);
     int N = pts.extent(1);
+    if(dim_ != 0 && M != dim_) {
+        throw std::runtime_error("GaussianDistribution::LogDensityImpl: The number of rows in pts must match the dimension of the distribution.");
+    }
     Kokkos::MDRangePolicy<Kokkos::Rank<2>, typename MemoryToExecution<MemorySpace>::Space> policy({{0, 0}}, {{N, M}});
     if(mean_.extent(0) == 0){
         diff = pts;
@@ -33,7 +39,7 @@ void GaussianDistribution<MemorySpace>::LogDensityImpl(StridedMatrix<const doubl
     }
 
     if(!idCov_) {
-        covLU_.solveInPlace(diff);
+        covChol_.solveLInPlace(diff);
     }
 
     Kokkos::parallel_for( "log terms", N, KOKKOS_LAMBDA (const int& j) {
