### 2022-11-09 7dc8618
Work on Gaussian
