+++ tests/Test_LinearAlgebra.cpp
@@ -198,7 +198,6 @@ TEST_CASE( "Testing Cholesky Factorization", "LinearAlgebra_Cholesky" ) {
     B(2,0) = 3.; B(2,1) = 6.;
 
     auto constA = ToConstKokkos(A.data(), 3, 3);
-    auto constB = ToConstKokkos(B.data(), 3, 2);
     auto eigA = ConstKokkosToMat(constA);
     auto eigB = KokkosToMat(B);
     unsigned int nrows = eigA.rows();
@@ -233,7 +232,7 @@ TEST_CASE( "Testing Cholesky Factorization", "LinearAlgebra_Cholesky" ) {
     }
     SECTION("Solve Cholesky out of place") {
         Cholesky<Kokkos::HostSpace> Achol(constA);
-        auto C = Achol.solve(constB);
+        auto C = Achol.solve(B);
         for(unsigned int j=0; j<C.extent(1); ++j){
             for(unsigned int i=0; i<C.extent(0); ++i){
                 CHECK(C(i,j) == Approx(eigX(i,j)).epsilon(1e-14).margin(1e-14));
@@ -344,4 +343,116 @@ TEST_CASE( "Testing Linear Mat-Mat product on Device", "[LinearAlgebra_MatMatDev
     }
 }
 
+TEST_CASE( "Testing LU Factorization on Device", "LinearAlgebra_LUDevice" ) {
+    Kokkos::View<double**, Kokkos::LayoutLeft, Kokkos::HostSpace> A("A", 3, 3);
+    Kokkos::View<double**, Kokkos::LayoutLeft, Kokkos::HostSpace> B("B", 3, 2);
+    A(0,0) = 1.; A(0,1) = 2.; A(0,2) = 3.;
+    A(1,0) = 4.; A(1,1) = 5.; A(1,2) = 6.;
+    A(2,0) = 7.; A(2,1) = 8.; A(2,2) = 0.;
+
+    B(0,0) = 1.; B(0,1) = 4.;
+    B(1,0) = 2.; B(1,1) = 5.;
+    B(2,0) = 3.; B(2,1) = 6.;
+
+    auto constA = ToConstKokkos(A.data(), 3, 3);
+    auto constA_d = ToDevice<mpart::DeviceSpace>(constA);
+    auto B_d = ToDevice<mpart::DeviceSpace>(B);
+    auto eigA = ConstKokkosToMat(constA);
+    auto eigB = KokkosToMat(B);
+    unsigned int nrows = eigA.rows();
+    unsigned int ncols = eigA.cols();
+    Eigen::MatrixXd eigX = eigA.lu().solve(eigB);
+
+    SECTION("Compute LU") {
+        PartialPivLU<mpart::DeviceSpace> Alu_d(constA_d);
+    }
+    SECTION("Solve LU inplace") {
+        Kokkos::View<double**, Kokkos::LayoutLeft, mpart::DeviceSpace> C_d("C", 3, 2);
+        Kokkos::deep_copy(C_d, B_d);
+        PartialPivLU<mpart::DeviceSpace> Alu_d(constA_d);
+        Alu_d.solveInPlace(C_d);
+        auto C_h = ToHost(C_d);
+        for(unsigned int j=0; j<C_h.extent(1); ++j){
+            for(unsigned int i=0; i<C_h.extent(0); ++i){
+                CHECK(C_h(i,j) == Approx(eigX(i,j)).epsilon(1e-14).margin(1e-14));
+            }
+        }
+    }
+    SECTION("Solve LU out of place") {
+        PartialPivLU<mpart::DeviceSpace> Alu_d(constA_d);
+        auto C_d = Alu.solve(B_d);
+        auto C_h = ToHost(C_d);
[…]
