### 2022-11-09 d2f559c
Work on Gaussian
