### 2022-11-09 3249f4e
Work on compiling GaussianDistribution
