+++ src/CMakeLists.txt
@@ -9,9 +9,9 @@ target_sources(mpart
     Utilities/Miscellaneous.cpp
     Utilities/LinearAlgebra.cpp
 
-    # Distributions/DistributionBase.cpp
-    # Distributions/SampleGenerator.cpp
-    # Distributions/GaussianDistribution.cpp
+    Distributions/DensityBase.cpp
+    Distributions/SampleGenerator.cpp
+    Distributions/GaussianDistribution.cpp
 
     ParameterizedFunctionBase.cpp
     ConditionalMapBase.cpp
