### 2022-11-10 d6c5b32
Revising ParameterizedFunctionBase
