+++ MParT/ParameterizedFunctionBase.h
@@ -40,7 +40,11 @@ namespace mpart {
             @detail Performs a deep copy of the input coefficients to the internally stored coefficients.
             @param coeffs A view containing the coefficients to copy.
         */
-        virtual void SetCoeffs(Kokkos::View<double*, Kokkos::HostSpace> coeffs);
+       virtual void SetCoeffs(Kokkos::View<double*, MemorySpace> coeffs);
+
+       #if defined(MPART_ENABLE_GPU)
+       void SetCoeffs(Kokkos::View<double*, std::conditional_t<std::is_same_v<MemorySpace,Kokkos::HostSpace>,mpart::DeviceSpace,Kokkos::HostSpace>> coeffs);
+       #endif
 
         /** @brief Wrap the internal coefficient view around another view.
             @detail Performs a shallow copy of the input coefficients to the internally stored coefficients.
@@ -48,16 +52,11 @@ namespace mpart {
             internally stored view.
             @param coeffs A view containing the coefficients we want to wrap.
         */
-        virtual void WrapCoeffs(Kokkos::View<double*, Kokkos::HostSpace> coeffs);
-
-        #if defined(MPART_ENABLE_GPU)
-        virtual void SetCoeffs(Kokkos::View<double*, mpart::DeviceSpace> coeffs);
-        virtual void WrapCoeffs(Kokkos::View<double*, mpart::DeviceSpace> coeffs);
-        #endif 
+        virtual void WrapCoeffs(Kokkos::View<double*, MemorySpace> coeffs);
 
         /** SetCoeffs function with conversion from Eigen to Kokkos vector types.*/
         virtual void SetCoeffs(Eigen::Ref<Eigen::VectorXd> coeffs);
-        virtual void WrapCoeffs(Eigen::Ref<Eigen::VectorXd> coeffs);
+        void WrapCoeffs(Eigen::Ref<Eigen::VectorXd> coeffs);
 
         /** Returns an eigen map wrapping around the coefficient vector, which is stored in a Kokkos::View.  Updating the
             components of this map should also update the view.
