+++ src/ComposedMap.cpp
@@ -155,9 +155,10 @@ ComposedMap<MemorySpace>::ComposedMap(std::vector<std::shared_ptr<ConditionalMap
     }
 }
 
-template<typename MemorySpace>
-void ComposedMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, Kokkos::HostSpace> coeffs)
+template<typename MemorySpace, typename MemorySpace2>
+void ComposedMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, MemorySpace2> coeffD)
 {
+    Kokkos::View<double*, MemorySpace> coeffs = ToHost(coeffD);
     // First, call the ConditionalMapBase version of this function to copy the view into the savedCoeffs member variable
     ConditionalMapBase<MemorySpace>::SetCoeffs(coeffs);
 
@@ -171,25 +172,6 @@ void ComposedMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, Kokkos::HostSpace
         cumNumCoeffs += maps_.at(i)->numCoeffs;
     }
 }
-
-#if defined(MPART_ENABLE_GPU)
-template<typename MemorySpace>
-void ComposedMap<MemorySpace>::SetCoeffs(Kokkos::View<double*, Kokkos::DefaultExecutionSpace::memory_space> coeffs)
-{
-    // First, call the ConditionalMapBase version of this function to copy the view into the savedCoeffs member variable
-    ConditionalMapBase<MemorySpace>::SetCoeffs(coeffs);
-
-    // Now create subviews for each of the maps
-    unsigned int cumNumCoeffs = 0;
-    for(unsigned int i=0; i<maps_.size(); ++i){
-        assert(cumNumCoeffs+maps_.at(i)->numCoeffs <= this->savedCoeffs.size());
-
-        maps_.at(i)->savedCoeffs = Kokkos::subview(this->savedCoeffs,
-            std::make_pair(cumNumCoeffs, cumNumCoeffs+maps_.at(i)->numCoeffs));
-        cumNumCoeffs += maps_.at(i)->numCoeffs;
-    }
-}
-#endif
 
 template<typename MemorySpace>
 void ComposedMap<MemorySpace>::LogDeterminantImpl(StridedMatrix<const double, MemorySpace> const& pts,
