### 2022-11-10 a9ad967
Work on CUDA compatibility
