+++ src/CMakeLists.txt
@@ -9,10 +9,6 @@ target_sources(mpart
     Utilities/Miscellaneous.cpp
     Utilities/LinearAlgebra.cpp
 
-    Distributions/DensityBase.cpp
-    Distributions/SampleGenerator.cpp
-    Distributions/GaussianDistribution.cpp
-
     ParameterizedFunctionBase.cpp
     ConditionalMapBase.cpp
     TriangularMap.cpp
