### 2022-11-11 b1978a0
Revert "Starting on Cereal"
This reverts commit ec046023c08ca319ab8d3ff5a3030b2a040f8753.
