### 2022-11-11 3ba9f98
Removing print_ptr from bindings, removing CreateAffineLRCMap method
