+++ bindings/python/src/TriangularMap.cpp
@@ -20,7 +20,6 @@ void mpart::binding::TriangularMapWrapper(py::module &m)
     py::class_<TriangularMap<MemorySpace>, ConditionalMapBase<MemorySpace>, std::shared_ptr<TriangularMap<MemorySpace>>>(m, tName.c_str())
         .def(py::init<std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>>>())
         .def("GetComponent", &TriangularMap<MemorySpace>::GetComponent)
-        .def("print_ptr", &TriangularMap<MemorySpace>::print_ptr)
         ;
 
 }
