### 2022-11-12 adb5831
Removed some unnecessary deep_copies from SummaryMap Eval and Gradient
