### 2022-11-12 76f386d
Adding SummarizedMap Eval and LogDet tests
