### 2022-11-14 3922728
Debugging AffineFunction values changing
