+++ MParT/AffineFunction.h
@@ -45,6 +45,11 @@ public:
                       StridedMatrix<const double, MemorySpace> const& sens,
                       StridedMatrix<double, MemorySpace>              output) override;
 
+
+    void print_Aij(int i, int j) { std::cout << "A(" << i << "," << j << ") = " << A_(i,j) << std::endl; }
+    void print_ptr_A() { std::cout << "ptr A = " << A_.data() << std::endl; }
+
+
 protected:
 
     StridedMatrix<double,MemorySpace> A_;
