### 2022-11-15 65325d9
Deep_copying in AffineMap and AffineFunction constructors
