+++ src/AffineFunction.cpp
@@ -11,12 +11,15 @@ using namespace mpart;
 
 template<typename MemorySpace>
 AffineFunction<MemorySpace>::AffineFunction(StridedVector<double,MemorySpace> b) : ParameterizedFunctionBase<MemorySpace>(b.size(),b.size(),0), 
-                                                                  b_(b) 
-{}
+                                                                  b_("b",b.layout()) 
+{
+
+    Kokkos::deep_copy(b_, b);
+}
 
 template<typename MemorySpace>
 AffineFunction<MemorySpace>::AffineFunction(StridedMatrix<double,MemorySpace> A) : ParameterizedFunctionBase<MemorySpace>(A.extent(1),A.extent(0),0),
-                                                                                   A_(A)
+                                                                               A_("A", A.layout())
 {
     Kokkos::deep_copy(A_, A);
     assert(A_.extent(0)<=A_.extent(1));
@@ -26,10 +29,12 @@ AffineFunction<MemorySpace>::AffineFunction(StridedMatrix<double,MemorySpace> A)
 template<typename MemorySpace>
 AffineFunction<MemorySpace>::AffineFunction(StridedMatrix<double,MemorySpace> A,
                                   StridedVector<double,MemorySpace> b) : ParameterizedFunctionBase<MemorySpace>(A.extent(1),A.extent(0),0),
-                                                                  A_(A),
-                                                                  b_(b)
+                                                                  A_("A", A.layout()),
+                                                                  b_("b",b.layout())
 {   
-    assert(A_.extent(0)<=A_.extent(1));
+    Kokkos::deep_copy(A_, A);
+    Kokkos::deep_copy(b_, b);
+    assert(A_.extent(0) <= A_.extent(1));
     assert(A_.extent(0) == b_.extent(0));
 }
 
