### 2022-11-15 5c85f2d
Removing debugging functions
