### 2022-11-15 9f90bb2
Added commented out debugging functions
