### 2022-11-15 17b427b
Work on Serialization
