### 2022-11-15 c872683
Change matlab initialization
