+++ .github/workflows/build-bindings.yml
@@ -19,8 +19,8 @@ jobs:
 
       - name: TEMPORARY
         run: |
-        export JULIA_VERSION=$(julia -e "println(Sys.BINDIR)" | awk '{print $NF}')
-        ls /usr/local/julia$JULIA_VERSION
+          export JULIA_VERSION=$(julia -v | awk '{print $NF}')
+          ls /usr/local/julia$JULIA_VERSION
 
       - name: Setup Julia
         run: |
