+++ .github/workflows/build-bindings.yml
@@ -59,7 +59,7 @@ jobs:
 
       - name: Install Eigen, Catch2, Pybind11
         shell: bash -l {0}
-        run: conda install -c conda-forge -y eigen catch2 pybind11 pytest
+        run: conda install -c conda-forge -y eigen catch2 pybind11 pytest numpy
 
       - name: Configure MParT
         shell: bash -l {0}
