+++ .github/workflows/build-bindings.yml
@@ -41,13 +41,16 @@ jobs:
           export GITHUB_JULIA_PATH=$(julia -e "println(DEPOT_PATH[1])")
           cat $GITHUB_JULIA_PATH/artifacts/Overrides.toml
 
+      - name: Set up MATLAB
+        uses: matlab-actions/setup-matlab@v1
+
       - name: Checkout MParT
         uses: actions/checkout@v3
         with:
           path: mpart
 
       - name: Cache Kokkos
-        uses: actions/cache@v1.0.3
+        uses: actions/cache@v3
         id: cache-kokkos
         with:
           path: "${{ github.workspace }}/KOKKOS_INSTALL"
