+++ .github/workflows/build-external-lib-tests.yml
@@ -22,21 +22,30 @@ jobs:
         with:
           path: mpart
 
-      - name: Checkout Kokkos
+      - name: Cache Kokkos
+        uses: actions/cache@v3
+        id: cache-kokkos
+        with:
+          path: "${{ github.workspace }}/KOKKOS_INSTALL"
+          key: kokkos3.7.00
+
+      - if: ${{steps.cache-kokkos.outputs.cache-hit != 'true'}}
+        name: Checkout Kokkos
         uses: actions/checkout@v3
         with:
           repository: kokkos/kokkos
           path: kokkos
           ref: '3.7.00'
 
-      - name: Install Kokkos
+      - if: ${{steps.cache-kokkos.outputs.cache-hit != 'true'}}
+        name: Install Kokkos
         run: |
           cd $GITHUB_WORKSPACE/kokkos
           mkdir build && cd build
           cmake -DKokkos_ENABLE_THREADS=ON \
                 -DBUILD_SHARED_LIBS=ON     \
                 -DKokkos_ENABLE_SERIAL=ON  \
-                -DKokkos_CXX_STANDARD=17   \
+                -DCMAKE_CXX_STANDARD=17    \
                 -DCMAKE_INSTALL_PREFIX=$GITHUB_WORKSPACE/KOKKOS_INSTALL/ \
                 ../
           sudo make -j2 install
@@ -61,7 +70,7 @@ jobs:
         run: cd $GITHUB_WORKSPACE/mpart/build; ./RunTests --kokkos-threads=2 --reporter junit -o test-results-external.xml
 
       - name: Publish Unit Test Results
-        uses: EnricoMi/publish-unit-test-result-action@v1
+        uses: EnricoMi/publish-unit-test-result-action@v2
         if: always()
         with:
           check_name: "Test Results with Externally Built Libraries"
