### 2022-11-16 af117a7
conda cache in external lib tests, fix sed
