+++ .github/workflows/build-external-lib-tests.yml
@@ -17,6 +17,14 @@ jobs:
           python-version: "3.8"
           activate-environment: "test"
 
+      - name: Initialize Conda Variables
+        shell: bash -l {0}
+        run: |
+          export CONDA_EXE=$(which conda)
+          export CONDA_ROOT=${CONDA_EXE%bin*}
+          echo "CONDA_ROOT=$CONDA_ROOT" >> $GITHUB_ENV
+          echo "DATE=$(date +'%Y%m%d')" >> $GITHUB_ENV
+
       - name: Checkout MParT
         uses: actions/checkout@v3
         with:
@@ -50,9 +58,17 @@ jobs:
                 ../
           sudo make -j2 install
 
-      - name: Install Eigen, Catch2, Pybind11
+      - name: Cache Conda Deps
+        uses: actions/cache@v3
+        id: cache-conda
+        with:
+          path: "${{ env.CONDA_ROOT }}/pkgs"
+          key: conda-${{ hashFiles('**/environment.yml') }}-${{ env.DATE }}
+
+      - if: ${{steps.cache-conda.outputs.cache-hit != 'true'}}
+        name: Install Conda Dependencies
         shell: bash -l {0}
-        run: conda install -c conda-forge -y eigen catch2 pybind11
+        run: conda env update -n test -f $GITHUB_WORKSPACE/mpart/.github/environment.yml
 
       - name: Configure MParT
         shell: bash -l {0}
