### 2022-11-16 63189f2
Remove Kokkos_Layout from Julia, remove file extension
