+++ .github/workflows/build-bindings.yml
@@ -129,7 +129,7 @@ jobs:
         continue-on-error: true
         uses: matlab-actions/run-command@v1
         with:
-          command: $GITHUB_WORKSPACE/mpart/bindings/matlab/tests/runtests.m
+          command: $GITHUB_WORKSPACE/mpart/bindings/matlab/tests/runtests
 
       - name: Publish Unit Test Results
         uses: EnricoMi/publish-unit-test-result-action@v2
