### 2022-11-16 69179fa
Added test for SingleEntryMap
