+++ tests/Test_TriangularMap.cpp
@@ -75,7 +75,7 @@ TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_
             REQUIRE(outBlock.extent(1)==numSamps);
             REQUIRE(outBlock.extent(0)==1);
             for(unsigned int j=0; j<numSamps; ++j)
-                CHECK( out(i,j) == Approx(outBlock(0,j)).epsilon(1e-6));
+                CHECK( out(i,j) == Approx(outBlock(0,j)).margin(1e-6));
         }
     }
 
@@ -86,7 +86,7 @@ TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_
 
         for(unsigned int i=0; i<numBlocks; ++i){
             for(unsigned int j=0; j<numSamps; ++j)
-                CHECK( inv(i,j) == Approx(in(i+extraInputs,j)).epsilon(1e-6));
+                CHECK( inv(i,j) == Approx(in(i+extraInputs,j)).margin(1e-6));
         }
     }
 
@@ -104,7 +104,7 @@ TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_
         }
 
         for(unsigned int j=0; j<numSamps; ++j)
-            CHECK(logDet(j) == Approx(truth(j)).epsilon(1e-10));
+            CHECK(logDet(j) == Approx(truth(j)).margin(1e-10));
 
     }
 
@@ -139,7 +139,7 @@ TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_
                 for(unsigned int j=0; j<triMap->outputDim; ++j)
                     fdDeriv += sens(j,ptInd) * (evals2(j,ptInd)-evals(j,ptInd))/fdstep;
 
-                CHECK( coeffGrad(i,ptInd) == Approx(fdDeriv).epsilon(1e-3)); 
+                CHECK( coeffGrad(i,ptInd) == Approx(fdDeriv).margin(1e-3)); 
             }
             coeffs(i) -= fdstep;
         }
@@ -178,7 +178,7 @@ TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_
                 for(unsigned int j=0; j<triMap->outputDim; ++j)
                     fdDeriv += sens(j,ptInd) * (evals2(j,ptInd)-evals(j,ptInd))/fdstep;
 
-                CHECK( inputGrad(i,ptInd) == Approx(fdDeriv).epsilon(1e-3)); 
+                CHECK( inputGrad(i,ptInd) == Approx(fdDeriv).margin(1e-3)); 
             }
 
             for(unsigned int ptInd=0; ptInd<numSamps; ++ptInd)
@@ -212,7 +212,7 @@ TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_
             logDet2 = triMap->LogDeterminant(in);
 
             for(unsigned int ptInd=0; ptInd<numSamps; ++ptInd)
-                CHECK( detGrad(i,ptInd) == Approx((logDet2(ptInd)-logDet(ptInd))/fdstep).epsilon(1e-4)); 
+                CHECK( detGrad(i,ptInd) == Approx((logDet2(ptInd)-logDet(ptInd))/fdstep).margin(1e-4)); 
             
             coeffs(i) -= fdstep;
         }
@@ -440,6 +440,221 @@ TEST_CASE( "Testing TriangularMap made from smaller TriangularMaps", "[Triangula
 
             for(unsigned int ptInd=0; ptInd<numSamps; ++ptInd)
                 CHECK( detGrad(i,ptInd) == Approx((logDet2(ptInd)-logDet(ptInd))/fdstep).epsilon(1e-4)); 
+            
[…]
