+++ bindings/matlab/tests/runtests.m
@@ -2,6 +2,8 @@ import matlab.unittest.TestRunner
 import matlab.unittest.TestSuite
 import matlab.unittest.plugins.XMLPlugin
 
+KokkosInitialize(2);
+
 suite = TestSuite.fromClass(?FixedMultiIndexSetTest);
 runner = TestRunner.withNoPlugins;
 xmlFile = 'test-results-matlab.xml';
