### 2022-11-17 584557a
Fix Cache in External, remove TEMPORARY in binding
