### 2022-11-18 16040ef
Fix FixedMultiIndexSet when no MPART_ARCHIVE
