### 2022-11-18 d371700
Add cereal dependency
