### 2022-11-18 e2e9521
Fix cereal auto install
