+++ .github/workflows/build-bindings.yml
@@ -20,7 +20,7 @@ jobs:
         id: cache-conda
         with:
           path: "~/conda_pkgs_dir"
-          key: conda-${{ hashFiles('**/environment.yml') }}-${{ env.DATE }}
+          key: conda-${{ hashFiles('.github/environment.yml') }}-${{ env.DATE }}
 
       - name: Use Conda
         uses: conda-incubator/setup-miniconda@v2
