+++ .github/workflows/build-bindings.yml
@@ -15,6 +15,11 @@ jobs:
         shell: bash -l {0}
         run: echo "DATE=$(date +'%Y%m%d')" >> $GITHUB_ENV
 
+      - name: Checkout MParT
+        uses: actions/checkout@v3
+        with:
+          path: mpart
+
       - name: Cache Conda Deps
         uses: actions/cache@v3
         id: cache-conda
@@ -39,11 +44,6 @@ jobs:
 
       - name: Set up MATLAB
         uses: matlab-actions/setup-matlab@v1
-
-      - name: Checkout MParT
-        uses: actions/checkout@v3
-        with:
-          path: mpart
 
       - name: Cache Kokkos
         uses: actions/cache@v3
