### 2022-11-21 86c6f2b
Turns out I don't need Findcereal.
