### 2022-11-21 797afba
Fix Cereal cmake, Fix pytests, only cereal on host
