+++ bindings/python/tests/test_Serialization.py
@@ -3,6 +3,7 @@
 import mpart
 import numpy as np
 import os
+import pytest
 
 opts = mpart.MapOptions()
 
@@ -16,24 +17,37 @@ map_1 = mpart.CreateComponent(mset_1, opts)
 map_2 = mpart.CreateComponent(mset_2, opts)
 
 triangular = mpart.TriangularMap([map_1,map_2])
+triangular.SetCoeffs(np.concatenate([map_1.CoeffMap(), map_2.CoeffMap()]))
+
+@pytest.fixture(autouse=True)
+def run_around_tests():
+    os.mkdir("_test_serialization")
+    yield
+    os.remove("_test_serialization/opts.mpart")
+    os.remove("_test_serialization/mset_1.mpart")
+    os.remove("_test_serialization/mset_2.mpart")
+    os.remove("_test_serialization/map_1.mpart")
+    os.remove("_test_serialization/map_2.mpart")
+    os.remove("_test_serialization/triangular.mpart")
+    os.rmdir("_test_serialization")
+
 
 def test_serialization():
-    os.mkdir("_test_serialization")
     mpart.SerializeMapOptions(opts, "_test_serialization/opts.mpart")
     mpart.SerializeFixedMultiIndexSet(mset_1, "_test_serialization/mset_1.mpart")
     mpart.SerializeFixedMultiIndexSet(mset_2, "_test_serialization/mset_2.mpart")
-    mpart.SerializeComponent(map_1, "_test_serialization/map_1.mpart")
-    mpart.SerializeComponent(map_2, "_test_serialization/map_2.mpart")
+    mpart.SerializeMapCoeffs(map_1, "_test_serialization/map_1.mpart")
+    mpart.SerializeMapCoeffs(map_2, "_test_serialization/map_2.mpart")
     mpart.SerializeMapCoeffs(triangular, "_test_serialization/triangular.mpart")
 
     opts_s = mpart.DeserializeMapOptions("_test_serialization/opts.mpart")
     mset_1_s = mpart.DeserializeFixedMultiIndexSet("_test_serialization/mset_1.mpart")
     mset_2_s = mpart.DeserializeFixedMultiIndexSet("_test_serialization/mset_2.mpart")
-    map_1_s = mpart.DeserializeComponent("_test_serialization/map_1.mpart")
-    map_2_s = mpart.DeserializeComponent("_test_serialization/map_2.mpart")
+    map_1_s = mpart.DeserializeMapCoeffs("_test_serialization/map_1.mpart")
+    map_2_s = mpart.DeserializeMapCoeffs("_test_serialization/map_2.mpart")
     triangularCoeff_s = mpart.DeserializeMapCoeffs("_test_serialization/triangular.mpart")
-    assert mset_1_s.maxDegrees() == mset_1.maxDegrees()
-    assert mset_2_s.maxDegrees() == mset_2.maxDegrees()
+    assert (mset_1_s.MaxDegrees() == mset_1.MaxDegrees()).all()
+    assert (mset_2_s.MaxDegrees() == mset_2.MaxDegrees()).all()
     assert opts_s.basisType == opts.basisType
     assert opts_s.basisLB == opts.basisLB
     assert opts_s.basisUB == opts.basisUB
@@ -46,13 +60,6 @@ def test_serialization():
     assert opts_s.quadMinSub == opts.quadMinSub
     assert opts_s.quadPts == opts.quadPts
     assert opts_s.contDeriv == opts.contDeriv
-    assert map_1_s.CoeffMap() == map_1.CoeffMap()
-    assert map_2_s.CoeffMap() == map_2.CoeffMap()
-    assert triangularCoeff_s == triangular.CoeffMap()
[…]
