+++ bindings/python/CMakeLists.txt
@@ -26,7 +26,7 @@ endif()
 include_directories(./include ../common/include)
 pybind11_add_module(pympart SHARED NO_EXTRAS ${PYTHON_BINDING_SOURCES})
 
-target_link_libraries(pympart PRIVATE mpart Kokkos::kokkos Eigen3::Eigen)
+target_link_libraries(pympart PRIVATE mpart Kokkos::kokkos Eigen3::Eigen ${CEREAL_LIBRARIES})
 
 # Add an installation target for the python bindings
 install(TARGETS pympart DESTINATION "${PYTHON_INSTALL_PREFIX}/mpart")

[…]
