+++ tests/Test_Serialization.cpp
@@ -1,13 +1,81 @@
 #include <sstream>
 #include <catch2/catch_all.hpp>
 
+#include "MParT/MonotoneComponent.h"
+#include "MParT/PositiveBijectors.h"
+#include "MParT/OrthogonalPolynomial.h"
+#include "MParT/MultivariateExpansionWorker.h"
+#include "MParT/Quadrature.h"
 #include "MParT/TriangularMap.h"
 #include "MParT/MapFactory.h"
+#include "MParT/Utilities/Serialization.h"
 
 using namespace mpart;
 using namespace Catch;
+using DefaultMonotoneComponent = MonotoneComponent<MultivariateExpansionWorker<ProbabilistHermite,Kokkos::HostSpace>, SoftPlus, AdaptiveSimpson<Kokkos::HostSpace>, Kokkos::HostSpace>;
 
-TEST_CASE( "Test serializing 3d triangular map from MonotoneComponents", "[Serialization_TriangularMap]" ) {
+TEST_CASE( "Test serializing Kokkos Views", "[Serialization]" ) {
+    std::stringstream ss;
+    SECTION("Check Serialization vectors") {
+        Kokkos::View<double*, Kokkos::HostSpace> vec("vec", 10);
+        Kokkos::View<double*, Kokkos::HostSpace> vec2("vec2", 10);
+        Kokkos::deep_copy(vec, 1.0);
+        Kokkos::deep_copy(vec2, 2.0);
+        {
+            cereal::BinaryOutputArchive archive(ss);
+            archive(vec);
+        }
+        {
+            cereal::BinaryInputArchive archive(ss);
+            archive(vec2);
+        }
+        for (int i = 0; i < 10; i++) {
+            REQUIRE(vec2(i) == 1.0);
+        }
+    }
+
+    SECTION("Check Serialization Matrices: LayoutLeft") {
+        Kokkos::View<double**, Kokkos::LayoutLeft, Kokkos::HostSpace> mat("mat", 10, 10);
+        Kokkos::View<double**, Kokkos::LayoutLeft, Kokkos::HostSpace> mat2("mat2", 10, 10);
+        Kokkos::deep_copy(mat, 1.0);
+        Kokkos::deep_copy(mat2, 2.0);
+        {
+            cereal::BinaryOutputArchive archive(ss);
+            archive(mat);
+        }
+        {
+            cereal::BinaryInputArchive archive(ss);
+            archive(mat2);
+        }
+        for (int i = 0; i < 10; i++) {
+            for (int j = 0; j < 10; j++) {
+                REQUIRE(mat2(i, j) == 1.0);
+            }
+        }
+    }
+
+    SECTION("Check Serialization Matrices: LayoutRight") {
+        Kokkos::View<double**, Kokkos::LayoutRight, Kokkos::HostSpace> mat("mat", 10, 10);
+        Kokkos::View<double**, Kokkos::LayoutRight, Kokkos::HostSpace> mat2("mat2", 10, 10);
+        Kokkos::deep_copy(mat, 1.0);
+        Kokkos::deep_copy(mat2, 2.0);
+        {
+            cereal::BinaryOutputArchive archive(ss);
+            archive(mat);
+        }
+        {
+            cereal::BinaryInputArchive archive(ss);
+            archive(mat2);
+        }
+        for (int i = 0; i < 10; i++) {
+            for (int j = 0; j < 10; j++) {
+                REQUIRE(mat2(i, j) == 1.0);
+            }
+        }
+    }
+}
+
+TEST_CASE( "Test serializing 3d triangular map from MonotoneComponents", "[Serialization]" ) {
 
[…]
