+++ MParT/Utilities/Serialization.h
@@ -7,46 +7,51 @@
 
 #include "MParT/Utilities/ArrayConversions.h"
 
-namespace cereal {
-    template <typename ScalarType, typename MemorySpace, typename Archive>
-    std::enable_if_t<traits::is_output_serializable<BinaryData<ScalarType>, Archive>::value && std::is_same_v<MemorySpace, Kokkos::HostSpace>, void> save(
-        Archive &ar, Kokkos::View<ScalarType*, MemorySpace> const &vec) {
+template<typename What, typename ... Args>
+struct is_present {
+    static constexpr bool value {(std::is_same_v<What, Args> || ...)};
+};
 
-        Kokkos::View<ScalarType*,Kokkos::HostSpace> vec_h = vec;
+namespace cereal {
+    template <typename ScalarType, typename Archive, typename... Traits>
+    std::enable_if_t<traits::is_output_serializable<BinaryData<ScalarType>, Archive>::value && (std::is_same_v<Traits, Kokkos::HostSpace> || ...), void> save(
+        Archive &ar, Kokkos::View<ScalarType*, Traits...> const &vec) {
+        Kokkos::View<ScalarType*,Traits...> vec_h = vec;
+        static_assert(!(std::is_same_v<Traits, Kokkos::LayoutStride> || ...), "LayoutStride not supported");
         unsigned int sz = vec_h.extent(0);
         ar(sz);
         ar(binary_data(vec_h.data(), sz * sizeof(std::remove_cv_t<ScalarType>)));
     }
 
-    template<typename ScalarType, typename MemorySpace, typename Archive>
-    std::enable_if_t<traits::is_input_serializable<BinaryData<ScalarType>, Archive>::value && std::is_same_v<MemorySpace, Kokkos::HostSpace>, void> load(
-        Archive &ar, Kokkos::View<ScalarType*, MemorySpace> &vec) {
-
+    template<typename ScalarType, typename Archive, typename... Traits>
+    std::enable_if_t<traits::is_input_serializable<BinaryData<ScalarType>, Archive>::value && (std::is_same_v<Traits, Kokkos::HostSpace> || ...), void> load(
+        Archive &ar, Kokkos::View<ScalarType*, Traits...> &vec) {
+        static_assert(!(std::is_same_v<Traits, Kokkos::LayoutStride> || ...), "LayoutStride not supported");
         unsigned int sz;
         ar(sz);
-        Kokkos::View<ScalarType*,Kokkos::HostSpace> vec_h ("vec_h", sz);
+        Kokkos::View<ScalarType*,Traits...> vec_h ("vec_h", sz);
         ar(binary_data(vec_h.data(), sz * sizeof(ScalarType)));
         vec = std::move(vec_h);
     }
 
-    template <typename ScalarType, typename MemorySpace, typename Archive>
-    std::enable_if_t<traits::is_output_serializable<BinaryData<ScalarType>, Archive>::value && std::is_same_v<MemorySpace, Kokkos::HostSpace>, void> save(
-        Archive &ar, Kokkos::View<ScalarType**, MemorySpace> const &mat) {
-
+    template<typename ScalarType, typename Archive, typename... Traits>
+    std::enable_if_t<traits::is_output_serializable<BinaryData<ScalarType>, Archive>::value && (std::is_same_v<Traits, Kokkos::HostSpace> || ...), void> save(
+        Archive &ar, Kokkos::View<ScalarType**, Traits...> const &mat) {
+        static_assert(!(std::is_same_v<Traits, Kokkos::LayoutStride> || ...), "LayoutStride not supported");
         unsigned int m = mat.extent(0);
[…]
