### 2022-11-23 f0681ba
Debugging serial
