### 2022-11-23 dbf74fc
Fix scoping issues in unit tests
