+++ tests/Test_Serialization.cpp
@@ -105,44 +105,43 @@ TEST_CASE( "Test serializing 3d triangular map from MonotoneComponents", "[Seria
 
     std::stringstream ss;
 
-    SECTION("Check Serialization"){
-        cereal::BinaryOutputArchive oarchive(ss);
-        oarchive(options1);
-        oarchive(triMap1->inputDim, triMap1->outputDim, maxDegree);
-        save(oarchive, triMap1->Coeffs());
-        for(int i=0; i<numBlocks; ++i) {
-            oarchive(*msets1[i]);
+    SECTION("Check TriangularMap Serialization"){
+        {
+            cereal::BinaryOutputArchive oarchive(ss);
+            oarchive(options1);
+            oarchive(triMap1->inputDim, triMap1->outputDim, maxDegree);
+            save(oarchive, triMap1->Coeffs());
+            for(int i=0; i<numBlocks; ++i) {
+                oarchive(*msets1[i]);
+            }
         }
-    }
-    SECTION("Check Deserialization"){
-        cereal::BinaryInputArchive iarchive(ss);
-        MapOptions options2;
-        iarchive(options2);
-        // std::cerr << "Line 121\n";
-        // REQUIRE(options1.basisType == options2.basisType);
-        // REQUIRE(options1.basisNorm == options2.basisNorm);
-        // unsigned int inputDim2, outputDim2, maxDegree2;
-        // iarchive(inputDim2, outputDim2, maxDegree2);
-        // REQUIRE(inputDim2 == triMap1->inputDim);
-        // REQUIRE(outputDim2 == triMap1->outputDim);
-        // REQUIRE(maxDegree2 == maxDegree);
-        // Kokkos::View<double*,Kokkos::HostSpace> coeffs2;
-        // load(iarchive, coeffs2);
-        // REQUIRE(coeffs1.extent(0) == coeffs2.extent(0));
-        // std::cerr << "Line 131\n";
-        // for(unsigned int i=0; i<coeffs1.extent(0); ++i)
-        //     REQUIRE(coeffs1(i) == coeffs2(i));
-        // std::cerr << "Line 135\n";
-        // for(int i=0; i<numBlocks; ++i) {
-        //     auto mset1 = *msets1[i];
-        //     FixedMultiIndexSet<Kokkos::HostSpace> mset2 (1,1);
-        //     iarchive(mset2);
-        //     REQUIRE(mset2.Size() == mset1.Size());
-        //     auto maxDegrees1 = mset1.MaxDegrees();
-        //     auto maxDegrees2 = mset2.MaxDegrees();
-        //     REQUIRE(maxDegrees1.extent(0) == maxDegrees2.extent(0));
-        //     for(unsigned int j=0; j<maxDegrees1.extent(0); ++j)
-        //         REQUIRE(maxDegrees1(j) == maxDegrees2(j));
-        // }
+        {
+            cereal::BinaryInputArchive iarchive(ss);
+            MapOptions options2;
+            iarchive(options2);
+            REQUIRE(options1.basisType == options2.basisType);
+            REQUIRE(options1.basisNorm == options2.basisNorm);
+            unsigned int inputDim2, outputDim2, maxDegree2;
+            iarchive(inputDim2, outputDim2, maxDegree2);
+            REQUIRE(inputDim2 == triMap1->inputDim);
+            REQUIRE(outputDim2 == triMap1->outputDim);
+            REQUIRE(maxDegree2 == maxDegree);
+            Kokkos::View<double*,Kokkos::HostSpace> coeffs2;
+            load(iarchive, coeffs2);
+            REQUIRE(coeffs1.extent(0) == coeffs2.extent(0));
[…]
