### 2022-11-25 ed3dc4f
Work on documentation, move python to OOP
