+++ bindings/python/tests/test_Serialization.py
@@ -1,22 +1,22 @@
 # Serialization test
 
-import mpart
+import mpart as mt
 import numpy as np
 import os
 import pytest
 
-opts = mpart.MapOptions()
+opts = mt.MapOptions()
 
 multis_1 = np.array([[0],[1]])  # linear
 multis_2 = np.array([[0,0],[0,1],[2,0]])  # quadratic in x_1, linear in x_2, matches form of target
 
-mset_1 = mpart.MultiIndexSet(multis_1).fix(True)
-mset_2 = mpart.MultiIndexSet(multis_2).fix(True)
+mset_1 = mt.MultiIndexSet(multis_1).fix(True)
+mset_2 = mt.MultiIndexSet(multis_2).fix(True)
 
-map_1 = mpart.CreateComponent(mset_1, opts)
-map_2 = mpart.CreateComponent(mset_2, opts)
+map_1 = mt.CreateComponent(mset_1, opts)
+map_2 = mt.CreateComponent(mset_2, opts)
 
-triangular = mpart.TriangularMap([map_1,map_2])
+triangular = mt.TriangularMap([map_1,map_2])
 triangular.SetCoeffs(np.concatenate([map_1.CoeffMap(), map_2.CoeffMap()]))
 
 @pytest.fixture(autouse=True)
@@ -33,19 +33,19 @@ def run_around_tests():
 
 
 def test_serialization():
-    mpart.SerializeMapOptions(opts, "_test_serialization/opts.mpart")
-    mpart.SerializeFixedMultiIndexSet(mset_1, "_test_serialization/mset_1.mpart")
-    mpart.SerializeFixedMultiIndexSet(mset_2, "_test_serialization/mset_2.mpart")
-    mpart.SerializeMapCoeffs(map_1, "_test_serialization/map_1.mpart")
-    mpart.SerializeMapCoeffs(map_2, "_test_serialization/map_2.mpart")
-    mpart.SerializeMapCoeffs(triangular, "_test_serialization/triangular.mpart")
+    mt.SerializeMapOptions(opts, "_test_serialization/opts.mpart")
+    mt.SerializeFixedMultiIndexSet(mset_1, "_test_serialization/mset_1.mpart")
+    mt.SerializeFixedMultiIndexSet(mset_2, "_test_serialization/mset_2.mpart")
+    mt.SerializeMapCoeffs(map_1, "_test_serialization/map_1.mpart")
+    mt.SerializeMapCoeffs(map_2, "_test_serialization/map_2.mpart")
+    mt.SerializeMapCoeffs(triangular, "_test_serialization/triangular.mpart")
 
-    opts_s = mpart.DeserializeMapOptions("_test_serialization/opts.mpart")
-    mset_1_s = mpart.DeserializeFixedMultiIndexSet("_test_serialization/mset_1.mpart")
-    mset_2_s = mpart.DeserializeFixedMultiIndexSet("_test_serialization/mset_2.mpart")
-    map_1_s = mpart.DeserializeMapCoeffs("_test_serialization/map_1.mpart")
-    map_2_s = mpart.DeserializeMapCoeffs("_test_serialization/map_2.mpart")
-    triangularCoeff_s = mpart.DeserializeMapCoeffs("_test_serialization/triangular.mpart")
+    opts_s = mt.DeserializeMapOptions("_test_serialization/opts.mpart")
+    mset_1_s = mt.DeserializeFixedMultiIndexSet("_test_serialization/mset_1.mpart")
+    mset_2_s = mt.DeserializeFixedMultiIndexSet("_test_serialization/mset_2.mpart")
+    map_1_s = mt.DeserializeMapCoeffs("_test_serialization/map_1.mpart")
[…]
