### 2022-11-25 26cc718
Work on bindings for serialization
