### 2022-11-25 ae3b5c6
Fix ostream
