+++ bindings/python/tests/test_Serialization.py
@@ -5,61 +5,78 @@ import numpy as np
 import os
 import pytest
 
-opts = mt.MapOptions()
+# Initialize folder name for test
+foldername = "_test_serialization/"
 
-multis_1 = np.array([[0],[1]])  # linear
-multis_2 = np.array([[0,0],[0,1],[2,0]])  # quadratic in x_1, linear in x_2, matches form of target
+# Create MapOptions
+options = mt.MapOptions()
+options.basisType = mt.BasisTypes.ProbabilistHermite
+options.basisNorm = False;
 
-mset_1 = mt.MultiIndexSet(multis_1).fix(True)
-mset_2 = mt.MultiIndexSet(multis_2).fix(True)
+# Create FixedMultiIndexSet and map component
+multis = np.array([[0],[1]])
+mset= mt.MultiIndexSet(multis)
+fixed_mset = mset.fix(True)
 
-map_1 = mt.CreateComponent(mset_1, opts)
-map_2 = mt.CreateComponent(mset_2, opts)
+component = mt.CreateComponent(fixed_mset, options)
+coeffs = component.CoeffMap()
+for i in range(len(coeffs)):
+    coeffs[i] = 0.5*(i+1)
 
-triangular = mt.TriangularMap([map_1,map_2])
-triangular.SetCoeffs(np.concatenate([map_1.CoeffMap(), map_2.CoeffMap()]))
+# Serialize objects
+
+def DeserializeComponent():
+    # Deserialize the FixedMultiIndexSet
+    # Note that we need to construct the object before calling Deserialize
+    fixed_mset = mt.MultiIndexSet(np.array([[0]])).fix(True)
+    fixed_mset.Deserialize("fmset.mt")
+
+    # Deserialize the MapOptions
+    options = MapOptions()
+    options.Deserialize("opts.mt")
+
+    # Deserialize the Map Coefficients and construct the component
+    inputDim, outputDim, coeffs = mt.DeserializeMap("comp.mt")
+    component = mt.CreateComponent(fixed_mset, options)
+    component.SetCoeffs(coeffs)
 
 @pytest.fixture(autouse=True)
 def run_around_tests():
-    os.mkdir("_test_serialization")
+    os.mkdir(foldername)
     yield
-    os.remove("_test_serialization/opts.mpart")
-    os.remove("_test_serialization/mset_1.mpart")
-    os.remove("_test_serialization/mset_2.mpart")
-    os.remove("_test_serialization/map_1.mpart")
-    os.remove("_test_serialization/map_2.mpart")
-    os.remove("_test_serialization/triangular.mpart")
-    os.rmdir("_test_serialization")
+    os.remove(foldername + "opts.mt")
+    os.remove(foldername + "fmset.mt")
+    os.remove(foldername + "comp.mt")
+    os.rmdir(foldername)
 
 
 def test_serialization():
-    mt.SerializeMapOptions(opts, "_test_serialization/opts.mpart")
-    mt.SerializeFixedMultiIndexSet(mset_1, "_test_serialization/mset_1.mpart")
-    mt.SerializeFixedMultiIndexSet(mset_2, "_test_serialization/mset_2.mpart")
-    mt.SerializeMapCoeffs(map_1, "_test_serialization/map_1.mpart")
-    mt.SerializeMapCoeffs(map_2, "_test_serialization/map_2.mpart")
-    mt.SerializeMapCoeffs(triangular, "_test_serialization/triangular.mpart")
+    fixed_mset.Serialize(foldername + "fmset.mt")
+    options.Serialize(foldername + "opts.mt")
+    component.Serialize(foldername + "comp.mt")
 
[…]
