### 2022-11-26 15ad053
Slightly modify coeffs check serialization pytest
