### 2022-11-26 c64a7fe
Link to serialization in install docs
