### 2022-11-27 faf6b75
Initial work toward wavelet
