### 2022-11-29 fe83aef
Updated installations instructions to use PTHREAD instead of THREADS for kokkos backend.
