### 2022-12-02 ba01d0b
Working Wavelets
