+++ MParT/Wavelet.h
@@ -17,28 +17,38 @@ template<class Mixer>
 class Wavelet: public Mixer {
 
 public:
-    Wavelet(unsigned int c, unsigned int d): shift_(c-Mixer::a), scale_((d-c)/(Mixer::b-Mixer::a)) {}
+    Wavelet(double c, double d): shift_(c-Mixer::a), scale_((d-c)/(Mixer::b-Mixer::a)) {
+        std::cerr << "a = " << Mixer::a << ", b = " << Mixer::b << ", c = " << c << ", d = " << d << std::endl;
+        std::cerr << "shift_ = " << shift_ << std::endl;
+        std::cerr << "scale_ = " << scale_ << std::endl;
+    }
 
     /* Evaluates all wavelets up to a specified order. */
     KOKKOS_FUNCTION void EvaluateAll(double*              output,
                                     unsigned int         maxOrder,
                                     double               x) const
     {
-        unsigned int L,Q;
-        MaxOrder(maxOrder, L, Q);
-        unsigned int idx = 1;
-        double outscale = M_SQRT2;
-        unsigned int two_power_ell = 2;
         x -= shift_;
         x *= scale_;
-        output[0] = psi(x);
+        if(x > Mixer::b || x < Mixer::a) {
+            std::fill(output, output+maxOrder, 0.);
+            return;
+        }
+        unsigned int L,Q;
+        MaxOrder(maxOrder, L, Q);
+        unsigned int idx = 2;
+        double outscale = M_SQRT2;
+        unsigned int two_power_ell = 2;
+        output[0] = 0.;
+        output[1] = this->psi(x);
         x *= 2;
         for(unsigned int l=1; l<=L; ++l)
         {
             unsigned int Q_l = l == L ? Q+1 : two_power_ell;
+            // std::cerr << "l = " << l << ", Q_l = " << Q_l << ", outscale = " << outscale << std::endl;
             for(unsigned int q=0; q < Q_l; ++q)
             {
-                output[idx] = outscale*psi(x-q);
+                output[idx] = outscale*this->psi(x-q);
                 ++idx;
             }
             x *= 2;
@@ -56,25 +66,32 @@ public:
                              unsigned int maxDegree,
                              double       x) const
     {
+        x -= shift_;
+        x *= scale_;
+        if(x > Mixer::b || x < Mixer::a) {
+            std::fill(vals, vals+maxDegree, 0.);
+            std::fill(derivs, derivs+maxDegree, 0.);
+            return;
+        }
         unsigned int L,Q;
-        MaxOrder(maxOrder, L, Q);
-        unsigned int idx = 1;
+        MaxOrder(maxDegree, L, Q);
+        unsigned int idx = 2;
         double outscale = M_SQRT2;
         double outscale_deriv_inc = 2*M_SQRT2;
         double outscale_deriv = scale_*outscale_deriv_inc;
         unsigned int two_power_ell = 2;
-        x -= shift_;
-        x *= scale_;
-        vals[0] = psi(x);
-        derivs[0] = scale_*psi_derivative(x);
+        vals[0] = 0.;
+        vals[1] = this->psi(x);
+        derivs[0] = 0.;
+        derivs[1] = scale_*this->psi_derivative(x);
         x *= 2;
         for(unsigned int l=1; l<=L; ++l)
         {
             unsigned int Q_l = l == L ? Q+1 : two_power_ell;
             for(unsigned int q=0; q < Q_l; ++q)
             {
-                vals[idx] = outscale*psi(x-q);
-                derivs[idx] = outscale_deriv*psi_derivative(x-q);
[…]
