### 2022-12-05 #4843 @bonh
Thanks, @diehlpk! The DFG is funding my research stay at MIT.
