### 2022-12-05 #4843 @diehlpk
> Thanks, @diehlpk! The DFG is funding my research stay at MIT.

@bonh If you still related to a German institution, you might want to add that. If not, I think it is fine.
