### 2022-12-06 1d972cb
added affiliation and city
