### 2022-12-09 a09e682
Adding KeepCoeffs option for TriangularMap constructor.
