### 2022-12-12 fea7eec
Added cities to all affiliations.  Added Solea Energy as additional affiliation.
