### 2022-12-13 #4843 @mparno
@diehlpk I'm not quite sure what's going on.  The compilation works on the whedon preview page ([results are here](http://res.cloudinary.com/hkvhhndao/image/upload/v1670972129/klgyqeaitiz0n8kp7zih.pdf)).  Let me try again here.
