### 2022-12-14 #4843 @mparno
@diehlpk Thanks!

- MParT is now at [release v1.3.0](https://github.com/MeasureTransport/MParT/releases/tag/v1.3.0)
- This release has been uploaded to zenodo.  The zenodo DOI is 10.5281/zenodo.7435142

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7435142.svg)](https://doi.org/10.5281/zenodo.7435142)
