### 2022-12-15 #4843 @arfon
@mparno, just checking your paper here. The following sentence reads weirdly:

> The Monotone Parameterization Toolkit (MParT), pronounced aims to provide performance-portable implementations of such parameterizations.

Is `pronounced` meant to be in this sentence? If so, I'm not sure I quite know how to parse it :-)
