### 2022-12-15 #4843 @mparno
@arfon Yikes!  How did that slip through?   You're correct, "pronounced" should be removed.   Would you like me to update the paper in our repo?
