### 2022-12-16 f00b360
Adding tests for TriangularMaps
