+++ tests/Test_TriangularMap.cpp
@@ -7,7 +7,7 @@ using namespace mpart;
 using namespace Catch;
 using MemorySpace = Kokkos::HostSpace;
 
-TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_MonotoneComponents]" ) {
+TEST_CASE( "Testing 3d triangular map from MonotoneComponents with moveCoeffs=false", "[TriangularMap_MonotoneComponents]" ) {
 
     MapOptions options;
     options.basisType = BasisTypes::ProbabilistHermite;
@@ -221,9 +221,232 @@ TEST_CASE( "Testing 3d triangular map from MonotoneComponents", "[TriangularMap_
 
 }
 
+TEST_CASE( "Testing 3d triangular map from MonotoneComponents with moveCoeffs=true", "[TriangularMap_MonotoneComponents]" ) {
+
+    MapOptions options;
+    options.basisType = BasisTypes::ProbabilistHermite;
+    options.basisNorm = false;
+    
+    unsigned int numBlocks = 3;
+    unsigned int maxDegree = 2;
+    unsigned int extraInputs = 1;
+
+    unsigned int coeffSize = 0;
 
 
-TEST_CASE( "Testing TriangularMap made from smaller TriangularMaps", "[TriangularMap_TriangularMaps]" ) {
+
+    std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> blocks(numBlocks);
+    std::vector<Kokkos::View<double*,Kokkos::HostSpace>> coeffs_(numBlocks);
+    for(unsigned int i=0;i<numBlocks;++i){
+        
+        FixedMultiIndexSet<MemorySpace> mset(i+extraInputs+1,maxDegree);
+        coeffSize += mset.Size();
+
+        blocks.at(i) = MapFactory::CreateComponent<MemorySpace>(mset, options);
+
+        coeffs_.at(i) = Kokkos::View<double*,Kokkos::HostSpace>("Coefficients", blocks.at(i)->numCoeffs);
+
+        for(unsigned int j=0; j<blocks.at(i)->numCoeffs; ++j)
+            coeffs_.at(i)(j) = 0.1*(j+1);
+        
+        blocks.at(i)->SetCoeffs(coeffs_.at(i));
+        
+    }
+    bool moveCoeffs=true;
+    std::shared_ptr<ConditionalMapBase<MemorySpace>> triMap = std::make_shared<TriangularMap<MemorySpace>>(blocks, moveCoeffs);
+
+    CHECK(triMap->outputDim == numBlocks);
+    CHECK(triMap->inputDim == numBlocks+extraInputs);
+    CHECK(triMap->numCoeffs == coeffSize);
+
+        
+    SECTION("Coefficients"){
+
+
+        // Now make sure that the coefficients of each block were set
+        unsigned int cumCoeffInd = 0;
+        for(unsigned int i=0; i<numBlocks; ++i){
+            for(unsigned int j=0; j<blocks.at(i)->numCoeffs; ++j){
+                CHECK(blocks.at(i)->Coeffs()(j) == triMap->Coeffs()(cumCoeffInd)); // Values of coefficients should be equal
+                CHECK(blocks.at(i)->Coeffs()(j) == coeffs_.at(i)(j)); 
+                CHECK(&blocks.at(i)->Coeffs()(j) == &triMap->Coeffs()(cumCoeffInd)); // Memory location should also be the same (no copy)
+                cumCoeffInd++;
+            }
+        }
+    }
+
+    Kokkos::View<double*,Kokkos::HostSpace> coeffs(triMap->Coeffs().data(), triMap->numCoeffs);
+
+    unsigned int numSamps = 10;
[…]
