### 2022-12-17 3f75e30
Working on MoveCoeffs for composed map
weirdness: ComposedMap didn't have WrapCoeffs implemented
