+++ tests/Test_ComposedMap.cpp
@@ -9,7 +9,7 @@ using namespace mpart;
 using namespace Catch;
 using MemorySpace = Kokkos::HostSpace;
 
-TEST_CASE( "Testing 2 layer composed map", "[ShallowComposedMap]" ) {
+TEST_CASE( "Testing 2 layer composed map with moveCoeffs=false", "[ShallowComposedMap]" ) {
 
     MapOptions options;
     options.basisType = BasisTypes::ProbabilistHermite;
@@ -247,6 +247,249 @@ TEST_CASE( "Testing 2 layer composed map", "[ShallowComposedMap]" ) {
     }
 
 }
+
+TEST_CASE( "Testing 2 layer composed map with moveCoeffs=true", "[ShallowComposedMap]" ) {
+
+    MapOptions options;
+    options.basisType = BasisTypes::ProbabilistHermite;
+    options.basisNorm = false;
+    
+    unsigned int dim = 2;
+    unsigned int numMaps = 2;
+    unsigned int order = 2;
+    unsigned int coeffSize = 0;
+    
+
+    std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> maps(numMaps);
+    std::vector<Kokkos::View<double*,Kokkos::HostSpace>> coeffs_(numMaps);
+    for(unsigned int i=0;i<numMaps;++i){
+
+        maps.at(i) = MapFactory::CreateTriangular<Kokkos::HostSpace>(dim, dim, order, options);
+        coeffSize += maps.at(i)->numCoeffs;
+
+        coeffs_.at(i) = Kokkos::View<double*,Kokkos::HostSpace>("Coefficients", maps.at(i)->numCoeffs);
+
+        for(unsigned int j=0; j<maps.at(i)->numCoeffs; ++j)
+            coeffs_.at(i)(j) = 0.1*(j+1);
+        
+        maps.at(i)->SetCoeffs(coeffs_.at(i));
+    }
+    bool moveCoeffs=true;
+    std::shared_ptr<ConditionalMapBase<MemorySpace>> composedMap = std::make_shared<ComposedMap<MemorySpace>>(maps,moveCoeffs);
+
+    CHECK(composedMap->outputDim == dim);
+    CHECK(composedMap->inputDim == dim);
+    CHECK(composedMap->numCoeffs == coeffSize);
+
+    Kokkos::View<double*,Kokkos::HostSpace> coeffs(composedMap->Coeffs().data(), composedMap->numCoeffs);
+
+    SECTION("Coefficients"){
+
+        // Now make sure that the coefficients of each block were set
+        unsigned int cumCoeffInd = 0;
+        for(unsigned int i=0; i<numMaps; ++i){
+            for(unsigned int j=0; j<maps.at(i)->numCoeffs; ++j){
+                CHECK(maps.at(i)->Coeffs()(j) == coeffs(cumCoeffInd)); // Values of coefficients should be equal
+                CHECK(maps.at(i)->Coeffs()(j) == coeffs_.at(i)(j));
+                CHECK(&maps.at(i)->Coeffs()(j) == &coeffs(cumCoeffInd));// Memory location should also be the same (no copy)
+                cumCoeffInd++;
+            }
+        }
+    }
+
+
+    unsigned int numSamps = 10;
+    Kokkos::View<double**, Kokkos::HostSpace> in("Map Input", dim, numSamps);
+    //Eigen::RowMatrixXd in(dim, numSamps);
+    for(unsigned int i=0; i<dim; ++i){
+        for(unsigned int j=0; j<numSamps; ++j){
+            in(i,j) = double(i)/(dim) + double(j)/numSamps;
+        }
+    }
+
[…]
