### 2022-12-17 d8d439f
Finishing MoveCoeffs for ComposedMap
