+++ tests/Test_ComposedMap.cpp
@@ -307,7 +307,6 @@ TEST_CASE( "Testing 2 layer composed map with moveCoeffs=true", "[ShallowCompose
         }
     }
 
-    composedMap->SetCoeffs(coeffs);
     auto out = composedMap->Evaluate(in);
     
     SECTION("Evaluate"){
@@ -490,9 +489,7 @@ TEST_CASE( "Testing 2 layer composed map with moveCoeffs=true", "[ShallowCompose
 
 }
 
-
-
-TEST_CASE( "Testing 8 layer composed map", "[DeepComposedMap]" ) {
+TEST_CASE( "Testing 8 layer composed map with moveCoeffs=false", "[DeepComposedMap]" ) {
 
     MapOptions options;
     options.basisType = BasisTypes::ProbabilistHermite;
@@ -527,6 +524,237 @@ TEST_CASE( "Testing 8 layer composed map", "[DeepComposedMap]" ) {
     }
 
     composedMap->SetCoeffs(coeffs);
+    auto out = composedMap->Evaluate(in);
+    
+    SECTION("Evaluate"){
+        Kokkos::View<double**, Kokkos::HostSpace> trueOut("True output", dim, numSamps);
+        Kokkos::deep_copy(trueOut, in);
+        for(auto& comp : maps)
+            trueOut = comp->Evaluate(trueOut);
+
+        for(unsigned int i=0; i<in.extent(0); ++i){
+            for(unsigned int j=0; j<numSamps; ++j)
+                CHECK( out(i,j) == Approx(trueOut(i,j)).epsilon(1e-7).margin(1e-7));
+        }
+    }
+
+    SECTION("LogDeterminant"){
+        auto logDet = composedMap->LogDeterminant(in);
+
+        REQUIRE(logDet.extent(0)==numSamps);
+        Kokkos::View<double*, Kokkos::HostSpace> truth("True Log Det", numSamps);
+        Kokkos::View<double*, Kokkos::HostSpace> partialDet("Partial log det", numSamps);
+        Kokkos::View<double**, Kokkos::HostSpace> currPts("intermediate points", in.extent(0), in.extent(1));
+        Kokkos::deep_copy(currPts, in);
+
+        for(auto& map : maps){
+            partialDet = map->LogDeterminant(currPts);
+            currPts = map->Evaluate(currPts);
+            truth += partialDet;
+        }
+
+        for(unsigned int j=0; j<numSamps; ++j)
+            CHECK(logDet(j) == Approx(truth(j)).epsilon(1e-10));
+
+    }
+
+    SECTION("CoeffGrad"){
+
+        Kokkos::View<double**,Kokkos::HostSpace> sens("Sensitivities", composedMap->outputDim, numSamps);
+        
+        for(unsigned int j=0; j<numSamps; ++j){
+            for(unsigned int i=0; i<composedMap->outputDim; ++i){
+                sens(i,j) = 1.0 + 0.1*i + j;
+            }
+        }
+
+        Kokkos::View<double**,Kokkos::HostSpace> evals = composedMap->Evaluate(in);
+        Kokkos::View<double**,Kokkos::HostSpace> evals2;
+
+        Kokkos::View<double**,Kokkos::HostSpace> coeffGrad = composedMap->CoeffGrad(in, sens);
+
+        REQUIRE(coeffGrad.extent(0)==composedMap->numCoeffs);
+        REQUIRE(coeffGrad.extent(1)==numSamps);
+
+        // Compare with finite differences
+        double fdstep = 1e-5;
+        for(unsigned int i=0; i<composedMap->numCoeffs; ++i){
+            coeffs(i) += fdstep;
+
+            composedMap->SetCoeffs(coeffs);
[…]
