### 2022-12-18 19364f4
IdentityMap in python
