+++ bindings/python/CMakeLists.txt
@@ -12,6 +12,7 @@ set(PYTHON_BINDING_SOURCES
   src/SummarizedMap.cpp
   # src/DebugMap.cpp
   src/MapFactory.cpp
+  src/IdentityMap.cpp
 
 
   ../common/src/CommonUtilities.cpp
