### 2022-12-19 834e533
Add docs for moveCoeffs in ComposedMap
