### 2022-12-19 57b3e7e
Add docs for moveCoeffs in TriangularMap
