+++ MParT/TriangularMap.h
@@ -36,6 +36,9 @@ public:
 
     @param components A vector of ConditionalMapBase objects defining each \f$T_k\f$ in the block triangular map.
                       To maintain the correct block structure, the dimensions of the components must satisfy \f$N_k = N_{k-1}+M_{k}\f$.
+    @param moveCoeffs A boolean to choose whether to keep the coefficients from the maps in `components` or make new ones. If true,
+                      the new object takes ownership of all the coefficients vectors within all the maps in `components` (changing the coefficients
+                      in the new map will then change the coefficients in the original maps). If false, no coefficients are copied or created.
     */
     TriangularMap(std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>> const& components, bool moveCoeffs=false);
 
