+++ MParT/Wavelet.h
@@ -25,10 +25,10 @@ public:
     {
         x *= scale_;
         x -= shift_;
-        if(x > Mixer::b || x < Mixer::a) {
-            std::fill(output, output+maxOrder, 0.);
-            return;
-        }
+        // if(x > Mixer::b || x < Mixer::a) {
+        //     std::fill(output, output+maxOrder, 0.);
+        //     return;
+        // }
         unsigned int L,Q;
         MaxOrder(maxOrder, L, Q);
         unsigned int idx = 2;
@@ -43,7 +43,8 @@ public:
 
             for(unsigned int q=0; q < Q_l; ++q)
             {
-                output[idx] = outscale*this->psi(x-(Mixer::a + (2*q+1)*(Mixer::b-Mixer::a)/(2*two_power_ell)));
+                double shift_q = Mixer::calculate_shift(two_power_ell, q);
+                output[idx] = outscale*this->psi(x-shift_q);
                 ++idx;
             }
             x *= 2;
@@ -63,11 +64,11 @@ public:
     {
         x *= scale_;
         x -= shift_;
-        if(x > Mixer::b || x < Mixer::a) {
-            std::fill(vals, vals+maxDegree, 0.);
-            std::fill(derivs, derivs+maxDegree, 0.);
-            return;
-        }
+        // if(x > Mixer::b || x < Mixer::a) {
+        //     std::fill(vals, vals+maxDegree, 0.);
+        //     std::fill(derivs, derivs+maxDegree, 0.);
+        //     return;
+        // }
         unsigned int L,Q;
         MaxOrder(maxDegree, L, Q);
         unsigned int idx = 2;
@@ -85,8 +86,9 @@ public:
             unsigned int Q_l = l == L ? Q+1 : two_power_ell;
             for(unsigned int q=0; q < Q_l; ++q)
             {
-                vals[idx] = outscale*this->psi(x-q);
-                derivs[idx] = outscale_deriv*this->psi_derivative(x-q);
+                double shift_q = Mixer::calculate_shift(two_power_ell, q);
+                vals[idx] = outscale*this->psi(x-shift_q);
+                derivs[idx] = outscale_deriv*this->psi_derivative(x-shift_q);
                 ++idx;
             }
             x *= 2;
@@ -106,12 +108,12 @@ public:
     {
         x *= scale_;
         x -= shift_;
-        if(x > Mixer::b || x < Mixer::a) {
-            std::fill(vals, vals+maxDegree, 0.);
-            std::fill(derivs, derivs+maxDegree, 0.);
-            std::fill(derivs2, derivs2+maxDegree, 0.);
-            return;
-        }
+        // if(x > Mixer::b || x < Mixer::a) {
+        //     std::fill(vals, vals+maxDegree, 0.);
+        //     std::fill(derivs, derivs+maxDegree, 0.);
+        //     std::fill(derivs2, derivs2+maxDegree, 0.);
+        //     return;
+        // }
         unsigned int L,Q;
         MaxOrder(maxDegree, L, Q);
         unsigned int idx = 2;
@@ -133,9 +135,10 @@ public:
             unsigned int Q_l = l == L ? Q+1 : two_power_ell;
             for(unsigned int q=0; q < Q_l; ++q)
             {
-                vals[idx] = outscale*this->psi(x-q);
-                derivs[idx] = outscale_deriv*this->psi_derivative(x-q);
-                derivs2[idx] = outscale_deriv2*this->psi_second_derivative(x-q);
+                double shift_q = Mixer::calculate_shift(two_power_ell, q);
+                vals[idx] = outscale*this->psi(x-shift_q);
[…]
