+++ src/Utilities/LinearAlgebra.cpp
@@ -7,7 +7,7 @@ using namespace mpart;
 
 
 template<>
-void mpart::dgemm(double                                     alpha, 
+void mpart::dgemm(double                                     alpha,
                   TransposeObject<Kokkos::HostSpace>         A,
                   TransposeObject<Kokkos::HostSpace>         B,
                   double                                     beta,
@@ -59,7 +59,7 @@ void PartialPivLU<Kokkos::HostSpace>::compute(Kokkos::View<const double**,Kokkos
 
 template<>
 void PartialPivLU<Kokkos::HostSpace>::solveInPlace(Kokkos::View<double**,Kokkos::LayoutLeft,Kokkos::HostSpace> x)
-{   
+{
     auto eigX = KokkosToMat(x);
     eigX = (luSolver_->permutationP() * eigX).eval();
     luSolver_->matrixLU().template triangularView<Eigen::UnitLower>().solveInPlace(eigX);
@@ -70,26 +70,77 @@ void PartialPivLU<Kokkos::HostSpace>::solveInPlace(Kokkos::View<double**,Kokkos:
 template<>
 double PartialPivLU<Kokkos::HostSpace>::determinant() const
 {
-    assert(isComputed); 
+    assert(isComputed);
     return luSolver_->determinant();
 }
 
+
+template<>
+void Cholesky<Kokkos::HostSpace>::compute(Kokkos::View<const double**,Kokkos::LayoutLeft,Kokkos::HostSpace> A)
+{
+    auto eigA = ConstKokkosToMat(A);
+    // Use eigen to compute the Cholesky decomposition in place
+    cholSolver_ = std::make_shared<Eigen::LLT<Eigen::MatrixXd>>(eigA);
+    isComputed = true;
+}
+
+template<>
+void Cholesky<Kokkos::HostSpace>::solveInPlace(Kokkos::View<double**,Kokkos::LayoutLeft,Kokkos::HostSpace> x)
+{
+    auto eigX = KokkosToMat(x);
+    cholSolver_->matrixL().solveInPlace(eigX);
+    cholSolver_->matrixU().solveInPlace(eigX);
+}
+
+template<>
+void Cholesky<Kokkos::HostSpace>::solveLInPlace(Kokkos::View<double**,Kokkos::LayoutLeft,Kokkos::HostSpace> x)
+{
+    auto eigX = KokkosToMat(x);
+    cholSolver_->matrixL().solveInPlace(eigX);
+}
+
+template<>
+Kokkos::View<double**,Kokkos::LayoutLeft,Kokkos::HostSpace> Cholesky<Kokkos::HostSpace>::multiplyL(Kokkos::View<const double**,Kokkos::LayoutLeft,Kokkos::HostSpace> x)
+{
+    Kokkos::View<double**,Kokkos::LayoutLeft,Kokkos::HostSpace> y("y", x.extent(0),x.extent(1));
+    auto eigX = ConstKokkosToMat(x);
+    auto eigY = KokkosToMat(y);
+    eigY.noalias() = cholSolver_->matrixL() * eigX;
+    return y;
+}
+
+template<typename MemorySpace>
+Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace> Cholesky<MemorySpace>::solve(StridedMatrix<const double, MemorySpace> x)
+{
+    Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace> output("solution", x.extent(0),x.extent(1));
+    Kokkos::deep_copy(output, x);
+    solveInPlace(output);
+    return output;
+}
+
+template<>
+double Cholesky<Kokkos::HostSpace>::determinant() const
+{
+    assert(isComputed);
+    double sqrt_det = cholSolver_->matrixL().determinant();
+    return sqrt_det*sqrt_det;
+}
 
 
[…]
