+++ tests/Test_LinearAlgebra.cpp
@@ -38,7 +38,7 @@ TEST_CASE( "Testing Vector Sum", "[Vec+vec]" ) {
     A(0) = 1.0;
     A(1) = 0.5;
 
-    B(0) = 1.0; 
+    B(0) = 1.0;
     B(1) = 0.125;
 
     Kokkos::View<double*, Kokkos::HostSpace> C = A + B;
@@ -71,10 +71,10 @@ TEST_CASE( "Testing Linear Mat-Mat product", "[LinearAlgebra_MatMat]" ) {
     SECTION("No Transpose") {
         C = A*B;
         trueC = eigA*eigB;
-        
+
         REQUIRE(C.extent(0) == trueC.rows());
         REQUIRE(C.extent(1) == trueC.cols());
-        
+
         for(unsigned int j=0; j<C.extent(1); ++j){
             for(unsigned int i=0; i<C.extent(0); ++i){
                 CHECK(C(i,j) == Approx(trueC(i,j)).epsilon(1e-14).margin(1e-14));
@@ -85,10 +85,10 @@ TEST_CASE( "Testing Linear Mat-Mat product", "[LinearAlgebra_MatMat]" ) {
     SECTION("A Transpose") {
         C = transpose(A)*B;
         trueC = eigA.transpose()*eigB;
-        
+
         REQUIRE(C.extent(0) == trueC.rows());
         REQUIRE(C.extent(1) == trueC.cols());
-        
+
         for(unsigned int j=0; j<C.extent(1); ++j){
             for(unsigned int i=0; i<C.extent(0); ++i){
                 CHECK(C(i,j) == Approx(trueC(i,j)).epsilon(1e-14).margin(1e-14));
@@ -99,10 +99,10 @@ TEST_CASE( "Testing Linear Mat-Mat product", "[LinearAlgebra_MatMat]" ) {
     SECTION("B Transpose") {
         C = A*transpose(B);
         trueC = eigA*eigB.transpose();
-        
+
         REQUIRE(C.extent(0) == trueC.rows());
         REQUIRE(C.extent(1) == trueC.cols());
-        
+
         for(unsigned int j=0; j<C.extent(1); ++j){
             for(unsigned int i=0; i<C.extent(0); ++i){
                 CHECK(C(i,j) == Approx(trueC(i,j)).epsilon(1e-14).margin(1e-14));
@@ -113,10 +113,10 @@ TEST_CASE( "Testing Linear Mat-Mat product", "[LinearAlgebra_MatMat]" ) {
     SECTION("Both Transpose") {
         C = transpose(A)*transpose(B);
         trueC = eigA.transpose() * eigB.transpose();
-        
+
         REQUIRE(C.extent(0) == trueC.rows());
         REQUIRE(C.extent(1) == trueC.cols());
-        
+
         for(unsigned int j=0; j<C.extent(1); ++j){
             for(unsigned int i=0; i<C.extent(0); ++i){
                 CHECK(C(i,j) == Approx(trueC(i,j)).epsilon(1e-14).margin(1e-14));
@@ -127,10 +127,10 @@ TEST_CASE( "Testing Linear Mat-Mat product", "[LinearAlgebra_MatMat]" ) {
     SECTION("Extra Transpose") {
         C = transpose(A)*transpose(transpose(B));
         trueC = eigA.transpose() * eigB;
-        
+
         REQUIRE(C.extent(0) == trueC.rows());
         REQUIRE(C.extent(1) == trueC.cols());
-        
+
[…]
