+++ src/ComposedMap.cpp
@@ -8,7 +8,7 @@
 using namespace mpart;
 
 template<typename MemorySpace>
-ComposedMap<MemorySpace>::Checkpointer::Checkpointer(unsigned int maxSaves, 
+ComposedMap<MemorySpace>::Checkpointer::Checkpointer(unsigned int maxSaves,
                                                      StridedMatrix<const double, MemorySpace> initialPts,
                                                      std::vector<std::shared_ptr<ConditionalMapBase<MemorySpace>>>& comps) : maxSaves_(maxSaves),
                                                                                                                             maps_(comps)
@@ -23,7 +23,7 @@ ComposedMap<MemorySpace>::Checkpointer::Checkpointer(unsigned int maxSaves,
 
     checkpointLayers_.push_back(0);
 
-    // Allocate the workspace 
+    // Allocate the workspace
     workspace1_ = Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>("xi", initialPts.extent(0), initialPts.extent(1));
     workspace2_ = Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace>("xj", initialPts.extent(0), initialPts.extent(1));
 }
@@ -37,7 +37,7 @@ int ComposedMap<MemorySpace>::Checkpointer::GetNextCheckpoint(unsigned int layer
 
     int currLayer = checkpointLayers_.back();
     int oldLayer = currLayer;
-    
+
     //ds = snaps - *check;
     int availSaves = int(maxSaves_) - checkpoints_.size();
     if(availSaves==0){
@@ -46,12 +46,12 @@ int ComposedMap<MemorySpace>::Checkpointer::GetNextCheckpoint(unsigned int layer
 
     int reps = 0;
     int range = 1;
-    
-    while(range < layerInd+1 - currLayer) { 
+
+    while(range < layerInd+1 - currLayer) {
         reps += 1;
-        range = range*(reps + availSaves)/reps; 
+        range = range*(reps + availSaves)/reps;
     }
-    
+
     int bino1, bino2, bino3, bino4, bino5;
     bino1 = range*reps/(availSaves+reps);
     bino2 = (availSaves > 1) ? bino1*availSaves/(availSaves+reps-1) : 1;
@@ -71,15 +71,15 @@ int ComposedMap<MemorySpace>::Checkpointer::GetNextCheckpoint(unsigned int layer
     if(layerInd-currLayer <= bino1 + bino3){
         currLayer = currLayer + bino4;
     }else{
-        if(layerInd-currLayer >= range - bino5){ 
-            currLayer = currLayer + bino1; 
+        if(layerInd-currLayer >= range - bino5){
+            currLayer = currLayer + bino1;
         }else{
             currLayer = layerInd-bino2-bino3;
         }
     }
 
-    if (currLayer == oldLayer) 
-        currLayer = oldLayer + 1;  
+    if (currLayer == oldLayer)
+        currLayer = oldLayer + 1;
 
     return currLayer;
 }
@@ -87,7 +87,7 @@ int ComposedMap<MemorySpace>::Checkpointer::GetNextCheckpoint(unsigned int layer
 
 template<typename MemorySpace>
 Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace> ComposedMap<MemorySpace>::Checkpointer::GetLayerInput(unsigned int layerInd)
-{   
+{
     // Check if the layer asked for is the last checkpoint
     if(layerInd==checkpointLayers_.back()){
         return checkpoints_.back();
@@ -99,21 +99,21 @@ Kokkos::View<double**, Kokkos::LayoutLeft, MemorySpace> ComposedMap<MemorySpace>
         return GetLayerInput(layerInd);
 
     }else{
[…]
