+++ MParT/Utilities/GPUtils.h
@@ -1,7 +1,6 @@
 #include <Kokkos_Core.hpp>
 
 namespace mpart{
-#define MPART_ENABLE_GPU 1
 #if defined(MPART_ENABLE_GPU)
 
 // Only enable DeviceSpace if the DefaultExecutionSpace is a GPU space.
