### 2022-12-20 15791bd
Change how kokkos chooses to use Kokkos::numbers
