### 2022-12-20 768c42b
Fix ComposedMap.cpp
