### 2022-12-20 b370d3e
Add basic tests for DensityBase
