### 2022-12-20 a94df3d
Remove extraneous comments
