### 2022-12-21 0c3d20f
Working on testing segfault
