### 2022-12-21 7c5a6ac
Fix sample generation in test
