### 2022-12-21 a2f8072
Fixed typo in joss paper.
