### 2022-12-21 #4843 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
