+++ MParT/Distributions/GaussianDistribution.h
@@ -24,7 +24,13 @@ class GaussianSamplerDensity {
     private:
 
 #if (KOKKOS_VERSION / 10000 < 4)
+
+#if (KOKKOS_VERSION / 10000 == 3) && (KOKKOS_VERSION / 100 % 100 > 5)
     const double logtau_ = std::log(2*Kokkos::Experimental::pi_v<double>);
+#else
+    const double logtau_ = std::log(2*M_PI);
+#endif
+
 #else
     const double logtau_ = std::log(2*Kokkos::numbers::pi_v<double>);
 #endif // KOKKOS_VERSION
