+++ tests/Distributions/Test_GaussianDistribution.cpp
@@ -1,0 +1,20 @@
+#include<algorithm>
+#include <catch2/catch_all.hpp>
+#include "MParT/Utilities/ArrayConversions.h"
+#include "MParT/Distributions/GaussianDistribution.h"
+
+
+TEST_CASE( "Testing Gaussian Distribution", "[GaussianDist]") {
+    SECTION( "Default Covariance, Default mean" ) {
+
+    }
+    SECTION( "Default Covariance, unit mean in all dimensions" ) {
+
+    }
+    SECTION( "Diagonal Covariance, Default mean" ) {
+
+    }
+    SECTION( "Diagonal Covariance, unit mean in all dimensions" ) {
+
+    }
+}
