+++ tests/Distributions/Test_Distribution.cpp
@@ -1,0 +1,3 @@
+#include "Test_Distributions_Common.h"
+
+TEST_CASE( "Testing Distribution")
