+++ MParT/Distributions/DensityBase.h
@@ -15,7 +15,7 @@ namespace mpart {
 
     public:
 
-    DensityBase() {};
+    DensityBase(unsigned int dim): dim_(dim) {};
 
     virtual ~DensityBase() = default;
 
@@ -57,6 +57,8 @@ namespace mpart {
     template<typename AnyMemorySpace>
     StridedMatrix<double, AnyMemorySpace> GradLogDensity(StridedMatrix<const double, AnyMemorySpace> const &X);
 
+    protected:
+    unsigned int dim_;
     };
 }
 
